% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{abstract_properties}
\alias{abstract_properties}
\alias{shape_abstract}
\title{Abstract Properties}
\usage{
shape_abstract(x)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\c}
}
\description{
Calls the extractor after converting the input to an \code{\link{AbstractTensor}}.
}
