% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize.R
\name{add_libs}
\alias{add_libs}
\title{Add to .libPaths from a local path}
\usage{
add_libs(local_path)
}
\arguments{
\item{local_path}{character, identifying a local path}
}
\value{
character with latest .libPaths()
}
\description{
Add to .libPaths from a local path
}
\details{
Add the synced google bucket path(which is now local) to
    the library path .libPaths() so that R can see the hosted
    libraries.
}
