% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{ambiguous}
\alias{ambiguous}
\title{Get Ambiguity of an Array}
\usage{
ambiguous(x, ...)
}
\arguments{
\item{x}{An array object}

\item{...}{Additional arguments (unused)}
}
\value{
\code{logical(1)} - \code{TRUE} if the dtype is ambiguous, \code{FALSE} otherwise
}
\description{
Returns whether the array's dtype is ambiguous.
}
