% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnVIL-defunct.R
\name{AnVIL-defunct}
\alias{AnVIL-defunct}
\alias{avworkflow_configuration}
\alias{avworkflow_import_configuration}
\alias{Gen3Fence}
\alias{Gen3Indexd}
\alias{Gen3Sheepdog}
\alias{Gen3Peregrine}
\alias{install}
\alias{repository}
\alias{BINARY_BASE_URL}
\alias{repositories}
\title{Defunct functions in package \sQuote{AnVIL}}
\usage{
avworkflow_configuration(
  configuration_namespace,
  configuration_name,
  namespace = avworkspace_namespace(),
  name = avworkspace_name()
)

avworkflow_import_configuration(
  config,
  namespace = avworkspace_namespace(),
  name = avworkspace_name()
)

Gen3Fence()

Gen3Indexd()

Gen3Sheepdog()

Gen3Peregrine()

install(
  pkgs = character(),
  ...,
  version = BiocManager::version(),
  binary_base_url = BINARY_BASE_URL
)

repository(version = BiocManager::version(), binary_base_url = BINARY_BASE_URL)

repositories(
  version = BiocManager::version(),
  binary_base_url = BINARY_BASE_URL
)
}
\arguments{
\item{configuration_namespace}{character(1).}

\item{configuration_name}{character(1).}

\item{namespace}{character(1).}

\item{name}{character(1).}

\item{config}{\code{avworkflow_configuration} object.}

\item{pkgs}{\code{character()} packages to install from binary repository.}

\item{...}{additional arguments. \code{install()} passes additional
arguments to
\code{BiocManager::install()}. \code{print.repository_stats()} ignores
the additional arguments.}

\item{version}{\code{character(1)} or \code{package_version} Bioconductor
version, e.g., "3.12".}

\item{binary_base_url}{\code{character(1)} host and base path for binary
package 'CRAN-style' repository; not usually required by the
end-user.}
}
\value{
\verb{gen3_*} APIs are not fully implemented, because a service
endpoint has not been identified.

\code{Gen3Fence()} returns the authentication API at
https://raw.githubusercontent.com/uc-cdis/fence/master/openapis/swagger.yaml

\code{Gen3Indexd()} returns the indexing service API documented at
https://raw.githubusercontent.com/uc-cdis/indexd/master/openapis/swagger.yaml

\code{Gen3Sheepdog} returns the submission services API at
https://raw.githubusercontent.com/uc-cdis/sheepdog/master/openapi/swagger.yml

\code{Gen3Peregrine} returns the graphQL query services API at
https://raw.githubusercontent.com/uc-cdis/peregrine/master/openapis/swagger.yaml
}
\description{
These functions are provided for compatibility with
older versions of \sQuote{AnVIL} only, and will be defunct at
the next release.

\code{install()} is deprecated in favor of
\code{BiocManager::install()}.

\code{repository()} is deprecated in favor of
\code{BiocManager::containerRepository()}.

\code{repositories()} is deprecated in favor of
\code{BiocManager::repositories()}.
}
\details{
The following functions are defunct and will be deleted after the
next Bioconductor release of \sQuote{AnVIL}. Use the replacement indicated
below:
\itemize{
\item \code{avworkflow_configuration()}: \code{\link{avworkflow_configuration_get}}
\item \code{avworkflow_import_configuration()}: \code{\link{avworkflow_configuration_set}}
}
}
