% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avworkflow_configuration.R
\name{AnVIL-deprecated}
\alias{AnVIL-deprecated}
\alias{avworkflow_configuration}
\alias{avworkflow_import_configuration}
\title{Deprecated functions in package \sQuote{AnVIL}}
\usage{
avworkflow_configuration(
  configuration_namespace,
  configuration_name,
  namespace = avworkspace_namespace(),
  name = avworkspace_name()
)

avworkflow_import_configuration(
  config,
  namespace = avworkspace_namespace(),
  name = avworkspace_name()
)
}
\arguments{
\item{configuration_namespace}{character(1).}

\item{configuration_name}{character(1).}

\item{namespace}{character(1).}

\item{name}{character(1).}

\item{config}{\code{avworkflow_configuration} object.}
}
\description{
These functions are provided for compatibility with
older versions ofi \sQuote{AnVIL} only, and will be defunct at
the next release.
}
\details{
The following functions are deprecated and will be made
defunct; use the replacement indicated below:\preformatted{\\itemize\{
    \\item\{avworkflow_configuration: \code{\link{avworkflow_configuration_get}}\}
    \\item\{avworkflow_import_configuration: \code{\link{avworkflow_configuration_set}}\}
\}
}
}
