% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnVIL-package.R
\docType{package}
\name{AnVIL-package}
\alias{AnVIL}
\alias{AnVIL-package}
\title{AnVIL: Bioconductor on the AnVIL compute environment}
\description{
The AnVIL is a cloud computing resource developed in part by the National Human Genome Research Institute. The AnVIL package provides programatic access to the Dockstore, Leonardo, Rawls, TDR, and Terra RESTful programming interfaces. For platform-specific user-level functionality, see either the AnVILGCP or AnVILAz package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Bioconductor/AnVIL}
  \item Report bugs at \url{https://github.com/Bioconductor/AnVIL/issues}
}

}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramos@sph.cuny.edu} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})

Authors:
\itemize{
  \item Martin Morgan (\href{https://orcid.org/0000-0002-5874-8148}{ORCID})
  \item Kayla Interdonato
  \item Yubo Cheng
  \item Nitesh Turaga
}

Other contributors:
\itemize{
  \item BJ Stubbs [contributor]
  \item Vincent Carey [contributor]
  \item Sehyun Oh [contributor]
  \item Sweta Gopaulakrishnan [contributor]
  \item Valerie Obenchain [contributor]
}

}
\keyword{internal}
