% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anvil_options.R
\name{anvil_options}
\alias{anvil_options}
\title{Get or set AnVIL options}
\usage{
anvil_options(key, value)
}
\arguments{
\item{key}{character(1) name of option.}

\item{value}{ANY value associated with key.}
}
\value{
`anvil_options()` returns a named list of key-value
    pairs.

    `anvil_options("leonardo_host")` returns the value of option
    `"leonardo_host"`.

    `anvil_options("key", "value")` updates the option `key` to
    `value`, returning the previous value invisibly.
}
\description{
View available options and current values with
    `anvil_options()`.  Retrieve specific option with, e.g.,
    `anvil_options("leonardo_host")`.  Set anvil with
    `anvil_options("leonardo_host", <value>)`.
}
\details{
AnVIL requires OAuth 2.0 credentials identifying the
    application. These must be added to the package source before
    installation. Do this by visiting
    `https://console.cloud.google.com/apis/credentials?authuser=1&project=anvil-leo-dev` and
    downloading the "Bioconductor-AnVIL" credentials to a file
    `inst/extdata/leonardo_access.json`.
}
\examples{
anvil_options("leonardo_host")
}
