% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{AnvilBox}
\alias{AnvilBox}
\title{AnvilBox}
\description{
Virtual base class for GraphBox and DebugBox.
This class is used to represent values during graph construction and debugging.
Cannot be instantiated directly - use \code{\link[=GraphBox]{GraphBox()}} or \code{\link[=DebugBox]{DebugBox()}} instead.
}
