% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{AnvlPrimitive}
\alias{AnvlPrimitive}
\title{AnvlPrimitive}
\usage{
AnvlPrimitive(name, subgraphs = character())
}
\arguments{
\item{name}{(\code{character()})\cr
The name of the primitive.}

\item{subgraphs}{(\code{character()})\cr
Names of parameters that are subgraphs. Only used if \code{higher_order = TRUE}.}
}
\value{
(\code{AnvlPrimitive})
}
\description{
Primitive interpretation rule.
Note that \code{[[} and \verb{[[<-} access the interpretation rules.
To access other fields, use \code{$} and \verb{$<-}.

A primitive is considered higher-order if it contains subgraphs.
}
