% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{arr}
\alias{arr}
\title{Create an R array}
\usage{
arr(..., shape = NULL)
}
\arguments{
\item{...}{(any)\cr
Values of new array.}

\item{shape}{(\code{NULL} | \code{integer()})\cr
Shape of new array. If \code{NULL} (default), uses length of elements to create a 1D array.}
}
\description{
Create an R array without having to wrap data in \code{c()}
}
\examples{
arr(1, 2, 3)
arr(1, 2, 3, 4, shape = c(2, 2))
}
