% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{arrayish}
\alias{arrayish}
\alias{is_arrayish}
\title{Array-like Objects}
\usage{
is_arrayish(x, convert_ok = TRUE)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}

\item{convert_ok}{(\code{logical(1)})\cr
Whether to accept \code{numeric(1)} and \code{logical(1)} and R arrays of type \code{numeric} and \code{logical}.}
}
\value{
\code{logical(1)}
}
\description{
A \code{arrayish} value is any object that can be input to a primitive such as \code{\link{prim_add}}.

During runtime of a JIT-compiled function, these are \code{\link{AnvilArray}} objects.

The following types are arrayish (during tracing):
\itemize{
\item \code{\link{AnvilArray}}: a concrete array holding data on a device.
\item \code{\link{GraphBox}}: a boxed abstract array representing a value in a graph.
\item Length-1 vectors: \code{numeric(1)} and \code{logical(1)}
\item R arrays of types: \code{numeric} and \code{logical}.
}

Use \code{\link[=is_arrayish]{is_arrayish()}} to check whether a value is arrayish.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
# AnvilArrays are arrayish
is_arrayish(nv_array(1:4))

# Scalar R literals are arrayish by default
is_arrayish(1.5)
# R arrays are arrayish by default
is_arrayish(array(1.5))

# R arrays
is_arrayish(array(1:4), convert_ok = TRUE)
is_arrayish(array(1:4), convert_ok = FALSE)

# Length 1 vectors
is_arrayish(1.5, convert_ok = FALSE)
is_arrayish(1.5, convert_ok = TRUE)
\dontshow{\}) # examplesIf}
}
\seealso{
\link{AnvilArray}, \link{GraphBox}
}
