% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{as_anvil_array}
\alias{as_anvil_array}
\alias{as_anvil_arrays}
\title{Convert to AnvilArray}
\usage{
as_anvil_array(x, device = NULL)

as_anvil_arrays(...)
}
\arguments{
\item{x}{(\code{\link{arrayish}})\cr
Input to standardize.}

\item{device}{(\code{NULL} | \code{\link{device}})\cr
Target device. If \code{x} is an \code{AnvilArray} on a different device, an error
is raised.}

\item{...}{(\code{\link{arrayish}})\cr
Inputs to align.}
}
\value{
(\code{\link{AnvilArray}} for \code{as_anvil_array()}, \code{list} of \code{\link{AnvilArray}}s
for \code{as_anvil_arrays()}).
}
\description{
Use this to canonicalize inputs at the start of a function so it works
both with eager executing and in combination with \code{\link[=jit]{jit()}}.
Use \code{\link[=as_anvil_array]{as_anvil_array()}} for a single input and \code{\link[=as_anvil_arrays]{as_anvil_arrays()}} for multiple inputs.
The latter will also ensure all arrays are from the same backend and live on the same device.
}
\details{
During tracing, \link[=GraphBox]{boxes} are returned as is.
During eager mode, R literals and arrays are converted to \code{AnvilArray}s on the specified device.
For \code{AnvilArray} inputs, we check that they live on provided device (if specified).
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
as_anvil_array(1L)
as_anvil_arrays(nv_array(1:3), 1L)
\dontshow{\}) # examplesIf}
}
