% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-functions.R
\name{as_r_function}
\alias{as_r_function}
\title{Convert to an R function}
\usage{
as_r_function(graph)
}
\arguments{
\item{graph}{(\code{\link{Graph}})\cr
Graph to convert.}
}
\value{
(\code{function})
}
\description{
Convenience wrapper around \code{\link[=graph_to_r_function]{graph_to_r_function()}} that always includes the
\code{declare(type(...))} header.
}
