% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-converters.R
\name{at2vt}
\alias{at2vt}
\title{Convert AbstractArray to ValueType}
\usage{
at2vt(x)
}
\arguments{
\item{x}{(\code{AbstractArray})}
}
\value{
(\code{ValueType})
}
\description{
Convert an \code{AbstractArray} to a \code{ValueType}.
}
