% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{atype}
\alias{atype}
\title{Abstract Type}
\usage{
atype(x)
}
\arguments{
\item{x}{(\code{tensorish})\cr
Object to convert.}
}
\value{
\code{\link{AbstractTensor}}
}
\description{
Convert a tensor-like object to its abstract type.
}
