% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate against the google cloud app `anvil-leo-dev`}
\usage{
authenticate(cache = getOption("httr_oauth_cache"))
}
\arguments{
\item{cache}{(optional) logical(1) or character(1) describing how
the OAuth 2.0 token will be managed. See
`?httr:oauth2.0_token`.}
}
\value{
An 'R6' token instance containing the authorization,
    invisibly.
}
\description{
Authenticate against the google cloud app `anvil-leo-dev`
}
