% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av.R
\name{avstop_for_status}
\alias{avstop_for_status}
\title{Check HTTP code before continuing}
\usage{
avstop_for_status(response, op)
}
\arguments{
\item{response}{A \code{response} object returned by the API service request.}

\item{op}{character(1) A description of the operation that failed.}
}
\value{
called for the side effect of \code{stop()}-ping with an informative
message if the HTTP status code is not 2xx.
}
\description{
\code{avstop_for_status()} checks the HTTP status code of a
response object and stops with an informative message if the
status code is not 2xx. This utility function is mainly called by other
functions that make API requests.
}
\details{
\code{avstop_for_status()} uses \code{httr}'s \code{http_condition()} and
\code{status_code} to obtain the condition message and identify the condition,
respectively.
}
