% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av.R
\name{avworkspace_namespace}
\alias{avworkspace_namespace}
\alias{avworkspace_name}
\alias{avworkspace_bucket}
\alias{avtables}
\alias{avtable}
\title{Functions for convenient user interaction with AnVIL resources}
\usage{
avworkspace_namespace()

avworkspace_name()

avworkspace_bucket()

avtables(namespace = avworkspace_namespace(),
  name = avworkspace_name())

avtable(table, namespace = avworkspace_namespace(),
  name = avworkspace_name())
}
\arguments{
\item{namespace}{character(1) AnVIL workspace namespace as returned
by, e.g., `avworkspace_namespace()`}

\item{name}{character(1) AnVIL workspace name as returned by, eg.,
`avworkspace_name()`.}

\item{table}{character(1) table name as returned by, e.g., `avtables()`.}
}
\value{
`avtables()`: A tibble with columns identifying the table,
    the number of records, and the column names.

`avtable()`: a tibble of data corresponding to the AnVIL
    table `table` in the specified workspace.
}
\description{
Functions for convenient user interaction with AnVIL resources
}
