% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_async_collect.R
\name{av_async_collect}
\alias{av_async_collect}
\title{Execute Athena query asynchronously}
\usage{
av_async_collect(x)
}
\arguments{
\item{x}{A tbl or valid SQL character string.}
}
\value{
A query execution ID
}
\description{
Uses AWS API to send a query to Athena for execution. Returns the query
execution id for later retrieval of the results using \code{\link[=av_get_result]{av_get_result()}}.
}
\examples{
qid <-
  av_gt("data", "appends") \%>\%
  dplyr::filter(!is.na(household_id)) \%>\%
  head(50) \%>\%
  av_async_collect()
av_get_result(qid, .f = read.csv)

qid <- av_async_collect("SELECT * FROM data.appends LIMIT 50")
av_get_result(qid, .f = read.csv)
}
\seealso{
\itemize{
\item \code{\link[=av_get_result]{av_get_result()}} retrieves results from Athena using a execution ID.
\item \code{\link[=av_get_status]{av_get_status()}} returns the status of a query based on execution ID.
}

Other get data functions: 
\code{\link{av_get_query}()},
\code{\link{av_get_result}()},
\code{\link{av_get_status}()},
\code{\link{av_get_table}()}
}
\author{
Nicholas DiQuattro \email{nicholas@alloy.us}
}
\concept{get data functions}
