% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_connect.R
\name{av_connect}
\alias{av_connect}
\title{Connect to a remote database}
\usage{
av_connect(dsn = NULL, connection_string = NULL, set_default = TRUE, ...)
}
\arguments{
\item{dsn}{Character that corresponds to a DSN defined in \verb{~/.odbc.ini}.}

\item{connection_string}{Character. ODBC standard connection string.}

\item{set_default}{Boolean. Sets connection object as default for session. It
is stored in options(anvil.con).}

\item{...}{Additional ODBC name/value pairs. These will be joined with the
other arguments.}
}
\value{
A connection object
}
\description{
Creates a database connection from a given a Data Source Name (DSN) or ODBC
connection string.
}
\details{
Only set the \code{dsn} or \code{connection_string} arguments. Add addional arguments
with \code{...} if needed. If using \code{connection_string} it should take the
following format: \code{"<name>=<value>;<name>=<value>"}.
}
\examples{
# For a DSN named "athena"
con <- av_connect("athena")

# If you need a separate connection
con2 <- av_connect("other_db", set_default = FALSE)
}
\author{
Nicholas DiQuattro \email{nicholas@alloy.us}
}
\concept{connection functions}
