% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_drop_table.R
\name{av_drop_table}
\alias{av_drop_table}
\title{Drop a table from Athena}
\usage{
av_drop_table(name, schema = "data-science", con = getOption("anvil.con"))
}
\arguments{
\item{name}{Character. Name of table to drop.}

\item{schema}{Character. Name of the schema belonging to the table.}

\item{con}{Object. Database connection to use. Defaults to most recent call
of \code{\link[=av_connect]{av_connect()}}.}
}
\description{
Removes a table from Athena.
}
\examples{
av_drop_table("anvil_test")
}
\seealso{
Other table modifiers: 
\code{\link{av_create_table}()}
}
\author{
Nicholas DiQuattro \email{nicholas@alloy.us}
}
\concept{table modifiers}
