% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_get_result.R
\name{av_get_result}
\alias{av_get_result}
\title{Get previously executed Athena query}
\usage{
av_get_result(query_id, file, .f, ...)
}
\arguments{
\item{query_id}{Character. Athena query id from the history page on the
console.}

\item{file}{Character. File name and location to save local copy.}

\item{.f}{Function. A function that will read results into memory. Likely a
csv-reader.}

\item{...}{arguments supplied to \code{.f}.}
}
\value{
The saved file path or results of the provided function.
}
\description{
Either save a local copy of query results from S3 or read into memory.
}
\details{
When defining \code{.f} a local copy is still temporarily saved to a tmp
directory. You may want to change the location of where R makes tempfiles if
this will fill your instance's space. See \code{help("tempdir")} details.
}
\examples{
# Save local file
tq <- av_get_result("09b03bf1-f596-4cbe-8c04-06f380d4f3f1", "/data/test.csv")
read.csv(tq)

# Load data into memory -- No saving
av_get_result("09b03bf1-f596-4cbe-8c04-06f380d4f3f1", .f = read.csv)

# Alternately, load data into memory with fread
av_get_result("09b03bf1-f596-4cbe-8c04-06f380d4f3f1", .f = fread, na.strings = "NA")

}
\seealso{
\itemize{
\item \code{\link[=av_async_collect]{av_async_collect()}} returns an execution ID before a query completes.
\item \code{\link[=av_get_status]{av_get_status()}} returns the status of a query based on execution ID.
}

Other get data functions: 
\code{\link{av_async_collect}()},
\code{\link{av_get_query}()},
\code{\link{av_get_status}()},
\code{\link{av_get_table}()}
}
\author{
Nicholas DiQuattro \email{nicholas@alloy.us}
}
\concept{get data functions}
