% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_get_status.R
\name{av_get_status}
\alias{av_get_status}
\title{Check Athena query status}
\usage{
av_get_status(query_id)
}
\arguments{
\item{query_id}{Character. Query Execution ID to check.}
}
\value{
Character.
}
\description{
Query for the status of an Athena query.
}
\details{
Returns values could be "SUCCEEDED", "FAILED", "CANCELLED", or "RUNNING".
}
\examples{
av_get_status("8ba4e41e-3ad4-486c-a351-e8b62b1fd3f2")  # Succeeded
av_get_status("485316e3-931c-4288-b833-c8b3429e688d")  # Failed
}
\seealso{
\itemize{
\item \code{\link[=av_async_collect]{av_async_collect()}} returns an execution ID before a query completes.
\item \code{\link[=av_get_status]{av_get_status()}} returns the status of a query based on execution ID.
}

Other get data functions: 
\code{\link{av_async_collect}()},
\code{\link{av_get_query}()},
\code{\link{av_get_result}()},
\code{\link{av_get_table}()}
}
\author{
Nicholas DiQuattro \email{nicholas@alloy.us}
}
\concept{get data functions}
