% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_get_table.R
\name{av_get_table}
\alias{av_get_table}
\alias{av_gt}
\title{Get Table}
\usage{
av_get_table(schema, table, full_name = NULL, con = getOption("anvil.con"))

av_gt(schema, table, full_name = NULL, con = getOption("anvil.con"))
}
\arguments{
\item{schema}{Character. Name of desired schema.}

\item{table}{Character. Name of desired table.}

\item{full_name}{Character. Full name of a table in the form of \verb{<schema>.<table>}.}

\item{con}{Object. Database connection to use. Defaults to most recent call
of \code{\link[=av_connect]{av_connect()}}.}
}
\value{
A tbl.
}
\description{
Queries for data from a given schema tablename.
}
\examples{
# Things are quotes automatically
av_get_table("rds-export-graphite", "graphite_address") # No extra quotes!

# Run collect to bring a full dataset into memory
library(dplyr)
sample <-
  av_get_table("rds-export-graphite", "graphite_address") \%>\%
  head(20) \%>\% # Same as LIMIT 20
  dplyr::collect()

# Results can be joined with other dplyr verbs for manipulation
av_get_table("rds-export-graphite", "graphite_address") \%>\%
  dplyr::filter(state == "CA") \%>\%
  dplyr::select(state, starts_with("address"))

# Copy and paste a full name
av_gt(full_name = "data-science.data_avail")
av_gt("rds-export-graphite", "graphite_address")
}
\seealso{
Other get data functions: 
\code{\link{av_async_collect}()},
\code{\link{av_get_query}()},
\code{\link{av_get_result}()},
\code{\link{av_get_status}()}
}
\author{
Nicholas DiQuattro \email{nicholas@alloy.us}
}
\concept{get data functions}
