% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_ls_results.R
\name{av_ls_results}
\alias{av_ls_results}
\title{List previous Athena queries}
\usage{
av_ls_results(n = 50)
}
\arguments{
\item{n}{Integer. Number of queries to return. Should be in multiples of 50.}
}
\value{
A tibble.
}
\description{
Explore recent Athena queries
}
\details{
Beware that larger values of n may result in incomplete results due to rate
limiting from AWS.
}
\examples{
av_ls_results(150)
}
\seealso{
Other list functions: 
\code{\link{av_ls_tables}()}
}
\author{
Nicholas DiQuattro \email{nicholas@alloy.us}
}
\concept{list functions}
