% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_ls_s3.R
\name{av_ls_s3}
\alias{av_ls_s3}
\title{List S3 buckets and contents}
\usage{
av_ls_s3(s3path = NULL, max = 100, list_all = FALSE, ...)
}
\arguments{
\item{s3path}{s3:// URI or bucket/prefix style character.}

\item{max}{Number of results to return.}

\item{list_all}{When true will list entire key of each file.}

\item{...}{Other parameters passed down to \code{\link[aws.s3:get_bucket]{aws.s3::get_bucket_df()}}.}
}
\value{
A tibble
}
\description{
Find S3 buckets you can access and the contents within a bucket.
}
\details{
If \code{s3path} is left as \code{NULL} a tibble of available buckets will be returned.
When \code{s3path} is defined the contents at the given path will be returned.
}
\examples{
# Calling with no arguments returns a list of buckets
av_ls_s3()

# A bucket name will show top-level folders and root files
av_ls_s3("alloy-datascience")

# Add slashes to dive deeper, like on Linux
av_ls_s3("alloy-datascience/models/")

# s3:// url schemes can be present or not
av_ls_s3("s3://alloy-datascience/models/")

# Number of responses are limited by default to keep responses snappy, adjust with max
av_ls_s3("alloy-datascience", max = 700, list_all = TRUE)
}
