% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/av_sample.R
\name{av_sample}
\alias{av_sample}
\title{Randomly sample rows}
\usage{
av_sample(.tbl, pct = 1)
}
\arguments{
\item{.tbl}{A tbl_sql.}

\item{pct}{Double. Desired proportion of full table to randomly return.}
}
\value{
A tbl_sql.
}
\description{
Sample a percentage of an Athena table
}
\details{
This uses \verb{TABLESAMPLE BERNOULLI()} under the hood on the current query
specified by the \code{tbl}. The number of rows returned is not deterministic.
That is, asking for .1 of a 1,000 row table may not be equal to 100 rows
every run.
}
\examples{
# Sample 10\% of Colorado
av_gt("data-science", "graphite_sample") \%>\%
  dplyr::filter(registration_state == "CO") \%>\%
  av_sample(.1)
}
\author{
Nicholas DiQuattro \email{nicholas@alloy.us}
}
