% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avnotebooks.R
\name{avnotebooks}
\alias{avnotebooks}
\alias{avnotebooks_localize}
\alias{avnotebooks_delocalize}
\title{Notebook management}
\usage{
avnotebooks(
  local = FALSE,
  namespace = avworkspace_namespace(),
  name = avworkspace_name()
)

avnotebooks_localize(
  destination,
  namespace = avworkspace_namespace(),
  name = avworkspace_name(),
  dry = TRUE
)

avnotebooks_delocalize(
  source,
  namespace = avworkspace_namespace(),
  name = avworkspace_name(),
  dry = TRUE
)
}
\arguments{
\item{local}{= \code{logical(1)} notebooks located on the workspace
(\code{local = FALSE}, default) or runtime / local instance (\code{local = TRUE}). When \code{local = TRUE}, the notebook path is
\verb{<avworkspace_name>/notebooks}.}

\item{namespace}{character(1) AnVIL workspace namespace as returned
by, e.g., \code{avworkspace_namespace()}}

\item{name}{character(1) AnVIL workspace name as returned by, eg.,
\code{avworkspace_name()}.}

\item{destination}{missing or character(1) file path to the local
file system directory for synchronization. The default location
is \verb{~/<avworkspace_name>/notebooks}. Out-of-date local files
are replaced with the workspace version.}

\item{dry}{\code{logical(1)}, when \code{TRUE} (default), return the
consequences of the operation without actually performing the
operation.}

\item{source}{missing or character(1) file path to the local file
system directory for synchronization. The default location is
\verb{~/<avworkspace_name>/notebooks}. Out-of-date local files are
replaced with the workspace version.}
}
\value{
\code{avnotebooks()} returns a character vector of buckets /
files located in the workspace 'Files/notebooks' bucket path,
or on the local file system.

\code{avnotebooks_localize()} returns the exit status of
\code{gsutil_rsync()}.

\code{avnotebooks_delocalize()} returns the exit status of
\code{gsutil_rsync()}.
}
\description{
\code{avnotebooks()} returns the names of the notebooks
associated with the current workspace.

\code{avnotebooks_localize()} synchronizes the content of
the workspace bucket to the local file system.

\code{avnotebooks_delocalize()} synchronizes the content of
the notebook location of the local file system to the workspace
bucket.
}
\examples{
if (gcloud_exists() && nzchar(avworkspace_name()))
    avnotebooks()

if (gcloud_exists() && nzchar(avworkspace_name()))
    avnotebooks_localize()  # dry run

if (gcloud_exists() && nzchar(avworkspace_name()))
    try(avnotebooks_delocalize())  # dry run, fails if no local resource

}
