% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avworkflow.R
\name{avworkflow_jobs}
\alias{avworkflow_jobs}
\alias{avworkflow_files}
\title{Working with AnVIL Workflow submissions and file outputs}
\usage{
avworkflow_jobs(namespace = avworkspace_namespace(), name = avworkspace_name())

avworkflow_files(submissionId = NULL, bucket = avbucket())
}
\arguments{
\item{namespace}{character(1) AnVIL workspace namespace as returned
by, e.g., \code{avworkspace_namespace()}}

\item{name}{character(1) AnVIL workspace name as returned by, eg.,
\code{avworkspace_name()}.}

\item{submissionId}{a character() of workflow submission ids, or a
tibble with column \code{submissionId}, or NULL / missing. See
'Details'.}

\item{bucket}{character(1) name of the google bucket in which the
workflow products are available, as \verb{gs://...}. Usually the
bucket of the active workspace, returned by \code{avbucket()}.}
}
\value{
\code{avworkflow_jobs()} returns a tibble, sorted by
\code{submissionDate}, with columns
\itemize{
\item submissionId character() job identifier from the workflow runner.
\item submitter character() AnVIL user id of individual submitting the job.
\item submissionDate POSIXct() date (in local time zone) of job submission.
\item succeeded integer() number of workflows succeeding.
\item failed integer() number of workflows failing.
}

\code{avworkflow_files()} returns a tibble with columns
\itemize{
\item file: character() 'base name' of the file in the bucket.
\item workflow: character() name of the workflow the file is associated
with.
\item task: character() name of the task in the workflow that generated
the file.
\item path: charcter() full path to the file in the google bucket.
}
}
\description{
\code{avworkflow_jobs()} returns a tibble summarizing
submitted workflow jobs for a namespace and name.

\code{avworkflow_files()} returns a tibble containing
information and file paths to workflow outputs.
}
\details{
For \code{avworkflow_files()}, the \code{submissionId} is the
identifier associated with the workflow job, and is present in
the return value of \code{avworkflow_jobs()}; the example
illustrates how the first row of \code{avworkflow_jobs()} (i.e., the
most recenltly completed workflow) can be used as input to
\code{avworkflow_files()}. When \code{submissionId} is not provided, the
return value is for the most recently submitted workflow of the
namespace and name of \code{avworkspace()}.
}
\examples{
if (gcloud_exists() && nzchar(avworkspace_name()))
    ## from within AnVIL
    avworkflow_jobs()

if (gcloud_exists() && nzchar(avworkspace_name())) {
    ## e.g., from within AnVIL
    avworkflow_jobs() \%>\%
    ## select most recent workflow
    head(1) \%>\%
    ## find paths to output and log files on the bucket
    avworkflow_files()
}
}
