% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avworkspace.R
\name{avworkspace}
\alias{avworkspace}
\alias{avworkspace_clone}
\title{Workspace management}
\usage{
avworkspace_clone(
  namespace = avworkspace_namespace(),
  name = avworkspace_name(),
  to_namespace = namespace,
  to_name,
  bucket_location = "US"
)
}
\arguments{
\item{namespace}{character(1) AnVIL workspace namespace as returned
by, e.g., \code{avworkspace_namespace()}}

\item{name}{character(1) AnVIL workspace name as returned by, eg.,
\code{avworkspace_name()}.}

\item{to_namespace}{character(1) workspace (billing account) in
which to make the clone.}

\item{to_name}{character(1) name of the cloned workspace.}

\item{bucket_location}{character(1) region (NO multi-region, except
the default) in which bucket attached to the workspace should
be created.}
}
\value{
\code{avworkspace_clone()} returns the namespace and name, in
the format \code{namespace/name}, of the cloned workspace.
}
\description{
\code{avworkspace_clone()} clones (copies) an existing
workspace, possibly into a new namespace (billing account).
}
