% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{BackendConfig}
\alias{BackendConfig}
\title{Create a backend configuration}
\usage{
BackendConfig(...)
}
\arguments{
\item{...}{Named functions implementing the backend interface.
Required fields: \code{constructor}, \code{dtype}, \code{shape}, \code{ambiguous},
\code{as_array}, \code{as_raw}, \code{platform}, \code{device}, \code{print_data}.}
}
\value{
A \code{BackendConfig} object.
}
\description{
Create a backend configuration
}
\keyword{internal}
