% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-builder.R
\name{Builder}
\alias{Builder}
\title{Here we need reference semantic, because this will be a global variable.
The problem is that we can't have one builder always at the bottom of the stack
Because}
\description{
Here we need reference semantic, because this will be a global variable.
The problem is that we can't have one builder always at the bottom of the stack
Because

Here we need reference semantic, because this will be a global variable.
The problem is that we can't have one builder always at the bottom of the stack
Because
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Builder-new}{\code{Builder$new()}}
\item \href{#method-Builder-add_equation}{\code{Builder$add_equation()}}
\item \href{#method-Builder-add_variable}{\code{Builder$add_variable()}}
\item \href{#method-Builder-get_variable}{\code{Builder$get_variable()}}
\item \href{#method-Builder-add_constant}{\code{Builder$add_constant()}}
\item \href{#method-Builder-build}{\code{Builder$build()}}
\item \href{#method-Builder-new_box}{\code{Builder$new_box()}}
\item \href{#method-Builder-clone}{\code{Builder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Builder-new"></a>}}
\if{latex}{\out{\hypertarget{method-Builder-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Builder$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Builder-add_equation"></a>}}
\if{latex}{\out{\hypertarget{method-Builder-add_equation}{}}}
\subsection{Method \code{add_equation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Builder$add_equation(equation)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Builder-add_variable"></a>}}
\if{latex}{\out{\hypertarget{method-Builder-add_variable}{}}}
\subsection{Method \code{add_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Builder$add_variable(box)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Builder-get_variable"></a>}}
\if{latex}{\out{\hypertarget{method-Builder-get_variable}{}}}
\subsection{Method \code{get_variable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Builder$get_variable(box)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Builder-add_constant"></a>}}
\if{latex}{\out{\hypertarget{method-Builder-add_constant}{}}}
\subsection{Method \code{add_constant()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Builder$add_constant(box, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Builder-build"></a>}}
\if{latex}{\out{\hypertarget{method-Builder-build}{}}}
\subsection{Method \code{build()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Builder$build(in_boxes, out_boxes)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Builder-new_box"></a>}}
\if{latex}{\out{\hypertarget{method-Builder-new_box}{}}}
\subsection{Method \code{new_box()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Builder$new_box(interpreter, aval)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Builder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Builder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Builder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
