% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{compare_proxy.AnvilTensor}
\alias{compare_proxy.AnvilTensor}
\title{Waldo Comparison Proxy for AnvilTensor}
\usage{
compare_proxy.AnvilTensor(x, path)
}
\description{
Waldo Comparison Proxy for AnvilTensor
}
\keyword{internal}
