% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{data}
\name{data_types}
\alias{data_types}
\alias{dt_i1}
\alias{dt_i8}
\alias{dt_i16}
\alias{dt_i32}
\alias{dt_i64}
\alias{dt_ui8}
\alias{dt_ui16}
\alias{dt_ui32}
\alias{dt_ui64}
\alias{dt_f32}
\alias{dt_f64}
\title{Tensor Data Types}
\description{
Data types for tensors:
\itemize{
\item \code{dt_i1}: boolean.
\item \verb{dt_i\{8, 16, 32, 64\}}: signed integer.
\item \verb{dt_ui\{8, 16, 32, 64\}}: unsigned integer.
\item \verb{dt_f\{32, 64\}}: float.
}
}
\keyword{datasets}
