% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-amazon_athena.R
\name{db_data_type.amazon_athena}
\alias{db_data_type.amazon_athena}
\title{Convert R data type to Athena}
\usage{
\method{db_data_type}{amazon_athena}(con, fields, ...)
}
\arguments{
\item{con}{Athena connection}

\item{fields}{fields to type reference}

\item{...}{ignored}
}
\description{
Convert R data type to Athena
}
\keyword{internal}
