% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{DebugBox}
\alias{DebugBox}
\title{Abstract Value Box Class}
\usage{
DebugBox(aval)
}
\arguments{
\item{aval}{(\code{AbstractTensor})\cr
The abstract tensor representing the value.}
}
\value{
(\code{DebugBox})
}
\description{
Box representing a value in debug mode.
}
