% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize.R
\name{delocalize}
\alias{delocalize}
\title{Delocalize a Google storage bucket from the container.}
\usage{
delocalize(source, destination, remove_local_volume = FALSE, ...)
}
\arguments{
\item{source}{character vector, representing a local path}

\item{destination}{character, a valid path to a google storage
bucket.}

\item{remove_local_volume}{logical, remove the local mount volume.}

\item{...}{arguments for gsutil_rsync like, delete, recursive,
parallel.}
}
\value{
Exit status of function, `gsutil_rsync()`
}
\description{
Delocalize a Google storage bucket from the container.
}
\details{
delocalize the volume which is mounted with the help of
    gsutil. This syncs the library a final time and then removes
    the volume from the directory. Once a path is delocalized, it
    has to be localized again to regain access.
}
\examples{

\dontrun{
    delocalize("/tmp/test-localize/", "gs://anvil-bioc")
}

}
