% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-generics.R
\name{diag}
\alias{diag}
\title{Matrix Diagonals}
\usage{
diag(x = 1, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Further arguments passed to methods.}
}
\description{
S3 generic dispatching to \code{\link[=nv_diag]{nv_diag()}} (1-D operand) or
\code{\link[=nv_extract_diag]{nv_extract_diag()}} (2-D operand) for \code{\link{arrayish}} inputs, and
\code{\link[base:diag]{base::diag()}} otherwise.
}
