% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{.current_descriptor}
\alias{.current_descriptor}
\title{Get the current graph}
\usage{
.current_descriptor(silent = FALSE)
}
\arguments{
\item{silent}{(\code{logical(1)})\cr
Whether to return \code{NULL} if no graph is currently being built (as opposed to aborting).}
}
\value{
A \code{\link{GraphDescriptor}} object.
}
\description{
Get the current graph being built (via \code{\link{local_descriptor}}).
}
