% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{.gsutil_do}
\alias{.gsutil_do}
\title{helper function to pass arguments to `gsutil`}
\usage{
.gsutil_do(args)
}
\arguments{
\item{args}{arguments to helper function}
}
\description{
helper function to pass arguments to `gsutil`
}
\details{
private helper function to pass arguments to `gsutil`
    command line utility and throw the appropriate errors.
}
