% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{dtype}
\alias{dtype}
\title{Get the data type of a tensor}
\usage{
dtype(x, ...)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
A tensor-like object.}

\item{...}{Additional arguments passed to methods (unused).}
}
\value{
A \code{\link[tengen:TensorDataType]{TensorDataType}}.
}
\description{
Returns the data type of a tensor (e.g. \code{f32}, \code{i64}).
}
\details{
This is implemented via the generic \code{\link[tengen:dtype]{tengen::dtype()}}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_tensor(1:4, dtype = "f32")
dtype(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[tengen:dtype]{tengen::dtype()}}
}
