% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{ensure_nv_tensor}
\alias{ensure_nv_tensor}
\title{Create an AnvilTensor struct}
\usage{
ensure_nv_tensor(x, ambiguous = FALSE)
}
\arguments{
\item{x}{The underlying PJRT buffer (or an existing AnvilTensor)}

\item{ambiguous}{Whether the dtype is ambiguous}
}
\value{
AnvilTensor struct
}
\description{
Create an AnvilTensor struct
}
\keyword{internal}
