% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{eq_type}
\alias{eq_type}
\alias{neq_type}
\title{Compare AbstractTensor Types}
\usage{
eq_type(e1, e2, ambiguity)

neq_type(e1, e2, ambiguity)
}
\arguments{
\item{e1}{(\code{\link{AbstractTensor}})\cr
First tensor to compare.}

\item{e2}{(\code{\link{AbstractTensor}})\cr
Second tensor to compare.}

\item{ambiguity}{(\code{logical(1)})\cr
Whether to consider the ambiguous field when comparing.
If \code{TRUE}, tensors with different ambiguity are not equal.
If \code{FALSE}, only dtype and shape are compared.}
}
\value{
\code{logical(1)} - \code{TRUE} if the tensors are equal, \code{FALSE} otherwise.
}
\description{
Compare two AbstractTensors for type equality.
}
