% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud.R
\name{gcloud}
\alias{gcloud}
\alias{gcloud_exists}
\alias{gcloud_account}
\alias{gcloud_project}
\alias{gcloud_help}
\alias{gcloud_cmd}
\title{gcloud command line utility interface}
\usage{
gcloud_exists()

gcloud_account(account = NULL)

gcloud_project(project = NULL)

gcloud_help(...)

gcloud_cmd(cmd, ...)
}
\arguments{
\item{account}{character(1) Google account (e.g., `user@gmail.com`)
to use for authentication.}

\item{project}{character(1) billing project name.}

\item{...}{Additional arguments appended to gcloud commands.}

\item{cmd}{`character(1)` representing a command used to evaluate
`gcloud cmd ...`.}
}
\value{
`gcloud_exists()` returns `TRUE` when the `gcloud`
    application can be found, FALSE otherwise.

`gcloud_account()` returns a `character(1)` vector
    containing the active gcloud account, typically a gmail email
    address.

`gcloud_project()` returns a `character(1)` vector
    containing the active gcloud project.

`gcloud_help()` returns an unquoted `character()` vector
    representing the text of the help manual page returned by
    `gcloud help ...`.

`gcloud_cmd()` returns a `character()` vector representing
    the text of the output of `gcloud cmd ...`
}
\description{
These functions invoke the `gcloud` command line
    utility. See \link{gsutil} for details on how `gcloud` is
    located.

`gcloud_exists()` tests whether the `gcloud()` command
    can be found on this system. See 'Details' section of `gsutil`
    for where the application is searched.

`gcloud_account()`: report the current gcloud account
    via `gcloud config get-value account`.

`gcloud_project()`: report the current gcloud project
    via `gcloud config get-value project`.

`gcloud_help()`: queries `gcloud` for help for a
    command or sub-comand via `gcloud help ...`.

`gcloud_cmd()` allows arbitrary `gcloud` command
    execution via `gcloud ...`. Use pre-defined functions in
    preference to this.
}
\examples{
gcloud_exists()

if (gcloud_exists())
    gcloud_account()

if (gcloud_exists())
    gcloud_help()

}
