% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{Graph}
\alias{Graph}
\title{Graph of Primitive Calls}
\usage{
Graph(
  calls = list(),
  in_tree = NULL,
  out_tree = NULL,
  inputs = list(),
  outputs = list(),
  constants = list()
)
}
\description{
Computational graph consisting exclusively of primitive calls.
}
\section{Fields}{

\itemize{
\item \code{calls} :: (\code{list(PrimitiveCall)})\cr
The primitive calls that make up the graph.
This can also be another call into a graph when the primitive is a \code{p_call}.
\item \code{in_tree} :: (\code{NULL | Node})\cr
The tree of inputs.
\item \code{out_tree} :: (\code{NULL | Node})\cr
The tree of outputs.
\item \code{inputs} :: (\code{list(GraphValue)})\cr
The inputs to the graph.
\item \code{outputs} :: (\code{list(GraphValue)})\cr
The outputs of the graph.
}
}

