% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{GraphBox}
\alias{GraphBox}
\title{Graph Box}
\usage{
GraphBox(gnode = `mut<GraphValue>`(), desc = `mut<GraphDescriptor>`())
}
\arguments{
\item{gnode}{(\code{GraphNode})\cr
The node representing the value in the graph.}

\item{desc}{(\code{GraphDescriptor})\cr
The descriptor of the graph.}
}
\description{
Box representing a value in a \code{\link{Graph}}.
}
