% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graphify}
\alias{graphify}
\title{Graphify an R function}
\usage{
graphify(f, args, desc = NULL)
}
\arguments{
\item{f}{(\code{function})\cr
The function to graphify.}

\item{args}{(\code{list})\cr
The arguments to the function.}

\item{desc}{(\code{NULL} | \code{GraphDescriptor})\cr
The descriptor to use for the graph.}
}
\description{
Graphify an R function.
}
