% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{GraphLiteral}
\alias{GraphLiteral}
\title{Graph Literal}
\usage{
GraphLiteral(aval = NULL, dtype = stablehlo::BooleanType())
}
\description{
Literal in a \code{\link{Graph}}.
}
\section{Fields}{

\itemize{
\item \code{aval} :: (\code{any})\cr
The value of the literal.
\item \code{dtype} :: (\code{stablehlo::TensorDataType})\cr
The dtype of the literal.
}
}

