% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\docType{data}
\name{GraphNode}
\alias{GraphNode}
\title{Graph Node}
\format{
An object of class \code{S7_union} of length 1.
}
\usage{
GraphNode
}
\description{
Node in a \code{\link{Graph}}.
Is either a \code{\link{GraphValue}} or a \code{\link{GraphLiteral}}.
}
\keyword{datasets}
