% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{GraphTransformation}
\alias{GraphTransformation}
\title{Graph Transformation}
\usage{
GraphTransformation(input = function() NULL)
}
\description{
Abstract base class for a (chained) graph transformation.
To apply such a (chained) transformation, use \code{\link[=apply_transform]{apply_transform()}}.
}
