% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{GraphValue}
\alias{GraphValue}
\title{Graph Value}
\usage{
GraphValue(aval = ShapedTensor())
}
\description{
Value in a \code{\link{Graph}}.
}
\section{Fields}{

\itemize{
\item \code{state} :: (\code{any})\cr
The state of the variable. Populated when the graph is executed.
}
}

