% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsutil.R
\name{gsutil}
\alias{gsutil}
\alias{gsutil_ls}
\alias{gsutil_cp}
\alias{gsutil_rm}
\alias{gsutil_rsync}
\title{Internal functions to interact with the gsutil command line}
\usage{
gsutil_ls(google_bucket, path = "", recursive = TRUE)

gsutil_cp(google_bucket, path, recursive = TRUE)

gsutil_rm(google_bucket, sub_directory, recursive = TRUE)

gsutil_rsync(source, destination, match = TRUE, recursive = TRUE)
}
\arguments{
\item{google_bucket}{character(1), a valid path to a google storage
bucket}

\item{path}{character(1), a path or regular expression listing files
of paths.}

\item{recursive}{logical(1), should the operation be recursive?}

\item{sub_directory}{character(1), representing a sub directory
path inside the google bucket.}

\item{source}{character(1), a path to a source url, either a google
bucket or a local path}

\item{destination}{character(1), a path to a destination url, either a
google bucket or a local path.}

\item{match}{logical(1), a flag which matches the source to the
    destination exactly. (rsync -d)

    The rsync -d option is very useful and commonly used, because
    it provides a means of making the contents of a destination
    bucket or directory match those of a source bucket or
    directory. However, please exercise caution when you use this
    option: It's possible to delete large amounts of data
    accidentally if, for example, you erroneously reverse source
    and destination.}
}
\value{
All functions return the exit status of the corresponding
    `gsutil` command, invisibly.
}
\description{
These are internal (non-exported) helper functions to
    pass arguments to `gsutil` command line utility. Functions
    either succeed or return the appropriate error.
}
\details{
To make gs://mybucket/data match the contents of the local
directory "data" you could do:

\code{gsutil_rsync(data, "gs://mybucket/data", match = TRUE)}

To make the local directory "data" the same as the contents of
gs://mybucket/data:

\code{gsutil_rsync("gs://mybucket/data", "data", match = TRUE)}
}
