% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{gsutil_cp}
\alias{gsutil_cp}
\title{Copy contents of a gcs bucket}
\usage{
gsutil_cp(google_bucket, target_path, recursive = TRUE)
}
\arguments{
\item{google_bucket}{character, a valid path to a google storage
bucket}

\item{recursive}{logical, if the operation should be recursive}

\item{local_path}{character vector, representing a local path}
}
\value{
Exit status of `gsutil_cp()`, invisibly
}
\description{
Copy contents of a gcs bucket
}
