% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{gsutil_ls}
\alias{gsutil_ls}
\title{List contents of a gcs bucket}
\usage{
gsutil_ls(google_bucket, path = "", recursive = TRUE)
}
\arguments{
\item{google_bucket}{character, a valid path to a google storage
bucket}

\item{path}{character, a path or regular expression listing files
of paths.}

\item{recursive}{logical, if the operation should be recursive}
}
\value{
Exit status of `gsutil_ls()`, invisibly
}
\description{
List contents of a gcs bucket
}
