% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{gsutil_rm}
\alias{gsutil_rm}
\title{Remove contents of a gcs bucket}
\usage{
gsutil_rm(google_bucket, sub_directory, recursive = TRUE)
}
\arguments{
\item{google_bucket}{character, a valid path to a google storage
bucket}

\item{sub_directory}{character vector, representing a sub directory
path inside the google bucket}

\item{recursive}{logical, if the operation should be recursive}
}
\value{
Exit status of `gsutil_rm()`, invisibly
}
\description{
Remove contents of a gcs bucket
}
