% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{gsutil_rsync}
\alias{gsutil_rsync}
\title{rsync contents of a gcs bucket}
\usage{
gsutil_rsync(source, destination, match = TRUE, recursive = TRUE)
}
\arguments{
\item{source}{character, a path to a source url. Either a google
bucket or a local path}

\item{destination}{character, a path to a destination url. Either a
google bucket or a local path.}

\item{match}{logical, a flag which matches the source to the
    destination exactly. (rsync -d)

    The rsync -d option is very useful and commonly used, because
    it provides a means of making the contents of a destination
    bucket or directory match those of a source bucket or
    directory. However, please exercise caution when you use this
    option: It's possible to delete large amounts of data
    accidentally if, for example, you erroneously reverse source
    and destination.}

\item{recursive}{logical, if the operation should be recursive}
}
\value{
Exit status of `gsutil_rsync()`, invisbly
}
\description{
rsync contents of a gcs bucket
}
\details{
To make gs://mybucket/data match the contents of the local
    directory "data" you could do:
       `gsutil rsync -d data gs://mybucket/data`

To make the local directory "data" the same as the contents of
   gs://mybucket/data:
`gsutil rsync -d -r gs://mybucket/data data`
}
