% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{HigherOrderPrimitive}
\alias{HigherOrderPrimitive}
\title{HigherOrderPrimitive}
\usage{
HigherOrderPrimitive(name, subgraphs = character())
}
\arguments{
\item{name}{(\code{character()})\cr
The name of the primitive.}

\item{subgraphs}{(\code{character()})\cr
Names of parameters that are subgraphs.}
}
\value{
(\code{HigherOrderPrimitive})
}
\description{
A primitive that contains subgraphs.
}
