% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jit.R
\name{hlo}
\alias{hlo}
\title{Lower a function to StableHLO}
\usage{
hlo(f, static = character())
}
\arguments{
\item{f}{(\code{function} | \code{Gr})\cr
Flattened function to lower.}

\item{.avals}{(\code{list()})\cr
List of flattened abstract values (avals) for the function arguments.
ShapedTensors will be traced (wrapped in HloBox), while other values
will be passed through as-is (static).}
}
\value{
(\code{list}) with elements:
\itemize{
\item \code{func}: The StableHLO \code{Func} object
\item \code{out_tree}: The output tree structure
}
}
\description{
Immediately lower a flattened function to a StableHLO Func object.
}
