% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablehlo.R
\name{HloEnv}
\alias{HloEnv}
\title{HloEnv}
\usage{
HloEnv(parent = NULL, gval_to_fval = NULL)
}
\arguments{
\item{parent}{(\code{HloEnv} | \code{NULL})\cr
Parent environment for lookups.}

\item{gval_to_fval}{(\code{hashtab})\cr
Mapping from graph values to func values.}
}
\value{
(\code{HloEnv})
}
\description{
Environment for storing graph value to func value mappings.
This is a mutable class.
}
