% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-passes.R
\name{inline_literals}
\alias{inline_literals}
\title{Inline Literals Pass}
\usage{
inline_literals(graph)
}
\arguments{
\item{graph}{(\code{Graph})\cr
The graph to process.}
}
\value{
(\code{Graph})\cr
A new graph with scalar constants converted to literals.
}
\description{
Converts all scalar constants (shape \code{c()}) in a graph to \code{GraphLiteral} objects.
This pass also recursively processes subgraphs in higher-order primitives.
}
