% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize.R
\name{install}
\alias{install}
\title{Install libraries and it's dependencies from a bucket.}
\usage{
install(pkgs = character(), lib = .libPaths()[1], lib.loc = NULL)
}
\arguments{
\item{lib}{character, file path to .libPaths()[1], primary location
to install packages.}

\item{packages}{character, names of packages to install from a
bucket.}

\item{google_bucket}{character, a valid path to a google storage
bucket.}
}
\value{
character, location of packages installed invisibly
}
\description{
Install libraries and it's dependencies from a bucket.
}
\details{
Given a pre-existing google bucket with a host of
    bioconductor packages built on a specific matching version of R
    and Bioconductor, the user is able to install packages his
    container using the 'install()` function from the google
    bucket.

eg:

add_libs("/tmp/host-site-library")

install(packages = c('BiocParallel', ='BiocGenerics'),
            google_bucket = "bioc_release_volume",
            lib = .libPaths()[1])
}
