% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{is_tensorish}
\alias{is_tensorish}
\title{Check if an Object is a Tensor-ish Object}
\usage{
is_tensorish(x, literal = TRUE)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}

\item{literal}{(\code{logical(1)})\cr
Whether to allow R literals (i.e., \code{1L}, \code{1.0}, \code{TRUE}, etc.) to be considered tensor-ish.
Defaults to \code{TRUE}.}
}
\value{
\code{logical(1)}
}
\description{
Check if an object is a tensor-ish object.
}
