% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{life_cycle}
\alias{life_cycle}
\alias{.life_cycle}
\title{Internal helpers}
\usage{
.life_cycle(
  newfun = oldfun,
  newpackage,
  package = "AnVIL",
  cycle = c("deprecated", "defunct"),
  title = package
)
}
\arguments{
\item{newfun}{The name of the function to use instead. It can be a specific
function within another package (e.g., \code{AnVIL::avstorage}) or a function in
the current package (e.g., \code{avstorage}).}

\item{newpackage}{If a function is moved to a new package, the name of the
new package can be specified here. This is equivalent to specifying
\code{newfun = paste0(newpackage, "::", newfun)}.}

\item{package}{The name of the package where the deprecated function resides.}

\item{cycle}{The life cycle stage of the function. This can be either
\code{"deprecated"} or \code{"defunct"}.}

\item{title}{The Rd name prefix of the documentation page where the
deprecation is documented (e.g., "av" for "av-deprecated").}
}
\description{
These functions are not intended to be called directly by users.
They are used internally by the package.
}
\keyword{internal}
