% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{LiteralTensor}
\alias{LiteralTensor}
\title{Literal Tensor Class}
\usage{
LiteralTensor(data, shape, dtype = default_dtype(data), ambiguous)
}
\arguments{
\item{data}{(\code{double(1)} | \code{integer(1)} | \code{logical(1)} | \code{\link{AnvilTensor}})\cr
The scalar value or scalarish AnvilTensor (contains 1 element).}

\item{shape}{(\code{\link[stablehlo:Shape]{stablehlo::Shape}} | \code{integer()})\cr
The shape of the tensor.}

\item{dtype}{(\code{\link[tengen:TensorDataType]{tengen::TensorDataType}})\cr
The data type. Defaults to the current backend's default floating dtype,
\code{i32} for integer, and \code{bool} for logical.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the type is ambiguous.
Ambiguous types usually arise from R literals (e.g., \code{1L}, \code{1.0}) and follow special promotion rules.
See the \code{vignette("type-promotion")} for more details.}
}
\description{
An \code{\link{AbstractTensor}} where all elements have the same constant value.
This either arises when using literals in traced code (e.g. \code{x + 1}) or when using
\code{\link[=nv_fill]{nv_fill()}} to create a constant.
}
\section{Type Ambiguity}{

When arising from R literals, the resulting \code{LiteralTensor} is ambiguous because no type
information was available. See the \code{vignette("type-promotion")} for more details.
}

\section{Lowering}{

\code{LiteralTensor}s become constants inlined into the stableHLO program.
I.e., they lower to \code{\link[stablehlo:hlo_constant]{stablehlo::hlo_tensor()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
x <- LiteralTensor(1L, shape = integer(), ambiguous = TRUE)
x
ambiguous(x)
shape(x)
ndims(x)
dtype(x)
# How it appears during tracing:
# 1. via R literals
graph <- trace_fn(function() 1, list())
graph
graph$outputs[[1]]$aval
# 2. via nv_fill()
graph <- trace_fn(function() nv_fill(2L, shape = c(2, 2)), list())
graph
graph$outputs[[1]]$aval
\dontshow{\}) # examplesIf}
}
