% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize-deprecated.R
\name{localize-deprecated}
\alias{localize-deprecated}
\alias{localize}
\alias{delocalize}
\title{Copy packages, folders, or files to or from Google buckets
(DEPRECATED).}
\usage{
localize(source, destination, dry = TRUE)

delocalize(source, destination, unlink = FALSE, dry = TRUE)
}
\arguments{
\item{source}{\code{character(1)}, a google storage bucket or local file
system directory location.}

\item{destination}{\code{character(1)}, a google storage bucket or local
file system directory location.}

\item{dry}{\code{logical(1)}, when \code{TRUE} (default), return the
consequences of the operation without actually performing the
operation.}

\item{unlink}{\code{logical(1)} remove (unlink) the file or directory
in \code{source}. Default: \code{FALSE}.}
}
\value{
\code{localize()}: exit status of function \code{gsutil_rsync()}.

\code{delocalize()}: exit status of function \code{gsutil_rsync()}
}
\description{
\code{localize()}: recursively synchronizes files from a
Google storage bucket (\code{source}) to the local file system
(\code{destination}). This command acts recursively on the \code{source}
directory, and does not delete files in \code{destination} that are
not in `source.

\code{delocalize()}: synchronize files from a local file
system (\code{source}) to a Google storage bucket
(\code{destination}). This command acts recursively on the \code{source}
directory, and does not delete files in \code{destination} that are
not in \code{source}.
}
