% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_constring.R
\name{make_constring}
\alias{make_constring}
\title{Make a ODBC connection string from a DSN file}
\usage{
make_constring(dsn, dsn_file = "/etc/odbc.ini")
}
\arguments{
\item{dsn}{Name of target DSN.}

\item{dsn_file}{Path to ODBC DSN file.}
}
\value{
A connection string.
}
\description{
Converts a defined Data Source Name (DSN) into a connection string for use as an ENV.
}
\examples{
# Set env
Sys.setenv(ATHENA_CON = make_constring("athena"))

av_connect(connection_string = Sys.getenv("ATHENA_CON"))
}
\author{
Nicholas DiQuattro \email{nicholas@alloy.us}
}
