% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{maybe_box_input}
\alias{maybe_box_input}
\title{Maybe Box Input}
\usage{
maybe_box_input(x, desc, toplevel)
}
\arguments{
\item{x}{(any)\cr
The input to box.}

\item{desc}{(\code{\link{GraphDescriptor}})\cr
The descriptor of the graph.}

\item{toplevel}{(\code{logical(1)})\cr
Whether the function is being traced at the top level.
If this is \code{TRUE}, inputs that are \code{AnvilTensor}s are treated as unknown, i.e. as
\code{GraphValue} instead of \code{GraphLiteral} objects.}
}
\description{
This function is called onto the inputs of a function that is being traced via \code{\link[=trace_fn]{trace_fn()}}.
The function boxes the provided value and registers it as an input to the graph.
}
\keyword{internal}
