% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{maybe_box_tensorish}
\alias{maybe_box_tensorish}
\title{Maybe Box Tensorish}
\usage{
maybe_box_tensorish(x)
}
\description{
This function boxes tensorish values passed to a primitive.
Note that length-1 vectors are converted to scalar AnvilTensor objects.
}
\keyword{internal}
