% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut.R
\name{mut}
\alias{mut}
\title{Convert an \code{S7} class to a mutable \code{S7} object}
\usage{
mut(x)
}
\arguments{
\item{x}{\code{S7_class} an \code{S7_class} class constructor - the function you
would otherwise call to create a new object of that class.}
}
\description{
Convert an \code{S7} class to a mutable \code{S7} object
}
\examples{
library(S7)

class_ex <- new_class(
  "class_example",
  properties = list(
    # validators
    value = new_property(
      class = class_integer,
      validator = function(value) {
        if (value == 0L) "value cannot be exactly 0"
      }
    ),
    # read-only properties
    is_gt_0 = new_property(
      class = class_logical,
      getter = function(self) {
        self@value > 0
      }
    )
  )
)

# make a mutable version of our class
ex <- mut(class_ex)(value = 3L)
ex@value

# we can make a copy and update our value property
ex_ref <- ex
ex_ref@value <- 30L

# all values reference the same data, our original is updated
ex@value

}
