% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_acos}
\alias{nv_acos}
\title{Arc Cosine}
\usage{
nv_acos(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise inverse cosine. You can also use \code{acos()}.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(-1, 0, 1))
acos(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_acos]{prim_acos()}} for the underlying primitive.
}
