% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_acosh}
\alias{nv_acosh}
\title{Inverse Hyperbolic Cosine}
\usage{
nv_acosh(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise inverse hyperbolic cosine. You can also use \code{acosh()}.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(1, 2, 10))
acosh(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_acosh]{prim_acosh()}} for the underlying primitive.
}
