% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_add}
\alias{nv_add}
\title{Addition}
\usage{
nv_add(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise addition of two tensors.
}
