% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_and}
\alias{nv_and}
\title{Logical And}
\usage{
nv_and(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logical AND operation.
}
