% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_argmin}
\alias{nv_argmin}
\title{Index of the Minimum}
\usage{
nv_argmin(x, dim = NULL)
}
\arguments{
\item{x}{(\code{\link{arrayish}})\cr
The array.}

\item{dim}{(\code{integer(1)} | \code{NULL})\cr
Dimension along which to find the index. If \code{NULL} (default), uses
the last dimension.}
}
\value{
\code{\link{arrayish}} of dtype \code{i64}\cr
Same shape as \code{x} with \code{dim} removed.
}
\description{
Returns the index of the minimum value along a dimension. Ties are broken
by returning the smallest index.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
nv_argmin(nv_array(c(3, 1, 4, 1, 5, 9, 2, 6)))
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_argmax]{nv_argmax()}}, \code{\link[=nv_reduce_min]{nv_reduce_min()}}.
}
