% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_asinh}
\alias{nv_asinh}
\title{Inverse Hyperbolic Sine}
\usage{
nv_asinh(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise inverse hyperbolic sine. You can also use \code{asinh()}.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(-1, 0, 1))
asinh(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_asinh]{prim_asinh()}} for the underlying primitive.
}
