% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_atan}
\alias{nv_atan}
\title{Arc Tangent}
\usage{
nv_atan(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise inverse tangent. You can also use \code{atan()}.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(-1, 0, 1))
atan(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_atan]{prim_atan()}} for the underlying primitive.
}
