% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_atan2}
\alias{nv_atan2}
\title{Arctangent 2}
\usage{
nv_atan2(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise two-argument arctangent.
}
