% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_bessel_i1e}
\alias{nv_bessel_i1e}
\title{Exponentially Scaled Modified Bessel I_1}
\usage{
nv_bessel_i1e(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise exponentially scaled modified Bessel function of the
first kind, order 1: \code{bessel_i1e(x) = bessel_i1(x) * exp(-abs(x))}.
Currently does not support automatic differentiation.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(-1, 0, 1, 5))
nv_bessel_i1e(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_bessel_i1e]{prim_bessel_i1e()}} for the underlying primitive.
}
