% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_broadcast}
\alias{nv_broadcast}
\title{Broadcast}
\usage{
nv_broadcast(x, shape)
}
\arguments{
\item{x}{(\code{\link{nv_tensor}})}

\item{shape}{(\code{integer()})\cr
Output shape.}
}
\value{
(\code{\link{nv_tensor}})
}
\description{
Broadcast
}
