% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_broadcast_all}
\alias{nv_broadcast_all}
\title{Broadcast}
\usage{
nv_broadcast_all(...)
}
\arguments{
\item{...}{(\code{\link{nv_tensor}})}

\item{shape}{(\code{integer()})\cr
Output shape. If \code{NULL} (default),}
}
\value{
(\code{\link{nv_tensor}})
}
\description{
Broadcast
}
