% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_broadcast_scalars}
\alias{nv_broadcast_scalars}
\title{Broadcast Scalars to Common Shape}
\usage{
nv_broadcast_scalars(...)
}
\arguments{
\item{...}{(\code{\link{arrayish}})\cr
Arrays to broadcast. Scalars will be broadcast to the common non-scalar shape.}
}
\value{
(\code{list()} of \code{\link{arrayish}})\cr
List of broadcasted arrays.
}
\description{
Broadcast scalar arrays to match the shape of non-scalar arrays.
All non-scalar arrays must have the same shape.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_array(c(1, 2, 3))
  # scalar 1 is broadcast to shape [3]
  nv_broadcast_scalars(x, 1)
})
\dontshow{\}) # examplesIf}
}
