% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_cbrt}
\alias{nv_cbrt}
\title{Cube Root}
\usage{
nv_cbrt(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise cube root.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(1, 8, 27))
nv_cbrt(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_cbrt]{prim_cbrt()}} for the underlying primitive.
}
