% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_ceil}
\alias{nv_ceil}
\title{Ceiling}
\usage{
nv_ceil(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise ceiling (round toward positive infinity). You can also use \code{ceiling()}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1.2, 2.7, -1.5))
  ceiling(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_ceil]{nvl_ceil()}} for the underlying primitive.
}
