% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_clamp}
\alias{nv_clamp}
\title{Clamp}
\usage{
nv_clamp(min_val, operand, max_val)
}
\arguments{
\item{min_val}{(\code{\link{tensorish}})\cr
Minimum value.}

\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{max_val}{(\code{\link{tensorish}})\cr
Maximum value.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise clamp: max(min_val, min(operand, max_val)).
}
