% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_cos}
\alias{nv_cos}
\title{Cosine}
\usage{
nv_cos(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise cosine. You can also use \code{cos()}.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(0, pi / 2, pi))
cos(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_cosine]{prim_cosine()}} for the underlying primitive.
}
