% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_cosh}
\alias{nv_cosh}
\title{Hyperbolic Cosine}
\usage{
nv_cosh(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise hyperbolic cosine. You can also use \code{cosh()}.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(-1, 0, 1))
cosh(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_cosh]{prim_cosh()}} for the underlying primitive.
}
