% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_cosine}
\alias{nv_cosine}
\title{Cosine}
\usage{
nv_cosine(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise cosine.
}
