% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_digamma}
\alias{nv_digamma}
\title{Digamma}
\usage{
nv_digamma(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise digamma function (logarithmic derivative of the gamma
function). You can also use \code{digamma()}.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(0.5, 1, 2, 5))
digamma(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_digamma]{prim_digamma()}} for the underlying primitive.
}
