% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{nv_dtype}
\alias{nv_dtype}
\title{Create a TensorDataType}
\usage{
nv_dtype(x)
}
\arguments{
\item{x}{(any)\cr
Object convertible to a \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}} (via \code{\link[stablehlo:as_dtype]{stablehlo::as_dtype}})}
}
\value{
\code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}}
}
\description{
Create a \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}}.
}
