% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_exp}
\alias{nv_exp}
\title{Exponential}
\usage{
nv_exp(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise exponential. You can also use \code{exp()}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(0, 1, 2))
  exp(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_exp]{nvl_exp()}} for the underlying primitive.
}
