% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_expand}
\alias{nv_expand}
\title{Expand}
\usage{
nv_expand(operand, shape)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{shape}{(\code{integer()})\cr
Target shape.}
}
\value{
\code{\link{tensorish}}\cr
Has the given \code{shape} and the same data type as \code{operand}.
}
\description{
Broadcasts a tensor to a target shape, expanding dimensions of size 1.
This is an alias for \code{\link[=nv_broadcast_to]{nv_broadcast_to()}}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  nv_expand(x, shape = c(2, 3))
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_broadcast_to]{nv_broadcast_to()}}, \code{\link[=nv_broadcast_tensors]{nv_broadcast_tensors()}}
}
