% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_expm1}
\alias{nv_expm1}
\title{Exponential Minus One}
\usage{
nv_expm1(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise \code{exp(x) - 1}, more accurate for small \code{x}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(0, 0.001, 1))
  nv_expm1(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_expm1]{nvl_expm1()}} for the underlying primitive.
}
