% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_flip}
\alias{nv_flip}
\title{Flip}
\usage{
nv_flip(operand, dims)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dims}{(\code{integer()})\cr
Dimensions to reverse.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as \code{operand}.
}
\description{
Reverses the order of elements along specified dimensions.
This is an alias for \code{\link[=nv_reverse]{nv_reverse()}}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2, 3, 4, 5))
  nv_flip(x, dims = 1L)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_reverse]{nv_reverse()}}
}
