% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_floor}
\alias{nv_floor}
\title{Floor}
\usage{
nv_floor(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise floor (round toward negative infinity).
}
