% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_full}
\alias{nv_full}
\title{Constant}
\usage{
nv_full(value, shape, dtype = NULL)
}
\arguments{
\item{value}{(any)\cr
Value.}

\item{shape}{(integer())\cr
Shape.}

\item{dtype}{(character(1))\cr
Data type.}
}
\description{
Create a constant.
}
