% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_gt}
\alias{nv_gt}
\title{Greater Than}
\usage{
nv_gt(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise greater than comparison.
}
