% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_le}
\alias{nv_le}
\title{Less Than or Equal}
\usage{
nv_le(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise less than or equal comparison.
}
