% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_lgamma}
\alias{nv_lgamma}
\title{Log-Gamma}
\usage{
nv_lgamma(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise natural logarithm of the absolute value of the gamma
function. You can also use \code{lgamma()}.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(0.5, 1, 2, 5))
lgamma(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_lgamma]{prim_lgamma()}} for the underlying primitive.
}
