% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_log10}
\alias{nv_log10}
\title{Base-10 Logarithm}
\usage{
nv_log10(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise base-10 logarithm. You can also use \code{log10()}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 10, 100, 1000))
  nv_log10(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_log]{nv_log()}}, \code{\link[=nv_log2]{nv_log2()}}
}
