% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_lt}
\alias{nv_lt}
\title{Less Than}
\usage{
nv_lt(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.
Operands are \link[=nv_promote_to_common]{promoted to a common data type}.
Scalars are \link[=nv_broadcast_scalars]{broadcast} to the shape of the other operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape as the inputs and boolean data type.
}
\description{
Element-wise less than comparison. You can also use the \code{<} operator.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  y <- nv_tensor(c(3, 2, 1))
  x < y
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_lt]{nvl_lt()}} for the underlying primitive.
}
