% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_matmul}
\alias{nv_matmul}
\title{Matrix Multiplication}
\usage{
nv_matmul(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Tensors with at least 2 dimensions.
Operands are \link[=nv_promote_to_common]{promoted to a common data type}.}
}
\value{
\code{\link{tensorish}}
}
\description{
Matrix multiplication of two tensors. You can also use the \code{\%*\%} operator.
Supports batched matrix multiplication when inputs have more than 2 dimensions.
}
\section{Shapes}{

\itemize{
\item \code{lhs}: \verb{(b1, ..., bk, m, n)}
\item \code{rhs}: \verb{(b1, ..., bk, n, p)}
\item output: \verb{(b1, ..., bk, m, p)}
}
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(matrix(1:6, nrow = 2))
  y <- nv_tensor(matrix(1:6, nrow = 3))
  x \%*\% y
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_dot_general]{nvl_dot_general()}} for the underlying primitive.
}
