% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_max}
\alias{nv_max}
\title{Maximum}
\usage{
nv_max(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise maximum of two tensors.
}
