% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_min}
\alias{nv_min}
\title{Minimum}
\usage{
nv_min(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.
Operands are \link[=nv_promote_to_common]{promoted to a common data type}.
Scalars are \link[=nv_broadcast_scalars]{broadcast} to the shape of the other operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and the promoted common data type of the inputs.
}
\description{
Element-wise minimum of two tensors.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 5, 3))
  y <- nv_tensor(c(4, 2, 6))
  nv_min(x, y)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_min]{nvl_min()}} for the underlying primitive.
}
