% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_mul}
\alias{nv_mul}
\title{Multiplication}
\usage{
nv_mul(lhs, rhs)
}
\arguments{
\item{lhs}{(\code{\link{tensorish}})}

\item{rhs}{(\code{\link{tensorish}})}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise multiplication of two tensors.
}
