% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_ne}
\alias{nv_ne}
\title{Not Equal}
\usage{
nv_ne(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise inequality comparison.
}
