% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_negate}
\alias{nv_negate}
\title{Negation}
\usage{
nv_negate(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
}
\description{
Negates an array element-wise. You can also use the unary \code{-} operator.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(1, -2, 3))
-x
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_negate]{prim_negate()}} for the underlying primitive.
}
