% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_not}
\alias{nv_not}
\title{Logical Not}
\usage{
nv_not(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logical NOT operation.
}
