% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_or}
\alias{nv_or}
\title{Logical Or}
\usage{
nv_or(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logical OR operation.
}
