% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_outer}
\alias{nv_outer}
\title{Outer Product}
\usage{
nv_outer(x, y)
}
\arguments{
\item{x, y}{(\code{\link{tensorish}})\cr
1-D tensors.}
}
\value{
\code{\link{tensorish}}\cr
A 2-D tensor of shape \verb{(length(x), length(y))}.
}
\description{
Computes the outer product of two 1-D tensors.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  y <- nv_tensor(c(4, 5))
  nv_outer(x, y)
})
\dontshow{\}) # examplesIf}
}
