% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_pad}
\alias{nv_pad}
\title{Pad}
\usage{
nv_pad(
  operand,
  padding_value,
  edge_padding_low,
  edge_padding_high,
  interior_padding = NULL
)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{padding_value}{(\code{\link{tensorish}})\cr
Scalar value to use for padding. Must have the same dtype as \code{operand}.}

\item{edge_padding_low}{(\code{integer()})\cr
Amount of padding to add at the start of each dimension.}

\item{edge_padding_high}{(\code{integer()})\cr
Amount of padding to add at the end of each dimension.}

\item{interior_padding}{(\code{integer()} | \code{NULL})\cr
Amount of padding to add between elements in each dimension.
If \code{NULL} (default), no interior padding is applied.}
}
\value{
\code{\link{tensorish}}\cr
Has the same data type as \code{operand}.
}
\description{
Pads a tensor with a given value at the edges and optionally between elements.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  nv_pad(x, nv_scalar(0), edge_padding_low = 2L, edge_padding_high = 1L)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_pad]{nvl_pad()}} for the underlying primitive.
}
