% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_popcnt}
\alias{nv_popcnt}
\title{Population Count}
\usage{
nv_popcnt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise population count (number of set bits).
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(7L, 3L, 15L))
  nv_popcnt(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_popcnt]{nvl_popcnt()}} for the underlying primitive.
}
