% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_pow}
\alias{nv_pow}
\title{Power}
\usage{
nv_pow(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise exponentiation of two tensors.
}
