% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-rng.R
\name{nv_rbinom}
\alias{nv_rbinom}
\title{Sample from a Binomial Distribution}
\usage{
nv_rbinom(shape, initial_state, n = 1L, prob = 0.5, dtype = "i32")
}
\arguments{
\item{shape}{(\code{integer()})\cr
Shape.}

\item{initial_state}{(\code{\link{tensorish}})\cr
RNG state (\code{ui64[2]}).}

\item{n}{(\code{integer(1)})\cr
Number of trials.}

\item{prob}{(\code{numeric(1)})\cr
Probability of success on each trial.}

\item{dtype}{(\code{character(1)} | \code{\link[tengen:TensorDataType]{tengen::TensorDataType}})\cr
Data type.}
}
\value{
(\code{list()} of \code{\link{tensorish}})\cr
List of two elements: the updated RNG state and the sampled values.
}
\description{
Samples from a binomial distribution with \eqn{n} trials and success probability \eqn{p}.
When \code{n = 1} (the default), this is a Bernoulli distribution.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  state <- nv_rng_state(42L)
  # Bernoulli samples
  result <- nv_rbinom(c(2, 3), state)
  result[[2]]
})
\dontshow{\}) # examplesIf}
}
\seealso{
Other rng: 
\code{\link{nv_rdunif}()},
\code{\link{nv_rng_state}()},
\code{\link{nv_rnorm}()},
\code{\link{nv_runif}()}
}
\concept{rng}
