% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nvl_rng_bit_generator}
\alias{nvl_rng_bit_generator}
\title{Random Numbers}
\usage{
nvl_rng_bit_generator(
  initial_state,
  rng_algorithm = "THREE_FRY",
  dtype,
  shape_out
)
}
\arguments{
\item{initial_state}{state seed}

\item{rng_algorithm}{one of 'DEFAULT', 'THREE_FRY', 'PHILOX'}

\item{dtype}{datatype of output}

\item{shape_out}{output shape}
}
\description{
generate random bits of desired shape and dtype
}
