% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-rng.R
\name{nv_rnorm}
\alias{nv_rnorm}
\title{Sample from a Normal Distribution}
\usage{
nv_rnorm(shape, initial_state, dtype = "f32", mu = 0, sigma = 1)
}
\arguments{
\item{shape}{(\code{integer()})\cr
Shape.}

\item{initial_state}{(\code{\link{tensorish}})\cr
RNG state (\code{ui64[2]}).}

\item{dtype}{(\code{character(1)} | \code{\link[tengen:TensorDataType]{tengen::TensorDataType}})\cr
Data type.}

\item{mu}{(\code{\link{tensorish}})\cr
Mean.}

\item{sigma}{(\code{\link{tensorish}})\cr
Standard deviation. Must be positive, otherwise results are invalid.}
}
\value{
(\code{list()} of \code{\link{tensorish}})\cr
List of two elements: the updated RNG state and the sampled values.
}
\description{
Samples from a normal distribution with mean \eqn{\mu} and standard deviation \eqn{\sigma}
using the Box-Muller transform.
}
\section{Covariance}{

To implement a covariance structure use Cholesky decomposition.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  state <- nv_rng_state(42L)
  result <- nv_rnorm(c(2, 3), state)
  result[[2]]
})
\dontshow{\}) # examplesIf}
}
\seealso{
Other rng: 
\code{\link{nv_rbinom}()},
\code{\link{nv_rdunif}()},
\code{\link{nv_rng_state}()},
\code{\link{nv_runif}()}
}
\concept{rng}
