% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_round}
\alias{nv_round}
\title{Round}
\usage{
nv_round(operand, method = "nearest_even")
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{method}{(\code{character(1)})\cr
Rounding method.
Either \code{"nearest_even"} (default) or \code{"afz"} (away from zero).}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise rounding. You can also use the \code{round()} generic.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1.4, 2.5, 3.6))
  round(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_round]{nvl_round()}} for the underlying primitive.
}
