% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_rsqrt}
\alias{nv_rsqrt}
\title{Reciprocal Square Root}
\usage{
nv_rsqrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise reciprocal square root, i.e. \code{1 / sqrt(x)}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 4, 9))
  nv_rsqrt(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_rsqrt]{nvl_rsqrt()}} for the underlying primitive.
}
