% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-rng.R
\name{nv_runif}
\alias{nv_runif}
\title{Sample from a Uniform Distribution}
\usage{
nv_runif(shape, initial_state, dtype = "f32", lower = 0, upper = 1)
}
\arguments{
\item{shape}{(\code{integer()})\cr
Shape.}

\item{initial_state}{(\code{\link{tensorish}})\cr
RNG state (\code{ui64[2]}).}

\item{dtype}{(\code{character(1)} | \code{\link[tengen:TensorDataType]{tengen::TensorDataType}})\cr
Data type.}

\item{lower, upper}{(\code{numeric(1)})\cr
Lower and upper bound.}
}
\value{
(\code{list()} of \code{\link{tensorish}})\cr
List of two elements: the updated RNG state and the sampled values.
}
\description{
Samples from a uniform distribution in the open interval \verb{(lower, upper)}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  state <- nv_rng_state(42L)
  result <- nv_runif(c(2, 3), state)
  result[[2]]
})
\dontshow{\}) # examplesIf}
}
\seealso{
Other rng: 
\code{\link{nv_rbinom}()},
\code{\link{nv_rdunif}()},
\code{\link{nv_rng_state}()},
\code{\link{nv_rnorm}()}
}
\concept{rng}
