% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-rng.R
\name{nv_sample_int}
\alias{nv_sample_int}
\title{Random Discrete Sample}
\usage{
nv_sample_int(n, shape, initial_state, dtype = "i32")
}
\arguments{
\item{n}{(\code{integer(1)})\cr
Number of categories to sample from (samples integers 1 to n).}

\item{shape}{(\code{integer()})\cr
Shape.}

\item{initial_state}{(\code{\link{tensorish}})\cr
Tensor of type \code{ui64[2]} for the RNG state.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Output dtype (default "i32").}
}
\value{
(\code{list()} of \code{\link{tensorish}})\cr
List of two tensors: the new RNG state and the sampled integers (1 to n).
}
\description{
Sample from a discrete distribution, analogous to R's \code{sample()} function.
Samples integers from 1 to n with uniform probability and with replacement.
}
