% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{nv_save}
\alias{nv_save}
\title{Save tensors to a file}
\usage{
nv_save(tensors, path)
}
\arguments{
\item{tensors}{(named \code{list} of \code{\link{AnvilTensor}})\cr
Named list of tensors to save. Names must be unique.}

\item{path}{(\code{character(1)})\cr
File path to write to.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Saves a named list of tensors to a file in the
\href{https://huggingface.co/docs/safetensors/index}{safetensors} format.
}
\details{
This is a convenience wrapper around \code{\link[=nv_serialize]{nv_serialize()}} that opens and closes
a file connection.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded("cpu")) withAutoprint(\{ # examplesIf}
x <- nv_tensor(array(1:6, dim = c(2, 3)))
x
path <- tempfile(fileext = ".safetensors")
nv_save(list(x = x), path)
nv_read(path)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_read]{nv_read()}}, \code{\link[=nv_serialize]{nv_serialize()}}, \code{\link[=nv_unserialize]{nv_unserialize()}}
}
