% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_seed2state}
\alias{nv_seed2state}
\title{Make initial state}
\usage{
nv_seed2state(
  dtype = "ui64",
  shape_out = 2,
  random_seed = NULL,
  hash_algo = "sha512"
)
}
\arguments{
\item{dtype}{output dtype either "ui32" or "ui64"}

\item{shape_out}{output shape}

\item{random_seed}{explicitly provide the random seed of a R session. auto-detects if not provided.}

\item{hash_algo}{hash algorithm to hash the random state with. Default is 'sha512'.}
}
\description{
converts a random seed into a initial state tensor
}
