% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{nv_serialization}
\alias{nv_serialization}
\alias{nv_write}
\alias{nv_read}
\alias{nv_serialize}
\alias{nv_unserialize}
\title{Tensor serialization and I/O}
\usage{
nv_write(tensors, path)

nv_read(path, device = NULL)

nv_serialize(tensors)

nv_unserialize(con, device = NULL)
}
\arguments{
\item{tensors}{(named \code{list} of \code{\link{AnvilTensor}})\cr
Named list of tensors.}

\item{path}{(\code{character(1)})\cr
Path to the safetensors file.}

\item{device}{(\code{NULL} | \code{character(1)} | \code{\link[pjrt:pjrt_device]{PJRTDevice}})\cr
The platform name for the tensor (\code{"cpu"}, \code{"cuda"}, ...).
Default is to use the CPU.}

\item{con}{(connection)\cr
A connection object to read serialized tensors from.}
}
\value{
\itemize{
\item \code{nv_write()}: \code{NULL} (invisibly)
\item \code{nv_read()}: Named list of \code{\link{AnvilTensor}} objects
\item \code{nv_serialize()}: Raw vector containing serialized tensors
\item \code{nv_unserialize()}: Named list of \code{\link{AnvilTensor}} objects
}
}
\description{
Read and write tensors using the safetensors format.
to/from raw vectors in memory.
}
\details{
These functions wrap the safetensors format functionality provided by the
\href{https://github.com/mlverse/safetensors}{\code{safetensors}} package.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded("cpu")) withAutoprint(\{ # examplesIf}
x <- nv_tensor(array(1:6, dim = c(2, 3)))
raw_data <- nv_serialize(list(x = x))
reloaded <- nv_unserialize(raw_data)
expect_equal(x, reloaded$x)
\dontshow{\}) # examplesIf}
}
