% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{nv_serialize}
\alias{nv_serialize}
\title{Serialize tensors to raw bytes}
\usage{
nv_serialize(tensors, con = NULL)
}
\arguments{
\item{tensors}{(named \code{list} of \code{\link{AnvilTensor}})\cr
Named list of tensors to serialize. Names must be unique.}

\item{con}{(\code{NULL} | connection)\cr
An optional connection to write to.
If \code{NULL} (default), a raw vector is returned.}
}
\value{
A \code{\link{raw}} vector if \code{con} is \code{NULL}, otherwise \code{NULL} (invisibly).
}
\description{
Serializes a named list of tensors into the
\href{https://huggingface.co/docs/safetensors/index}{safetensors} format.
}
\details{
The ambiguity of the tensors is stored in the metadata and preserved in write-read roundtrips.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded("cpu")) withAutoprint(\{ # examplesIf}
x <- nv_tensor(array(1:6, dim = c(2, 3)))
x
raw_data <- nv_serialize(list(x = x))
raw_data
nv_unserialize(raw_data)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_unserialize]{nv_unserialize()}}, \code{\link[=nv_save]{nv_save()}}, \code{\link[=nv_read]{nv_read()}}
}
