% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_set_elt}
\alias{nv_set_elt}
\title{Set Element at Indices}
\usage{
nv_set_elt(operand, ..., value)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{...}{(\code{\link{tensorish}})\cr
Scalar tensors specifying the index for each dimension (1-based).}

\item{value}{(\code{\link{tensorish}})\cr
Scalar tensor with the new value.}
}
\value{
\code{\link{tensorish}}
}
\description{
Update a single element in a tensor at specific indices.
All indices must be 0-dimensional (scalar) tensors.
}
