% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_shift_left}
\alias{nv_shift_left}
\title{Shift Left}
\usage{
nv_shift_left(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise bitwise left shift.
}
