% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_shift_right_arithmetic}
\alias{nv_shift_right_arithmetic}
\title{Arithmetic Shift Right}
\usage{
nv_shift_right_arithmetic(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.
Operands are \link[=nv_promote_to_common]{promoted to a common data type}.
Scalars are \link[=nv_broadcast_scalars]{broadcast} to the shape of the other operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and the promoted common data type of the inputs.
}
\description{
Element-wise arithmetic right bit shift.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(8L, -16L, 32L))
  y <- nv_tensor(c(1L, 2L, 3L))
  nv_shift_right_arithmetic(x, y)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_shift_right_arithmetic]{nvl_shift_right_arithmetic()}} for the underlying primitive.
}
