% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_sine}
\alias{nv_sine}
\title{Sine}
\usage{
nv_sine(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise sine. You can also use \code{sin()}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(0, pi / 2, pi))
  sin(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_sine]{nvl_sine()}} for the underlying primitive.
}
