% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_tan}
\alias{nv_tan}
\title{Tangent}
\usage{
nv_tan(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise tangent. You can also use \code{tan()}.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(0, 0.5, 1))
tan(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_tan]{prim_tan()}} for the underlying primitive.
}
