% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_tanh}
\alias{nv_tanh}
\title{Hyperbolic Tangent}
\usage{
nv_tanh(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise hyperbolic tangent. You can also use \code{tanh()}.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(-1, 0, 1))
tanh(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_tanh]{prim_tanh()}} for the underlying primitive.
}
