% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{nv_tensor}
\alias{nv_tensor}
\alias{nv_scalar}
\alias{nv_empty}
\title{Tensor}
\usage{
nv_tensor(data, dtype = NULL, platform = NULL, shape = NULL)

nv_scalar(data, dtype = NULL, platform = "cpu")

nv_empty(dtype, shape, platform = "cpu")
}
\arguments{
\item{x}{(any)\cr
Object.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
\code{nv_tensor}
}
\description{
Create a tensor.
}
\details{
Internally calls \code{\link[pjrt:pjrt_buffer]{pjrt_buffer}}.
}
