% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_tril}
\alias{nv_tril}
\title{Lower Triangular Matrix}
\usage{
nv_tril(operand, diagonal = 0L)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}

\item{diagonal}{(\code{integer(1)})\cr
Diagonal offset. \code{0} (default) is the main diagonal, positive values
include diagonals above, negative values exclude diagonals below.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as \code{operand}.
}
\description{
Returns the lower triangular part of a 2-D array, setting elements above
the specified diagonal to zero.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_fill(1, c(3, 3))
nv_tril(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_triu]{nv_triu()}}
}
