% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_triu}
\alias{nv_triu}
\title{Upper Triangular Matrix}
\usage{
nv_triu(operand, diagonal = 0L)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}

\item{diagonal}{(\code{integer(1)})\cr
Diagonal offset. \code{0} (default) is the main diagonal, positive values
exclude diagonals above, negative values include diagonals below.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as \code{operand}.
}
\description{
Returns the upper triangular part of a 2-D array, setting elements below
the specified diagonal to zero.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_fill(1, c(3, 3))
  nv_triu(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_tril]{nv_tril()}}
}
