% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_unary_ops}
\alias{nv_unary_ops}
\alias{nv_neg}
\title{Unary Operations}
\usage{
nv_neg(operand)
}
\arguments{
\item{operand}{(\code{\link{nv_tensor}})}
}
\value{
\code{\link{nv_tensor}}
}
\description{
Unary operations on tensors.
}
