% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{nv_unserialize}
\alias{nv_unserialize}
\title{Deserialize tensors from raw bytes}
\usage{
nv_unserialize(con, device = NULL)
}
\arguments{
\item{con}{(connection | \code{\link{raw}})\cr
A connection or raw vector to read from.}

\item{device}{(\code{NULL} | \code{character(1)} | \code{\link[pjrt:pjrt_device]{PJRTDevice}})\cr
The device on which to place the loaded tensors (\code{"cpu"}, \code{"cuda"}, ...).
Default is to use the CPU.}
}
\value{
Named \code{list} of \code{\link{AnvilTensor}} objects.
}
\description{
Deserializes tensors from the
\href{https://huggingface.co/docs/safetensors/index}{safetensors} format.
}
\details{
The data type, shape, and \link[=ambiguous]{ambiguity} of each tensor are
restored from the serialized data.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded("cpu")) withAutoprint(\{ # examplesIf}
x <- nv_tensor(array(1:6, dim = c(2, 3)))
x
raw_data <- nv_serialize(list(x = x))
raw_data
nv_unserialize(raw_data)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_serialize]{nv_serialize()}}, \code{\link[=nv_save]{nv_save()}}, \code{\link[=nv_read]{nv_read()}}
}
