% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_unsqueeze}
\alias{nv_unsqueeze}
\title{Unsqueeze}
\usage{
nv_unsqueeze(operand, dim)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dim}{(\code{integer(1)})\cr
Position at which to insert the new dimension.}
}
\value{
\code{\link{tensorish}}\cr
Has the same data type as \code{operand} with an extra dimension of size 1.
}
\description{
Inserts a dimension of size 1 at the specified position.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  nv_unsqueeze(x, dim = 1L)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_squeeze]{nv_squeeze()}}, \code{\link[=nv_reshape]{nv_reshape()}}
}
