% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_add}
\alias{nvl_add}
\title{Primitive Addition}
\usage{
nvl_add(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Tensorish values of any data type.
Must have the same shape.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the inputs.
It is ambiguous if both inputs are ambiguous.
}
\description{
Adds two tensors element-wise.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_add]{stablehlo::hlo_add()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  y <- nv_tensor(c(4, 5, 6))
  nvl_add(x, y)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_add]{nv_add()}}, \code{+}
}
