% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_and}
\alias{nvl_and}
\title{Primitive And}
\usage{
nvl_and(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Tensorish values of data type boolean, integer, or unsigned integer.
Must have the same shape.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the inputs.
It is ambiguous if both inputs are ambiguous.
}
\description{
Element-wise logical AND.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_and]{stablehlo::hlo_and()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(TRUE, FALSE, TRUE))
  y <- nv_tensor(c(TRUE, TRUE, FALSE))
  nvl_and(x, y)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_and]{nv_and()}}, \code{&}
}
