% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{nvl_array}
\alias{nvl_array}
\title{Array}
\usage{
nvl_array(x, ...)
}
\arguments{
\item{x}{(any)\cr
Object.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
\code{nvl_array}
}
\description{
Create an array.
}
\details{
Internally calls \code{\link[pjrt:pjrt_buffer]{pjrt_buffer}}.
}
