% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_atan2}
\alias{nvl_atan2}
\title{Primitive Atan2}
\usage{
nvl_atan2(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Tensorish values of data type floating-point.
Must have the same shape.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the inputs.
It is ambiguous if both inputs are ambiguous.
}
\description{
Element-wise atan2 operation.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_atan2]{stablehlo::hlo_atan2()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  y <- nv_tensor(c(1, 0, -1))
  x <- nv_tensor(c(0, 1, 0))
  nvl_atan2(y, x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_atan2]{nv_atan2()}}
}
