% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_broadcast_in_dim}
\alias{nvl_broadcast_in_dim}
\title{Primitive Broadcast}
\usage{
nvl_broadcast_in_dim(operand, shape, broadcast_dimensions)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of any data type.}

\item{shape}{(\code{integer()})\cr
Target shape. Each mapped dimension must either match the corresponding
operand dimension or the operand dimension must be 1.}

\item{broadcast_dimensions}{(\code{integer()})\cr
Maps each dimension of \code{operand} to a dimension of the output.
Must have length equal to the number of dimensions of \code{operand}.}
}
\value{
\code{\link{tensorish}}\cr
Has the same data type as the input and the given \code{shape}.
It is ambiguous if the input is ambiguous.
}
\description{
Broadcasts a tensor to a new shape by replicating the data along new or size-1 dimensions.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_broadcast_in_dim]{stablehlo::hlo_broadcast_in_dim()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  nvl_broadcast_in_dim(x, shape = c(2, 3), broadcast_dimensions = 2L)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_broadcast_to]{nv_broadcast_to()}}
}
