% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_cbrt}
\alias{nvl_cbrt}
\title{Primitive Cube Root}
\usage{
nvl_cbrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise cube root.
}
