% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_clamp}
\alias{nvl_clamp}
\title{Primitive Clamp}
\usage{
nvl_clamp(min_val, operand, max_val)
}
\arguments{
\item{min_val}{(\code{\link{tensorish}})\cr
Minimum value (scalar or same shape as operand).}

\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{max_val}{(\code{\link{tensorish}})\cr
Maximum value (scalar or same shape as operand).}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise clamp: max(min_val, min(operand, max_val)).
}
