% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_convert}
\alias{nvl_convert}
\title{Primitive Convert}
\usage{
nvl_convert(operand, dtype, ambiguous = FALSE)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Arrayish value of any data type.}

\item{dtype}{(\code{character(1)} | \code{\link[tengen:DataType]{tengen::DataType}})\cr
Target data type.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the type is ambiguous.
Ambiguous types usually arise from R literals (e.g., \code{1L}, \code{1.0}) and follow special promotion rules.
See the \code{vignette("type-promotion")} for more details.}
}
\value{
\code{\link{arrayish}}\cr
Has the given \code{dtype} and the same shape as \code{operand}.
Ambiguity is controlled by the \code{ambiguous} parameter.
}
\description{
Converts the elements of an array to a different data type.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{reverse}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_convert]{stablehlo::hlo_convert()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_array(c(1L, 2L, 3L))
  nvl_convert(x, dtype = "f32")
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_convert]{nv_convert()}}
}
