% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_div}
\alias{nvl_div}
\title{Primitive Division}
\usage{
nvl_div(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Tensorish values of data type integer, unsigned integer, or floating-point.
Must have the same shape.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the inputs.
It is ambiguous if both inputs are ambiguous.
}
\description{
Divides two tensors element-wise.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_divide]{stablehlo::hlo_divide()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(10, 20, 30))
  y <- nv_tensor(c(2, 5, 10))
  nvl_div(x, y)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_div]{nv_div()}}, \code{/}
}
