% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_eq}
\alias{nvl_eq}
\title{Primitive Equal}
\usage{
nvl_eq(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}} (boolean)
}
\description{
Element-wise equality comparison.
}
