% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_floor}
\alias{nvl_floor}
\title{Primitive Floor}
\usage{
nvl_floor(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of data type floating-point.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise floor.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_floor]{stablehlo::hlo_floor()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1.2, 2.7, -1.5))
  nvl_floor(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_floor]{nv_floor()}}, \code{\link[=floor]{floor()}}
}
