% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_iota}
\alias{nvl_iota}
\title{Primitive Iota}
\usage{
nvl_iota(dim, dtype, shape, start = 1L, ambiguous = FALSE)
}
\arguments{
\item{dim}{(\code{integer(1)})\cr
Dimension along which values increase (1-indexed).}

\item{dtype}{(\code{character(1)} | \code{\link[tengen:TensorDataType]{tengen::TensorDataType}})\cr
Data type.}

\item{shape}{(\code{integer()})\cr
Shape of the output tensor.}

\item{start}{(\code{integer(1)})\cr
Starting value.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the type is ambiguous.
Ambiguous types usually arise from R literals (e.g., \code{1L}, \code{1.0}) and follow special promotion rules.
See the \code{vignette("type-promotion")} for more details.}
}
\value{
\code{\link{tensorish}}\cr
Has the given \code{dtype} and \code{shape}.
}
\description{
Creates a tensor with values increasing along the specified dimension.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_iota]{stablehlo::hlo_iota()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval(nvl_iota(dim = 1L, dtype = "i32", shape = 5L))
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_iota]{nv_iota()}}
}
