% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_is_finite}
\alias{nvl_is_finite}
\title{Primitive Is Finite}
\usage{
nvl_is_finite(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}} of boolean type
}
\description{
Element-wise check if values are finite (not Inf, -Inf, or NaN).
}
