% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_le}
\alias{nvl_le}
\title{Primitive Less Equal}
\usage{
nvl_le(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}} (boolean)
}
\description{
Element-wise less than or equal comparison.
}
