% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_log}
\alias{nvl_log}
\title{Primitive Logarithm}
\usage{
nvl_log(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of data type floating-point.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise natural logarithm.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_log]{stablehlo::hlo_log()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2.718, 7.389))
  nvl_log(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_log]{nv_log()}}, \code{\link[=log]{log()}}
}
