% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_log1p}
\alias{nvl_log1p}
\title{Primitive Log Plus One}
\usage{
nvl_log1p(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of data type floating-point.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise log(1 + x), more accurate for small x.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_log_plus_one]{stablehlo::hlo_log_plus_one()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(0, 0.001, 1))
  nvl_log1p(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_log1p]{nv_log1p()}}
}
