% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_lt}
\alias{nvl_lt}
\title{Primitive Less Than}
\usage{
nvl_lt(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}} (boolean)
}
\description{
Element-wise less than comparison.
}
