% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_max}
\alias{nvl_max}
\title{Primitive Maximum}
\usage{
nvl_max(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Tensorish values of any data type.
Must have the same shape.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the inputs.
It is ambiguous if both inputs are ambiguous.
}
\description{
Element-wise maximum of two tensors.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_maximum]{stablehlo::hlo_maximum()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 5, 3))
  y <- nv_tensor(c(4, 2, 6))
  nvl_max(x, y)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_max]{nv_max()}}
}
