% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_min}
\alias{nvl_min}
\title{Primitive Minimum}
\usage{
nvl_min(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise minimum of two tensors.
}
