% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_mul}
\alias{nvl_mul}
\title{Primitive Multiplication}
\usage{
nvl_mul(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Tensorish values of any data type.
Must have the same shape.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the inputs.
It is ambiguous if both inputs are ambiguous.
}
\description{
Multiplies two tensors element-wise.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_multiply]{stablehlo::hlo_multiply()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  y <- nv_tensor(c(4, 5, 6))
  nvl_mul(x, y)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_mul]{nv_mul()}}, \code{*}
}
