% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_neg}
\alias{nvl_neg}
\title{Primitive Negation}
\usage{
nvl_neg(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Negates a tensor element-wise.
}
