% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_negate}
\alias{nvl_negate}
\title{Primitive Negation}
\usage{
nvl_negate(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of data type integer or floating-point.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Negates a tensor element-wise.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_negate]{stablehlo::hlo_negate()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, -2, 3))
  nvl_negate(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_negate]{nv_negate()}}, unary \code{-}
}
