% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_or}
\alias{nvl_or}
\title{Primitive Or}
\usage{
nvl_or(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{arrayish}})\cr
Arrayish values of data type boolean, integer, or unsigned integer.
Must have the same shape.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the inputs.
It is ambiguous if both inputs are ambiguous.
}
\description{
Element-wise logical OR.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{reverse}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_or]{stablehlo::hlo_or()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_array(c(TRUE, FALSE, TRUE))
  y <- nv_array(c(TRUE, TRUE, FALSE))
  nvl_or(x, y)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_or]{nv_or()}}, \code{|}
}
