% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_pad}
\alias{nvl_pad}
\title{Primitive Pad}
\usage{
nvl_pad(
  operand,
  padding_value,
  edge_padding_low,
  edge_padding_high,
  interior_padding
)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of any data type.}

\item{padding_value}{(\code{\link{tensorish}})\cr
Scalar value to use for padding. Must have the same dtype as \code{operand}.}

\item{edge_padding_low}{(\code{integer()})\cr
Amount of padding to add at the start of each dimension.}

\item{edge_padding_high}{(\code{integer()})\cr
Amount of padding to add at the end of each dimension.}

\item{interior_padding}{(\code{integer()})\cr
Amount of padding to add between elements in each dimension.}
}
\value{
\code{\link{tensorish}}\cr
Has the same data type as \code{operand}.
For the output shape see the underlying stablehlo documentation (\code{\link[stablehlo:hlo_pad]{stablehlo::hlo_pad()}}).
It is ambiguous if the input is ambiguous.
}
\description{
Pads a tensor with a given padding value.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_pad]{stablehlo::hlo_pad()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  nvl_pad(x, nv_scalar(0),
    edge_padding_low = 2L, edge_padding_high = 1L, interior_padding = 0L
  )
})
\dontshow{\}) # examplesIf}
}
