% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_popcnt}
\alias{nvl_popcnt}
\title{Primitive Population Count}
\usage{
nvl_popcnt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise population count (number of set bits).
}
