% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_print}
\alias{nvl_print}
\title{Primitive Print}
\usage{
nvl_print(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of any data type.}
}
\value{
\code{\link{tensorish}}\cr
Returns \code{operand} as-is.
}
\description{
Prints a tensor value to the console during execution and returns the
input unchanged. This is useful for debugging JIT-compiled code.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_custom_call]{stablehlo::hlo_custom_call()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  nvl_print(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_print]{nv_print()}}
}
