% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_reduce_min}
\alias{nvl_reduce_min}
\title{Primitive Min Reduction}
\usage{
nvl_reduce_min(operand, dims, drop = TRUE)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of any data type.}

\item{dims}{(\code{integer()})\cr
Dimensions to reduce over.}

\item{drop}{(\code{logical(1)})\cr
Whether to drop the reduced dimensions from the output shape.
If \code{TRUE}, the reduced dimensions are removed.
If \code{FALSE}, the reduced dimensions are set to 1.}
}
\value{
\code{\link{tensorish}}\cr
Has the same data type as the input.
When \code{drop = TRUE}, the shape is that of \code{operand} with \code{dims} removed.
When \code{drop = FALSE}, the shape is that of \code{operand} with \code{dims} set to 1.
It is ambiguous if the input is ambiguous.
}
\description{
Finds the minimum of tensor elements along the specified dimensions.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_reduce]{stablehlo::hlo_reduce()}} with \code{\link[stablehlo:hlo_minimum]{stablehlo::hlo_minimum()}} as the reducer.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(matrix(1:6, nrow = 2))
  nvl_reduce_min(x, dims = 1L)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_reduce_min]{nv_reduce_min()}}
}
