% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_reduce_sum}
\alias{nvl_reduce_sum}
\title{Primitive Sum Reduction}
\usage{
nvl_reduce_sum(operand, dims, drop = TRUE)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dims}{(\code{integer()})\cr
Dimensions to reduce.}

\item{drop}{(\code{logical(1)})\cr
Whether to drop reduced dimensions.}
}
\value{
\code{\link{tensorish}}
}
\description{
Sums tensor elements along dimensions.
}
\section{Shapes}{

\itemize{
\item \code{operand}: any shape
\item output (\code{drop = TRUE}): shape of \code{operand} with \code{dims} removed
\item output (\code{drop = FALSE}): shape of \code{operand} with \code{dims} set to 1
}
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_reduce]{stablehlo::hlo_reduce()}} with \code{\link[stablehlo:hlo_add]{stablehlo::hlo_add()}} as the reducer.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(matrix(1:6, nrow = 2))
  nvl_reduce_sum(x, dims = 1L)
})
}
}
