% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_remainder}
\alias{nvl_remainder}
\title{Primitive Remainder}
\usage{
nvl_remainder(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Tensorish values of data type integer, unsigned integer, or floating-point.
Must have the same shape.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the inputs.
It is ambiguous if both inputs are ambiguous.
}
\description{
Element-wise remainder of division.
For a more user-friendly interface, see \code{\link[=nv_remainder]{nv_remainder()}}.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_remainder]{stablehlo::hlo_remainder()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(7, 10, 15))
  y <- nv_tensor(c(3, 4, 6))
  nvl_remainder(x, y)
})
\dontshow{\}) # examplesIf}
}
