% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_shift_left}
\alias{nvl_shift_left}
\title{Primitive Shift Left}
\usage{
nvl_shift_left(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}} of integer type)\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise left bit shift.
}
\section{Shapes}{

\code{lhs} and \code{rhs} must have the same shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_shift_left]{stablehlo::hlo_shift_left()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1L, 2L, 4L))
  y <- nv_tensor(c(1L, 2L, 1L))
  nvl_shift_left(x, y)
})
\dontshow{\}) # examplesIf}
}
