% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_shift_right_arithmetic}
\alias{nvl_shift_right_arithmetic}
\title{Primitive Arithmetic Shift Right}
\usage{
nvl_shift_right_arithmetic(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}} of integer type)\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise arithmetic right bit shift.
}
\section{Shapes}{

\code{lhs} and \code{rhs} must have the same shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_shift_right_arithmetic]{stablehlo::hlo_shift_right_arithmetic()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(8L, -16L, 32L))
  y <- nv_tensor(c(1L, 2L, 3L))
  nvl_shift_right_arithmetic(x, y)
})
}
}
