% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_shift_right_logical}
\alias{nvl_shift_right_logical}
\title{Primitive Logical Shift Right}
\usage{
nvl_shift_right_logical(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Tensorish values of data type boolean, integer, or unsigned integer.
Must have the same shape.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the inputs.
It is ambiguous if both inputs are ambiguous.
}
\description{
Element-wise logical right bit shift.
For a more user-friendly interface, see \code{\link[=nv_shift_right_logical]{nv_shift_right_logical()}}.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_shift_right_logical]{stablehlo::hlo_shift_right_logical()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(8L, 16L, 32L))
  y <- nv_tensor(c(1L, 2L, 3L))
  nvl_shift_right_logical(x, y)
})
\dontshow{\}) # examplesIf}
}
