% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_sine}
\alias{nvl_sine}
\title{Primitive Sine}
\usage{
nvl_sine(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of data type floating-point.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise sine.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_sine]{stablehlo::hlo_sine()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(0, pi / 2, pi))
  nvl_sine(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_sine]{nv_sine()}}, \code{\link[=sin]{sin()}}
}
