% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_tan}
\alias{nvl_tan}
\title{Primitive Tangent}
\usage{
nvl_tan(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of data type floating-point.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise tangent.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_tan]{stablehlo::hlo_tan()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(0, 0.5, 1))
  nvl_tan(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_tan]{nv_tan()}}, \code{\link[=tan]{tan()}}
}
