% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_tanh}
\alias{nvl_tanh}
\title{Primitive Hyperbolic Tangent}
\usage{
nvl_tanh(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of data type floating-point.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise hyperbolic tangent.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_tanh]{stablehlo::hlo_tanh()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(-1, 0, 1))
  nvl_tanh(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_tanh]{nv_tanh()}}, \code{\link[=tanh]{tanh()}}
}
