% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{nvl_tensor}
\alias{nvl_tensor}
\title{Tensor}
\usage{
nvl_tensor(x, ...)
}
\arguments{
\item{x}{(any)\cr
Object.}

\item{...}{(any)\cr
Additional arguments.}
}
\value{
\code{nvl_tensor}
}
\description{
Create an tensor.
}
\details{
Internally calls \code{\link[pjrt:pjrt_buffer]{pjrt_buffer}}.
}
