% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{platform.ConcreteTensor}
\alias{platform.ConcreteTensor}
\title{Platform for ConcreteTensor}
\usage{
\method{platform}{ConcreteTensor}(x, ...)
}
\arguments{
\item{x}{A ConcreteTensor.}

\item{...}{Additional arguments (unused).}
}
\value{
The platform string.
}
\description{
Get the platform of a ConcreteTensor.
}
