% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{platform}
\alias{platform}
\alias{platform.AnvilArray}
\title{Get the platform}
\usage{
\method{platform}{AnvilArray}(x, ...)
}
\arguments{
\item{x}{An array object.}

\item{...}{Additional arguments (unused).}
}
\value{
\code{character(1)}
}
\description{
Returns the platform name (e.g. \code{"cpu"}, \code{"cuda"}).
}
