% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\docType{data}
\name{prim}
\alias{prim}
\title{Primitive Registry}
\format{
An environment.
}
\usage{
prim
}
\description{
Environment containing all registered primitives. Access an individual
primitive by name via \code{prim$add}, \code{prim$mul}, etc. Iterate over all
primitives with \code{as.list(prim)} or \code{eapply(prim, f)}.
}
\keyword{datasets}
