% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{prim_bessel_i1e}
\alias{prim_bessel_i1e}
\title{Primitive Exponentially Scaled Modified Bessel I_1}
\usage{
prim_bessel_i1e(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Arrayish value of data type floating-point.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise exponentially scaled modified Bessel function of the
first kind, order 1: \code{bessel_i1e(x) = bessel_i1(x) * exp(-abs(x))}.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_bessel_i1e]{stablehlo::hlo_bessel_i1e()}}.
}

\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(-1, 0, 1, 5))
prim_bessel_i1e(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_bessel_i1e]{nv_bessel_i1e()}}
}
