% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{prim_ceil}
\alias{prim_ceil}
\title{Primitive Ceiling}
\usage{
prim_ceil(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Arrayish value of data type floating-point.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise ceiling.
}
\section{Access}{

Access this primitive via \code{prim$ceil}.
}

\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{reverse}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_ceil]{stablehlo::hlo_ceil()}}.
}

\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(1.2, 2.7, -1.5))
nvl_ceil(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_ceil]{nv_ceil()}}, \code{\link[=ceiling]{ceiling()}}
}
