% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{prim_eq}
\alias{prim_eq}
\title{Primitive Equal}
\usage{
prim_eq(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{arrayish}})\cr
Arrayish values of any data type.
Must have the same shape.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape as the inputs and boolean data type.
It is ambiguous if both inputs are ambiguous.
}
\description{
Element-wise equality comparison.
}
\section{Access}{

Access this primitive via \code{prim$equal}.
}

\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{reverse}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_compare]{stablehlo::hlo_compare()}} with \code{comparison_direction = "EQ"}.
}

\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(1, 2, 3))
y <- nv_array(c(1, 3, 2))
prim$eq(x, y)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_eq]{nv_eq()}}, \code{==}
}
