% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{prim_erf_inv}
\alias{prim_erf_inv}
\title{Primitive Inverse Error Function}
\usage{
prim_erf_inv(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Arrayish value of data type floating-point.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise inverse error function (the inverse of \code{erf} on \verb{(-1, 1)}).
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{reverse}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_erf_inv]{stablehlo::hlo_erf_inv()}}.
}

\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(-0.5, 0, 0.5))
prim_erf_inv(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_erf_inv]{nv_erf_inv()}}
}
