% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{prim_logistic}
\alias{prim_logistic}
\title{Primitive Logistic (Sigmoid)}
\usage{
prim_logistic(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Arrayish value of data type floating-point.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise logistic sigmoid: 1 / (1 + exp(-x)).
}
\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_logistic]{stablehlo::hlo_logistic()}}.
}

\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(-2, 0, 2))
nvl_logistic(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_logistic]{nv_logistic()}}
}
