% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{prim_lt}
\alias{prim_lt}
\title{Primitive Less Than}
\usage{
prim_lt(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{arrayish}})\cr
Arrayish values of any data type.
Must have the same shape.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape as the inputs and boolean data type.
It is ambiguous if both inputs are ambiguous.
}
\description{
Element-wise less than comparison.
}
\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_compare]{stablehlo::hlo_compare()}} with \code{comparison_direction = "LT"}.
}

\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(1, 2, 3))
y <- nv_array(c(3, 2, 1))
nvl_lt(x, y)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_lt]{nv_lt()}}, \code{<}
}
