% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{prim_min}
\alias{prim_min}
\title{Primitive Minimum}
\usage{
prim_min(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{arrayish}})\cr
Arrayish values of any data type.
Must have the same shape.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the inputs.
It is ambiguous if both inputs are ambiguous.
}
\description{
Element-wise minimum of two arrays.
}
\section{Access}{

Access this primitive via \code{prim$minimum}.
}

\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{reverse}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_minimum]{stablehlo::hlo_minimum()}}.
}

\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(1, 5, 3))
y <- nv_array(c(4, 2, 6))
prim$min(x, y)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_min]{nv_min()}}
}
