% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{prim_popcnt}
\alias{prim_popcnt}
\title{Primitive Population Count}
\usage{
prim_popcnt(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Arrayish value of data type integer or unsigned integer.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise population count (number of set bits).
}
\section{Access}{

Access this primitive via \code{prim$popcnt}.
}

\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{reverse}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_popcnt]{stablehlo::hlo_popcnt()}}.
}

\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(7L, 3L, 15L))
prim$popcnt(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_popcnt]{nv_popcnt()}}
}
