% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{prim_sine}
\alias{prim_sine}
\title{Primitive Sine}
\usage{
prim_sine(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Arrayish value of data type floating-point.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise sine.
}
\section{Access}{

Access this primitive via \code{prim$sine}.
}

\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{reverse}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_sine]{stablehlo::hlo_sine()}}.
}

\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(0, pi / 2, pi))
prim$sine(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_sine]{nv_sine()}}, \code{\link[=sin]{sin()}}
}
