% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{prim_square}
\alias{prim_square}
\title{Primitive Square}
\usage{
prim_square(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Arrayish value of data type floating-point.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise square \code{x^2}.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{reverse}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_square]{stablehlo::hlo_square()}}.
}

\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(-2, 0, 3))
prim_square(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_square]{nv_square()}}
}
