% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{prim_sub}
\alias{prim_sub}
\title{Primitive Subtraction}
\usage{
prim_sub(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{arrayish}})\cr
Arrayish values of data type integer, unsigned integer, or floating-point.
Must have the same shape.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the inputs.
It is ambiguous if both inputs are ambiguous.
}
\description{
Subtracts two arrays element-wise.
}
\section{Access}{

Access this primitive via \code{prim$sub}.
}

\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{reverse}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_subtract]{stablehlo::hlo_subtract()}}.
}

\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(1, 2, 3))
y <- nv_array(c(4, 5, 6))
nvl_sub(x, y)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_sub]{nv_sub()}}, \code{-}
}
