% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{Primitive}
\alias{Primitive}
\title{Primitive}
\usage{
Primitive(name)
}
\arguments{
\item{name}{(\code{character()})\cr
The name of the primitive.}
}
\value{
(\code{Primitive})
}
\description{
Primitive interpretation rule.
}
