% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-quickr.R
\name{QuickrDevice}
\alias{QuickrDevice}
\title{Quickr device}
\usage{
QuickrDevice(type = "cpu")
}
\arguments{
\item{type}{(\code{character(1)})\cr
Device type. Must be \code{"cpu"}.}
}
\value{
A \code{QuickrDevice} object.
}
\description{
Device descriptor for the quickr backend. The only supported \code{type} is
\code{"cpu"}.
}
\seealso{
\code{\link[=nv_device]{nv_device()}}, \code{\link[=AnvilBackendQuickr]{AnvilBackendQuickr()}}.
}
