% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{register_primitive}
\alias{register_primitive}
\title{Register a Primitive}
\usage{
register_primitive(name, primitive, overwrite = FALSE)
}
\arguments{
\item{name}{(\code{character()})\cr
The name of the primitive.}

\item{primitive}{(\code{AnvilPrimitive})\cr
The primitive to register.}

\item{overwrite}{(\code{logical(1)})\cr
Whether to overwrite the primitive if it is already registered.}
}
\description{
Register a primitive.
}
