% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{reindex_tree}
\alias{reindex_tree}
\title{Reindex Tree}
\usage{
reindex_tree(x, counter)
}
\arguments{
\item{x}{Tree node to reindex.}

\item{counter}{Counter object for generating new indices.}
}
\value{
Reindexed tree node.
}
\description{
Recursively reindex leaf nodes starting from a counter.
}
