% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{repository_stats}
\alias{repository_stats}
\alias{print.repository_stats}
\title{Obtain a comparative summary of the Bioconductor binary package
repository}
\usage{
repository_stats(
  version = BiocManager::version(),
  binary_base_url = BINARY_BASE_URL
)

\method{print}{repository_stats}(x)
}
\arguments{
\item{version}{`character(1)` or `package_version` Bioconductor
version, e.g., "3.12".}

\item{binary_base_url}{character(1) The base URL for installation of binary
packages. This argument is not usually set by the user.}
}
\value{
A printout of metrics that include the Bioconductor repository
version and location URL, the number of binary and source packages
currently available, the number missing and out-of-date binaries,
among other metrics.
}
\description{
\code{repository_stats} provides a summary of binary versus source
packages compatible with the Bioconductor containers and/or Terra
environments.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Show the summary of repositories
}}

\examples{

repository_stats()

}
