% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-generics.R
\name{sd}
\alias{sd}
\title{Standard Deviation}
\usage{
sd(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Further arguments passed to methods.}
}
\description{
S3 generic dispatching to \code{\link[=nv_sd]{nv_sd()}} for \code{\link{arrayish}} inputs and
\code{\link[stats:sd]{stats::sd()}} otherwise.
}
