% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{shape}
\alias{shape}
\title{Get the shape of a tensor}
\usage{
shape(x, ...)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
A tensor-like object.}

\item{...}{Additional arguments passed to methods (unused).}
}
\value{
\code{integer()}
}
\description{
Returns the shape of a tensor as an \code{integer()} vector.
}
\details{
This is implemented via the generic \code{\link[tengen:shape]{tengen::shape()}}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_tensor(1:4, dtype = "f32")
shape(x)
\dontshow{\}) # examplesIf}
}
