% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{ShapedTensor}
\alias{ShapedTensor}
\title{Shaped Tensor Class}
\usage{
ShapedTensor(dtype, shape, ambiguous = FALSE)
}
\arguments{
\item{dtype}{(\code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
The data type of the tensor.}

\item{shape}{(\code{\link[stablehlo:Shape]{stablehlo::Shape}} | \code{integer()})\cr
The shape of the tensor. Can be provided as an integer vector.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the type is ambiguous. Ambiguous usually arise from R literals
(e.g., \code{1L}, \code{1.0}, \code{TRUE}) and follow special promotion rules.
Only \code{f32}, \code{i32}, and \code{i1} (bool) can be ambiguous.}
}
\description{
Abstract representation of a tensor with a known dtype and shape, but no concrete data.
Used during tracing to represent tensor metadata without actual values.
}
\seealso{
\link{ConcreteTensor}, \link{LiteralTensor}, \code{\link[=st]{st()}}
}
