% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{st}
\alias{st}
\title{Convert to Shaped Tensor}
\usage{
st(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to convert.}
}
\value{
\code{\link{ShapedTensor}}
}
\description{
Convert an object to its abstract tensor representation (\code{\link{ShapedTensor}}).
}
