% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stablehlo.R
\name{stablehlo}
\alias{stablehlo}
\title{Lower a function to StableHLO}
\usage{
stablehlo(graph, constants_as_inputs = TRUE, env = NULL, donate = character())
}
\arguments{
\item{graph}{(\code{Graph})\cr
The graph to lower.}

\item{constants_as_inputs}{(\code{logical(1)})\cr
Whether to add constants as inputs.}

\item{env}{(\code{HloEnv} | \code{NULL})\cr
The environment for storing graph value to func variable mappings.}

\item{donate}{(\code{character()})\cr
Names of the arguments whose buffers should be donated.
Donated buffers can be aliased with outputs of the same type.}
}
\value{
(\code{list}) with elements:
\itemize{
\item \code{func}: The StableHLO \code{Func} object
\item \code{constants}: The constants of the graph
}
}
\description{
Immediately lower a flattened function to a StableHLO Func object.
}
