% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-generics.R
\name{[.AnvilBox}
\alias{[.AnvilBox}
\title{Get Element at Indices}
\usage{
\method{[}{AnvilBox}(x, ...)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
Tensor to extract from.}

\item{...}{(\code{\link{tensorish}})\cr
Scalar tensors specifying the index for each dimension (1-based).}
}
\value{
Scalar \code{\link{tensorish}}
}
\description{
Extract a single element from a tensor at specific indices.
All indices must be 0-dimensional (scalar) tensors.
}
