% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-generics.R
\name{sub-.set.AnvilBox}
\alias{sub-.set.AnvilBox}
\alias{[.AnvilBox}
\alias{[.AnvilTensor}
\alias{[<-.AnvilBox}
\alias{[<-.AnvilTensor}
\title{Subset and Subset-Assign for Tensors}
\usage{
\method{[}{AnvilBox}(x, ...)

\method{[}{AnvilTensor}(x, ...)

\method{[}{AnvilBox}(x, ...) <- value

\method{[}{AnvilTensor}(x, ...) <- value
}
\arguments{
\item{x}{A tensor object.}

\item{...}{Subset specifications (see \code{\link[=nv_subset]{nv_subset()}}).}

\item{value}{For \verb{[<-}: the replacement values (see \code{\link[=nv_subset_assign]{nv_subset_assign()}}).}
}
\value{
A tensor.
}
\description{
These methods enable \code{[} and \verb{[<-} syntax for \code{\link{AnvilTensor}} and \code{\link{AnvilBox}} objects,
delegating to \code{\link[=nv_subset]{nv_subset()}} and \code{\link[=nv_subset_assign]{nv_subset_assign()}} respectively.
}
\seealso{
\code{\link[=nv_subset]{nv_subset()}}, \code{\link[=nv_subset_assign]{nv_subset_assign()}}
}
