% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{subgraphs}
\alias{subgraphs}
\title{Get Subgraphs from Higher-Order Primitive}
\usage{
subgraphs(call)
}
\arguments{
\item{call}{(\code{PrimitiveCall})\cr
The primitive call.}
}
\value{
(\code{list(AnvilGraph)})\cr
List of subgraphs found in the parameters.
}
\description{
Extracts subgraphs from the parameters of a higher-order primitive call.
}
