% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-generics.R
\name{[<-.AnvilBox}
\alias{[<-.AnvilBox}
\title{Set Element at Indices}
\usage{
\method{[}{AnvilBox}(x, ...) <- value
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
Tensor to update.}

\item{...}{(\code{\link{tensorish}})\cr
Scalar tensors specifying the index for each dimension (1-based).}

\item{value}{(\code{\link{tensorish}})\cr
Scalar tensor with the new value.}
}
\value{
\code{\link{tensorish}}
}
\description{
Update a single element in a tensor at specific indices.
All indices must be 0-dimensional (scalar) tensors.
}
