% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tensorish}
\alias{tensorish}
\alias{is_tensorish}
\title{Tensor-like Objects}
\usage{
is_tensorish(x, literal = TRUE)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}

\item{literal}{(\code{logical(1)})\cr
Whether to accept R literals as tensorish.}
}
\value{
\code{logical(1)}
}
\description{
A \code{tensorish} value is any object that can be passed as an input to
anvil primitive functions such as \code{\link{nvl_add}}.

The following types are tensorish:
\itemize{
\item \code{\link{AnvilTensor}}: a concrete tensor holding data on a device.
\item \code{\link{GraphBox}}: a boxed abstract tensor representing a value in a graph.
\item Literals: \code{numeric(1)}, \code{integer(1)}, \code{logical(1)}: promoted to scalar tensors.
}

Use \code{\link[=is_tensorish]{is_tensorish()}} to check whether a value is tensorish.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
# AnvilTensors are tensorish
is_tensorish(nv_tensor(1:4))

# Scalar R literals are tensorish by default
is_tensorish(1.5)

# Non-scalar vectors are not tensorish
is_tensorish(1:4)

is_tensorish(DebugBox(nv_aten("f32", c(2L, 3L))))

# Disable literal promotion
is_tensorish(1.5, literal = FALSE)
\dontshow{\}) # examplesIf}
}
\seealso{
\link{AnvilTensor}, \link{GraphBox}
}
