% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{unwrap_if_tensor}
\alias{unwrap_if_tensor}
\title{Get the underlying PJRT buffer from an AnvilTensor or pass through other values}
\usage{
unwrap_if_tensor(x)
}
\arguments{
\item{x}{An AnvilTensor or any other value}
}
\value{
The underlying PJRT buffer if x is an AnvilTensor, otherwise x unchanged
}
\description{
Get the underlying PJRT buffer from an AnvilTensor or pass through other values
}
\keyword{internal}
