% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-generics.R
\name{var}
\alias{var}
\title{Variance}
\usage{
var(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Further arguments passed to methods.}
}
\description{
S3 generic dispatching to \code{\link[=nv_var]{nv_var()}} for \code{\link{arrayish}} inputs and
\code{\link[stats:cor]{stats::var()}} otherwise.
}
