% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-converters.R
\name{vt2at}
\alias{vt2at}
\title{Convert ValueType to AbstractTensor}
\usage{
vt2at(x)
}
\arguments{
\item{x}{(\code{ValueType})}
}
\value{
(\code{AbstractTensor})
}
\description{
Convert a \code{ValueType} to an \code{AbstractTensor}.
}
