% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{nv_aten}
\alias{nv_aten}
\alias{AbstractTensor}
\title{Abstract Tensor Class}
\usage{
nv_aten(dtype, shape, ambiguous = FALSE)

AbstractTensor(dtype, shape, ambiguous = FALSE)
}
\arguments{
\item{dtype}{(\code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
The data type of the tensor.}

\item{shape}{(\code{\link[stablehlo:Shape]{stablehlo::Shape}} | \code{integer()})\cr
The shape of the tensor. Can be provided as an integer vector.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the type is ambiguous.
Ambiguous types usually arise from R literals (e.g., \code{1L}, \code{1.0}) and follow special promotion rules.
See the vignette "Type Promotion" for more details.}
}
\description{
Abstract representation of a tensor with a (possibly ambiguous) dtype and shape, but no concrete data.
Used during tracing to represent tensor metadata without actual values.
}
\details{
Two tensors are considered equal (\code{==}) if they have the same dtype and shape, ignoring ambiguity.
}
\seealso{
\link{ConcreteTensor}, \link{LiteralTensor}, \code{\link[=to_abstract]{to_abstract()}}
}
