% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnVIL-package.R
\docType{package}
\name{AnVIL-package}
\alias{AnVIL}
\alias{AnVIL-package}
\title{AnVIL: Bioconductor on the AnVIL compute environment}
\description{
The AnVIL is a cloud computing resource developed in part by the National Human Genome Research Institute. The AnVIL package provides end-user and developer functionality. For the end-user, AnVIL provides fast binary package installation, utitlities for working with Terra / AnVIL table and data resources, and convenient functions for file movement to and from Google cloud storage. For developers, AnVIL provides programatic access to the Terra, Leonardo, Rawls, and Dockstore RESTful programming interface, including helper functions to transform JSON responses to formats more amenable to manipulation in R.
}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramos@sph.cuny.edu} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})

Authors:
\itemize{
  \item Martin Morgan (\href{https://orcid.org/0000-0002-5874-8148}{ORCID})
  \item Kayla Interdonato
  \item Yubo Cheng
  \item Nitesh Turaga
}

Other contributors:
\itemize{
  \item BJ Stubbs [contributor]
  \item Vincent Carey [contributor]
  \item Sehyun Oh [contributor]
  \item Sweta Gopaulakrishnan [contributor]
  \item Valerie Obenchain [contributor]
}

}
\keyword{internal}
