% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anvil-package.R
\docType{package}
\name{anvil-package}
\alias{anvil}
\alias{anvil-package}
\title{anvil: Access and Analyze Alloy Attributes}
\description{
Simplifies and standardizes data access through Athena, S3, and Arrow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alloy-commons/anvil}
  \item Report bugs at \url{https://github.com/alloy-commons/anvil/issues}
}

}
\author{
\strong{Maintainer}: Nicholas DiQuattro \email{nicholas@alloy.us}

Other contributors:
\itemize{
  \item Susanna Supalla \email{susanna@alloy.us} [contributor]
  \item Richard Barney \email{richard.barney@alloy.us} [contributor]
  \item Clara Fried \email{clara@alloy.us} [contributor]
}

}
\keyword{internal}
