% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{anvil_set_auth_json}
\alias{anvil_set_auth_json}
\title{Store and retrieve authentication credentials using a secure keyring}
\usage{
anvil_set_auth_json(service, path)
}
\arguments{
\item{service}{\code{character(1)} The name of the service (e.g., \code{"terra"},
\code{"dockstore"}) for which the credentials are being set.}

\item{path}{\code{character(1)} The path to the \code{auth.json} file.}
}
\value{
\code{anvil_set_auth_json()} returns \code{NULL} invisibly.
}
\description{
\code{anvil_set_auth_json()} stores the content of an \code{auth.json}
file in the system keyring. This is the recommended way to store
credentials safely.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
jsonlite::write_json(
   list(token = "example_token"),
   "terratcgadata-test-key.json",
)
anvil_set_auth_json(
    "terra",
    "terratcgadata-test-key.json"
)
AnVIL:::.authenticate_get_access("terra")
unlink("terratcgadata-test-key.json")
\dontshow{\}) # examplesIf}
}
