% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{AnvilBox}
\alias{AnvilBox}
\title{AnvilBox}
\description{
Virtual S3 base class for \code{\link{GraphBox}} and \code{\link{DebugBox}}.
}
\seealso{
\link{DebugBox}, \link{GraphBox}, \code{\link[=debug_box]{debug_box()}}
}
