% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{AnvilGraph}
\alias{AnvilGraph}
\alias{Graph}
\title{Graph of Primitive Calls}
\usage{
AnvilGraph(
  calls = list(),
  in_tree = NULL,
  out_tree = NULL,
  inputs = list(),
  outputs = list(),
  constants = list()
)

Graph(
  calls = list(),
  in_tree = NULL,
  out_tree = NULL,
  inputs = list(),
  outputs = list(),
  constants = list()
)
}
\arguments{
\item{calls}{(\code{list(PrimitiveCall)})\cr
The primitive calls that make up the graph.
This can also be another call into a graph when the primitive is a \code{p_call}.}

\item{in_tree}{(\code{NULL | Node})\cr
The tree of inputs.}

\item{out_tree}{(\code{NULL | Node})\cr
The tree of outputs.}

\item{inputs}{(\code{list(GraphValue)})\cr
The inputs to the graph.}

\item{outputs}{(\code{list(GraphValue)})\cr
The outputs of the graph.}

\item{constants}{(\code{list(GraphValue)})\cr
The constants of the graph.}
}
\value{
(\code{AnvilGraph})
}
\description{
Computational graph consisting exclusively of primitive calls.
This is a mutable class.
}
