% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{AnvilPrimitive}
\alias{AnvilPrimitive}
\title{AnvilPrimitive}
\usage{
AnvilPrimitive(name, higher_order = FALSE, subgraphs = character())
}
\arguments{
\item{name}{(\code{character()})\cr
The name of the primitive.}

\item{higher_order}{(\code{logical(1)})\cr
Whether the primitive is higher-order (contains subgraphs). Default is \code{FALSE}.}

\item{subgraphs}{(\code{character()})\cr
Names of parameters that are subgraphs. Only used if \code{higher_order = TRUE}.}
}
\value{
(\code{AnvilPrimitive})
}
\description{
Primitive interpretation rule.
Note that \code{[[} and \verb{[[<-} access the interpretation rules.
To access other fields, use \code{$} and \verb{$<-}.
}
