% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{AnvilPrimitive}
\alias{AnvilPrimitive}
\title{AnvilPrimitive}
\usage{
AnvilPrimitive(name)
}
\arguments{
\item{name}{(\code{character()})\cr
The name of the primitive.}
}
\value{
(\code{AnvilPrimitive})
}
\description{
Primitive interpretation rule.
}
