% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{AnvilPrimitive}
\alias{AnvilPrimitive}
\alias{Primitive}
\title{AnvilPrimitive}
\usage{
AnvilPrimitive(name)

Primitive(name)
}
\arguments{
\item{name}{(\code{character()})\cr
The name of the primitive.}
}
\value{
(\code{AnvilPrimitive})
}
\description{
Primitive interpretation rule.
Note that \code{[[} and \verb{[[<-} access the interpretation rules.
To access other fields, use \code{$} and \verb{$<-}.
}
