% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{AnvilTensor}
\alias{AnvilTensor}
\title{AnvilTensor}
\description{
Virtual base class for tensor objects in anvil.
This class is used to mark objects that can be used as tensors in anvil.
Cannot be instantiated directly - use \code{\link[=nv_tensor]{nv_tensor()}}, \code{\link[=nv_scalar]{nv_scalar()}}, or \code{\link[=nv_empty]{nv_empty()}} instead.
}
