% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{arrayish}
\alias{arrayish}
\alias{is_arrayish}
\title{Array-like Objects}
\usage{
is_arrayish(x, literal = TRUE)
}
\arguments{
\item{x}{(\code{any})\cr
Object to check.}

\item{literal}{(\code{logical(1)})\cr
Whether to accept R literals as arrayish.}
}
\value{
\code{logical(1)}
}
\description{
A \code{arrayish} value is any object that can be passed as an input to
anvil primitive functions such as \code{\link{nvl_add}} or is an output of such a function.

During runtime, these are \code{\link{AnvilArray}} objects.

The following types are arrayish (during compile-time):
\itemize{
\item \code{\link{AnvilArray}}: a concrete array holding data on a device.
\item \code{\link{GraphBox}}: a boxed abstract array representing a value in a graph.
\item Literals: \code{numeric(1)}, \code{integer(1)}, \code{logical(1)}: promoted to scalar arrays.
}

Use \code{\link[=is_arrayish]{is_arrayish()}} to check whether a value is arrayish.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
# AnvilArrays are arrayish
is_arrayish(nv_array(1:4))

# Scalar R literals are arrayish by default
is_arrayish(1.5)

# Non-scalar vectors are not arrayish
is_arrayish(1:4)

is_arrayish(DebugBox(nv_aten("f32", c(2L, 3L))))

# Disable literal promotion
is_arrayish(1.5, literal = FALSE)
\dontshow{\}) # examplesIf}
}
\seealso{
\link{AnvilArray}, \link{GraphBox}
}
