% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{as_array}
\alias{as_array}
\title{Convert a tensor to an R array}
\usage{
as_array(x, ...)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
A tensor-like object.}

\item{...}{Additional arguments passed to methods (unused).}
}
\value{
An R \code{\link{array}} or \code{vector} of length 1.
}
\description{
Transfers tensor data to R and returns it as an R \code{\link{array}}.
Only in the case of scalars is the result a vector of length 1, as R \code{arrays} cannot have 0 dimensions.
}
\details{
This is implemented via the generic \code{\link[tengen:as_array]{tengen::as_array()}}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_tensor(1:4, dtype = "f32")
as_array(x)
y <- nv_scalar(1L)
# R arrays can't have 0 dimensions:
as_array(y)
\dontshow{\}) # examplesIf}
}
