% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-functions.R
\name{as_quickr_function}
\alias{as_quickr_function}
\title{Convert to a quickr-compiled function}
\usage{
as_quickr_function(graph)
}
\value{
(\code{function})
}
\description{
Convenience wrapper around \code{\link[=graph_to_quickr_function]{graph_to_quickr_function()}}.
}
