% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-converters.R
\name{at2vt}
\alias{at2vt}
\title{Convert AbstractTensor to ValueType}
\usage{
at2vt(x)
}
\arguments{
\item{x}{(\code{AbstractTensor})}
}
\value{
(\code{ValueType})
}
\description{
Convert an \code{AbstractTensor} to a \code{ValueType}.
}
