% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate against the google cloud app `anvil-leo-dev`}
\usage{
authenticate(service, cache = getOption("httr_oauth_cache"))
}
\arguments{
\item{service}{character(1) name of AnVIL service (e.g.,
"leonardo", "terra") requiring authentication.}

\item{cache}{(optional) logical(1) or character(1) describing how
the OAuth 2.0 token will be managed. See
`?httr:oauth2.0_token`.}
}
\value{
An 'R6' token instance containing the authorization,
    invisibly.
}
\description{
Authenticate against the google cloud app `anvil-leo-dev`
}
\details{
AnVIL applications require OAuth 2.0 credentials
    identifying the application. These must be added to the package
    source before installation.

    For Terra and Leonardo, visit
    `https://console.cloud.google.com/apis/credentials?authuser=1&project=anvil-leo-dev`
    and download (click on the downward-facing arrow to the
    right) the "Bioconductor-AnVIL" credentials to a file
    `inst/service/{leonardo,terra}/auth.json`.
}
