% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avworkspace.R
\name{avworkspace}
\alias{avworkspace}
\alias{avworkspace_namespace}
\alias{avworkspace_name}
\alias{avworkspace_clone}
\title{Workspace management}
\usage{
avworkspace_namespace(namespace = NULL)

avworkspace_name(name = NULL)

avworkspace(workspace = NULL)

avworkspace_clone(
  namespace = avworkspace_namespace(),
  name = avworkspace_name(),
  to_namespace = namespace,
  to_name
)
}
\arguments{
\item{namespace}{character(1) AnVIL workspace namespace as returned
by, e.g., `avworkspace_namespace()`}

\item{name}{character(1) AnVIL workspace name as returned by, eg.,
`avworkspace_name()`.}

\item{workspace}{when present, a `character(1)` providing the
concatenated namespace and name, e.g.,
`"bioconductor-rpci-anvil/Bioconductor-Package-AnVIL"`}

\item{to_namespace}{character(1) workspace (billing account) in
which to make the clone.}

\item{to_name}{character(1) name of the cloned workspace.}
}
\value{
`avworkspace_namespace()`, and `avworkspace_name()` return
    `character(1)` identifiers. `avworkspace()` returns the
    character(1) concatenated namespace and name. The value
    returned by `avworkspace_name()` will be percent-encoded (e.g.,
    spaces `" "` replaced by `"%20"`).

`avworkspace_clone()` returns the namespace and name, in
    the format `namespace/name`, of the cloned workspace.
}
\description{
`avworkspace_namespace()` and `avworkspace_name()` are
    utiliity functions to retrieve workspace namespace and name
    from environment variables or interfaces usually available in
    AnVIL notebooks or RStudio sessions.  `avworkspace()` provides
    a convenient way to specify workspace namespace and name in a
    single command.

`avworkspace_clone()` clones (copies) an existing
    workspace, possibly into a new namespace (billing account).
}
\details{
`avworkspace_namespace()` is the billing account. If the
    `namespace=` argument is not provided, try `gcloud_project()`,
    and if that fails try `Sys.getenv("WORKSPACE_NAMESPACE")`.

    `avworkspace_name()` is the name of the workspace as it appears
    in \url{https://app.terra.bio/#workspaces}. If not provided,
    `avworkspace_name()` tries to use
    `Sys.getenv("WORKSPACE_NAME")`.

    Namespace and name values are cached across sessions, so
    explicitly providing `avworkspace_name*()` is required at most
    once per session. Revert to system settings with arguments
    `NA`.
}
\examples{
avworkspace_namespace()
avworkspace_name()
avworkspace()

}
