% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{build_tree}
\alias{build_tree}
\title{Build Tree}
\usage{
build_tree(x, counter = NULL)
}
\arguments{
\item{x}{Object to build tree from.}

\item{counter}{Internal counter for leaf indices.}
}
\value{
\code{Node}
}
\description{
Build a tree structure from a nested object for tracking structure during flattening/unflattening.
}
