% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promotion.R
\name{common_dtype}
\alias{common_dtype}
\title{Type Promotion Rules}
\usage{
common_dtype(
  lhs_dtype,
  rhs_dtype,
  lhs_ambiguous = FALSE,
  rhs_ambiguous = FALSE
)
}
\arguments{
\item{lhs_ambiguous}{(\code{logical(1)})\cr
Whether the left-hand side type is ambiguous.}

\item{rhs_ambiguous}{(\code{logical(1)})\cr
Whether the right-hand side type is ambiguous.}

\item{lhs}{(\code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
The left-hand side type.}

\item{rhs}{(\code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
The right-hand side type.}
}
\value{
(\verb{list(dtype = [}stablehlo::TensorDataType\verb{], ambiguous = }logical(1)`)\cr
}
\description{
Computes the common dtype for a set of abstract types, respecting whether a type is ambiguous or not.
A type is ambiguous if it comes from a literal (like 1, 1.0, TRUE) or was cast
to an ambiguous type.
Promoting to an ambiguous type can happen in scenarios like \code{x + 1.2}, where \code{x} is a bool or an int.
}
