% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{ConcreteTensor}
\alias{ConcreteTensor}
\title{Concrete Tensor Class}
\usage{
ConcreteTensor(data)
}
\arguments{
\item{data}{(\code{\link{AnvilTensor}})\cr
The actual tensor data.}
}
\description{
A \code{\link{ShapedTensor}} that also holds a reference to the actual tensor data.
Used to represent constants captured during tracing.
Because it comes from a concrete tensor, it's type is not ambiguous.
}
\seealso{
\link{ShapedTensor}, \link{LiteralTensor}
}
