% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{data_types}
\alias{data_types}
\title{Tensor Data Types}
\description{
Data types for tensors:
\itemize{
\item \code{dt_i1}: boolean.
\item \verb{dt_i\{8, 16, 32, 64\}}: signed integer.
\item \verb{dt_ui\{8, 16, 32, 64\}}: unsigned integer.
\item \verb{dt_f\{32, 64\}}: float.
}
}
