% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug_box}
\alias{debug_box}
\title{Create a Debug Box}
\usage{
debug_box(dtype, shape, ambiguous = FALSE)
}
\arguments{
\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{shape}{(\code{integer()})\cr
Shape.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the type is ambiguous.
Ambiguous types usually arise from R literals (e.g., \code{1L}, \code{1.0}) and follow special promotion rules.
See the vignette "Type Promotion" for more details.}
}
\value{
(\code{\link{DebugBox}})
}
\description{
Create a debug box.
}
