% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug_box}
\alias{debug_box}
\title{Create a Debug Box}
\usage{
debug_box(dtype, shape, ambiguous = FALSE)
}
\arguments{
\item{aval}{(\code{AbstractTensor})\cr
The abstract tensor representing the value.}
}
\value{
(\code{DebugBox})
}
\description{
Create a debug box.
}
