% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{DebugBox}
\alias{DebugBox}
\title{Debug Box Class}
\usage{
DebugBox(aval = AbstractTensor())
}
\arguments{
\item{aval}{(\code{AbstractTensor})\cr
The abstract tensor representing the value.}
}
\description{
Box representing a value in debug mode.
}
