% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{delocalize}
\alias{delocalize}
\title{Delocalize a Google storage bucket from the container.}
\usage{
delocalize(local_path, google_bucket, ...)
}
\arguments{
\item{local_path}{character vector, representing a local path}

\item{google_bucket}{character, a valid path to a google storage
bucket.}
}
\value{
Exit status of function, `gsutil_rsync()`
}
\description{
Delocalize a Google storage bucket from the container.
}
\details{
delocalize the volume which is mounted with the help of
    gsutil. This syncs the library a final time and then removes
    the volume from the directory. Once a path is delocalized, it
    has to be localized again to regain access.
}
