% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{device}
\alias{device}
\title{Get the device of a tensor}
\usage{
device(x, ...)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
A tensor-like object.}

\item{...}{Additional arguments passed to methods (unused).}
}
\value{
\code{\link[pjrt:pjrt_device]{PJRTDevice}}
}
\description{
Returns the device on which a tensor is allocated.
}
\details{
This is implemented via the generic \code{\link[tengen:device]{tengen::device()}}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_tensor(1:4, dtype = "f32")
device(x)
\dontshow{\}) # examplesIf}
}
