% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{.current_descriptor}
\alias{.current_descriptor}
\title{Get the current graph}
\usage{
.current_descriptor(silent = FALSE)
}
\value{
A \code{\link{Graph}} object.
}
\description{
Get the current graph being built (via \code{\link{local_descriptor}}).
}
