% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drs.R
\name{drs_stat}
\alias{drs_stat}
\alias{drs_access_url}
\alias{drs_cp}
\title{DRS (Data Repository Service) URL management}
\usage{
drs_stat(source = character(), region = "US")

drs_access_url(source = character(), region = "US")

drs_cp(source, destination, ...)
}
\arguments{
\item{source}{character() DRS URLs (beginning with 'drs://') to
resources managed by the 'martha' DRS resolution server.}

\item{region}{character(1) Google cloud 'region' in which the DRS
resource is located. Most data is located in \code{"US"} (the
default); in principle \code{"auto"} allows for discovery of
the region, but sometimes fails. Regions are enumerated at
\url{https://cloud.google.com/storage/docs/locations#available-locations}.}

\item{destination}{character(1) directory path in which to retrieve
files.}

\item{...}{additional arguments, passed to \code{gsutil_cp()} for file
copying.}
}
\value{
\code{drs_stat()} returns a tbl with the following columns:
\itemize{
\item fileName: character() (resolver sometimes returns null).
\item size: integer() (resolver sometimes returns null).
\item contentType: character() (resolver sometimes returns null).
\item gsUri: character() (resolver sometimes returns null).
\item timeCreated: character() (the time created formatted using ISO
8601; resolver sometimes returns null).
\item timeUpdated: character() (the time updated formatted using ISO
8601; resolver sometimes returns null).
\item bucket: character() (resolver sometimes returns null).
\item name: character() (resolver sometimes returns null).
\item googleServiceAccount: list() (null unless the DOS url belongs to
a Bond supported host).
\item hashes: list() (contains the hashes type and their checksum
value; if unknown. it returns null)
}

`drs_access_url()` returns a vector of https URLs
    corresponding to the vector of DRS URIs provided as inputs to
    the function.

\code{drs_cp()} returns a tibble like \code{drs_stat()}, but with
additional columns
\itemize{
\item simple: logical() value indicating whether resolution used a
simple signed URL (\code{TRUE}) or auxilliary service account.
\item destination: character() full path to retrieved object(s)
}
}
\description{
\code{drs_stat()} resolves zero or more DRS URLs to their
google bucket location.

`drs_access_url()` returns a vector of 'signed' URLs
    that allow access to restricted resources via standard https
    protocols.

\code{drs_cp()} copies 0 or more DRS URIs to a google
bucket or local folder
}
\details{
\code{drs_stat()} sends requests in parallel to the DRS server,
using 8 forked processes (by default) to speed up queries. Use
\code{options(mc.cores = 16L)}, for instance, to set the number of
processes to use.\preformatted{`drs_stat()` uses the AnVIL 'pet' account associated with a
runtime. The pet account is discovered by default when
evaluated on an AnVIL runtime (e.g., in RStudio or a Jupyter
notebook in the AnVIL), or can be found in the return value of
`avruntimes()`.
}
}
\examples{
drs_eg_anvil <- c(
    "drs://dg.ANV0/975bd45f-f022-4fad-b9a2-3a00c3b8792c",
    "drs://dg.ANV0/00008531-03d7-418c-b3d3-b7b22b5381a0"
)

if (gcloud_exists() && startsWith(gcloud_account(), "pet-")) {
    ## from within AnVIL
    drs_stat(drs_eg_anvil)
}


}
