% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud-deprecated.R
\name{gcloud-deprecated}
\alias{gcloud-deprecated}
\alias{gcloud}
\alias{gcloud_exists}
\alias{gcloud_account}
\alias{gcloud_project}
\alias{gcloud_help}
\alias{gcloud_cmd}
\title{gcloud and gsutil command line utility interfaces (DEPRECATED)}
\usage{
gcloud_exists()

gcloud_account(account = NULL)

gcloud_project(project = NULL)

gcloud_help(...)

gcloud_cmd(cmd, ...)
}
\arguments{
\item{account}{character(1) Google account (e.g., \code{user@gmail.com})
to use for authentication.}

\item{project}{character(1) billing project name.}

\item{...}{Additional arguments appended to gcloud commands.}

\item{cmd}{\code{character(1)} representing a command used to evaluate
\verb{gcloud cmd ...}.}
}
\value{
\code{gcloud_exists()} returns \code{TRUE} when the \code{gcloud}
application can be found, FALSE otherwise.

\code{gcloud_account()} returns a \code{character(1)} vector
containing the active gcloud account, typically a gmail email
address.

\code{gcloud_project()} returns a \code{character(1)} vector
containing the active gcloud project.

\code{gcloud_help()} returns an unquoted \code{character()} vector
representing the text of the help manual page returned by
\verb{gcloud help ...}.

\code{gcloud_cmd()} returns a \code{character()} vector representing
the text of the output of \verb{gcloud cmd ...}
}
\description{
These functions invoke the \code{gcloud} and \code{gsutil} command line
utilities. See \link{gsutil} for details on how \code{gcloud} is
located.

\code{gcloud_exists()} tests whether the \code{gcloud()} command
can be found on this system. See 'Details' section of \code{gsutil}
for where the application is searched.

\code{gcloud_account()}: report the current gcloud account
via \verb{gcloud config get-value account}.

\code{gcloud_project()}: report the current gcloud project
via \verb{gcloud config get-value project}.

\code{gcloud_help()}: queries \code{gcloud} for help for a
command or sub-comand via \verb{gcloud help ...}.

\code{gcloud_cmd()} allows arbitrary \code{gcloud} command
execution via \verb{gcloud ...}. Use pre-defined functions in
preference to this.
}
\examples{
gcloud_exists()

if (gcloud_exists())
    gcloud_account()

if (gcloud_exists())
    gcloud_help()

}
