% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpreter-pullback.R
\name{gradient}
\alias{gradient}
\title{Gradient of a function}
\usage{
gradient(f, wrt = NULL)
}
\arguments{
\item{f}{(\code{function})}

\item{wrt}{(\code{character()})
Names of arguments to differentiate with respect to. If \code{NULL},
compute gradients w.r.t. all differentiable arguments.}
}
\value{
(\code{function})
}
\description{
Compute the gradient of a function using reverse mode automatic differentiation.
Output must be a 0-dimensional tensor.
}
