% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullback.R
\name{gradient}
\alias{gradient}
\title{Gradient}
\usage{
gradient(f, wrt = character())
}
\description{
Returns a function such when called, it will build up the gradient graph, insert it into its parent graph and return the outputs of the gradient graph.
}
\section{Signature of the output Function}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{f: (Box | StaticInput).. -> (Box | StaticInput)..
}\if{html}{\out{</div>}}
}

