% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{gradient}
\alias{gradient}
\alias{value_and_gradient}
\title{Gradient}
\usage{
gradient(f, wrt = NULL)

value_and_gradient(f, wrt = NULL)
}
\arguments{
\item{f}{(\code{function})\cr
Function to compute the gradient of.}

\item{wrt}{(\code{character})\cr
Names of the arguments to compute the gradient with respect to.}
}
\description{
Transform a function to its gradient.
}
\section{Functions}{
\itemize{
\item \code{value_and_gradient()}: Returns both the value and the gradient

}}
