% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_desc_add}
\alias{graph_desc_add}
\title{Add a Primitive Call to a Graph Descriptor}
\usage{
graph_desc_add(prim, args, params = list(), infer_fn, desc = NULL)
}
\arguments{
\item{prim}{(\code{\link{AnvilPrimitive}})\cr
The primitive to add.}

\item{args}{(\code{list} of \code{\link{GraphNode}})\cr
The arguments to the primitive.}

\item{params}{(\code{list})\cr
The parameters to the primitive.}

\item{infer_fn}{(\code{function})\cr
The inference function to use.
Must output a list of \code{\link{AbstractArray}}s.}

\item{desc}{(\code{\link{GraphDescriptor}} | \code{NULL})\cr
The graph descriptor to add the primitive call to.
Uses the \link[=.current_descriptor]{current descriptor} if \code{NULL}.}
}
\value{
(\code{list} of \code{\link{GraphBox}})
}
\description{
Add a primitive call to a graph descriptor.
}
