% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_desc_add}
\alias{graph_desc_add}
\title{Add a primitive call to a graph descriptor}
\usage{
graph_desc_add(
  prim,
  args,
  params = list(),
  infer_fn,
  desc = NULL,
  debug_mode = NULL
)
}
\arguments{
\item{prim}{(\code{Primitive})\cr
The primitive to add.}

\item{args}{(\code{list})\cr
The arguments to the primitive.}

\item{params}{(\code{list})\cr
The parameters to the primitive.}

\item{infer_fn}{(\code{function})\cr
The inference function to use.}

\item{desc}{(\code{GraphDescriptor})\cr
The graph descriptor to add the primitive call to.}
}
\value{
A list of \code{GraphBox} objects.
}
\description{
Add a primitive call to a graph descriptor.
}
