% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_desc_add}
\alias{graph_desc_add}
\title{Add a Primitive Call to a Graph Descriptor}
\usage{
graph_desc_add(
  prim,
  args,
  params = list(),
  infer_fn,
  desc = NULL,
  debug_mode = NULL
)
}
\arguments{
\item{prim}{(\code{\link{AnvilPrimitive}})\cr
The primitive to add.}

\item{args}{(\code{list} of \code{\link{GraphNode}})\cr
The arguments to the primitive.}

\item{params}{(\code{list})\cr
The parameters to the primitive.}

\item{infer_fn}{(\code{function})\cr
The inference function to use.
Must output a list of \code{\link{AbstractTensor}}s.}

\item{desc}{(\code{\link{GraphDescriptor}} | \code{NULL})\cr
The graph descriptor to add the primitive call to.
Uses the \link[=.current_descriptor]{current descriptor} if \code{NULL}.}

\item{debug_mode}{(\code{logical(1)})\cr
Whether to just perform abstract evaluation for debugging.}
}
\value{
(\code{list} of \code{Box})\cr
Either \code{GraphBox} objects or \code{DebugBox} objects, depending on \code{debug_mode}.
}
\description{
Add a primitive call to a graph descriptor.
}
