% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-to-quickr.R
\name{graph_to_quickr_function}
\alias{graph_to_quickr_function}
\title{Convert a Graph to a quickr-compiled function}
\usage{
graph_to_quickr_function(graph)
}
\arguments{
\item{graph}{(\code{\link{Graph}})\cr
Graph to convert.}
}
\value{
(\code{function})
}
\description{
Convenience wrapper around \code{\link[=graph_to_r_function]{graph_to_r_function()}} that inserts a
\code{declare(type(...))} header and compiles the resulting function with
\code{quickr::quick()}.
}
