% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{GraphBox}
\alias{GraphBox}
\title{Graph Box}
\usage{
GraphBox(gnode = `mut<GraphValue>`(), desc = `mut<GraphDescriptor>`())
}
\arguments{
\item{gnode}{(\code{\link{GraphNode}})\cr
The node.}

\item{desc}{(\code{\link{GraphDescriptor}})\cr
The descriptor of the graph.}
}
\description{
Box that represents a node in a \code{\link{GraphDescriptor}}.
}
