% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{GraphDescriptor}
\alias{GraphDescriptor}
\title{Graph Descriptor}
\usage{
GraphDescriptor(
  calls = list(),
  tensor_to_gval = hashtab(),
  gval_to_box = hashtab(),
  constants = list(),
  in_tree = NULL,
  out_tree = NULL,
  inputs = list(),
  outputs = list()
)
}
\arguments{
\item{calls}{(\code{list(PrimitiveCall)})\cr
The primitive calls that make up the graph.}

\item{tensor_to_gval}{(\code{hashtab})\cr
Mapping: \code{AnvilTensor} -> \code{GraphValue}}

\item{gval_to_box}{(\code{hashtab})\cr
Mapping: \code{GraphValue} -> \code{GraphBox}}

\item{constants}{(\code{list(GraphValue)})\cr
The constants of the graph.}

\item{in_tree}{(\code{NULL | Node})\cr
The tree of inputs.}

\item{out_tree}{(\code{NULL | Node})\cr
The tree of outputs.}

\item{inputs}{(\code{list(GraphValue)})\cr
The inputs to the graph.}

\item{outputs}{(\code{list(GraphValue)})\cr
The outputs of the graph.}
}
\description{
Descriptor of a \code{\link{Graph}}.
}
