% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{GraphDescriptor}
\alias{GraphDescriptor}
\title{Graph Descriptor}
\usage{
GraphDescriptor(
  calls = list(),
  tensor_to_gval = hashtab(),
  gval_to_box = hashtab(),
  constants = list(),
  in_tree = NULL,
  out_tree = NULL,
  inputs = list(),
  outputs = list()
)
}
\description{
Descriptor of a \code{\link{Graph}}.
}
\details{
The trickiest thing in our setup are how we ensure that the same values receive the same identifier
(GraphValue) across nested graphs.
There are two cases:
\enumerate{
\item When a
}
}
\section{Fields}{

\itemize{
\item \code{calls} :: (\code{list(PrimitiveCall)})\cr
The primitive calls that make up the graph.
\item \code{tensor_to_gval} :: (\code{hashtab})\cr
Mapping: \code{AnvilTensor} -> \code{GraphValue}
\item \code{gval_to_box} :: (\code{hashtab})\cr
}
}

