% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{GraphLiteral}
\alias{GraphLiteral}
\title{Graph Literal}
\usage{
GraphLiteral(aval = NULL, dtype = stablehlo::BooleanType())
}
\arguments{
\item{aval}{(\code{any})\cr
The value of the literal.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}
}
\description{
Literal in a \code{\link{Graph}}.
}
