% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{GraphLiteral}
\alias{GraphLiteral}
\title{Graph Literal}
\usage{
GraphLiteral(aval = LiteralTensor())
}
\arguments{
\item{aval}{(\code{any})\cr
The value of the literal.}
}
\description{
Literal in a \code{\link{Graph}}.
}
