% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{GraphLiteral}
\alias{GraphLiteral}
\title{Graph Literal}
\usage{
GraphLiteral(aval)
}
\arguments{
\item{aval}{(\code{\link{LiteralTensor}})\cr
The value of the literal.}
}
\value{
(\code{GraphLiteral})
}
\description{
Literal in a \code{\link{Graph}}. This is a mutable class.
}
