% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{GraphNode}
\alias{GraphNode}
\title{Graph Node}
\description{
Virtual base class for nodes in an \code{\link{AnvilGraph}}.
Is either a \code{\link{GraphValue}} or a \code{\link{GraphLiteral}}.
Cannot be instantiated directly - use \code{\link[=GraphValue]{GraphValue()}} or \code{\link[=GraphLiteral]{GraphLiteral()}} instead.
}
