% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{GraphValue}
\alias{GraphValue}
\title{Graph Value}
\usage{
GraphValue(aval = ShapedTensor())
}
\arguments{
\item{aval}{(\code{\link{ShapedTensor}})\cr
The abstract value of the variable.}
}
\description{
Value in a \code{\link{Graph}}.
}
