% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{GraphValue}
\alias{GraphValue}
\title{Graph Value}
\usage{
GraphValue(aval)
}
\arguments{
\item{aval}{(\code{\link{AbstractTensor}})\cr
The abstract value of the variable.}
}
\value{
(\code{GraphValue})
}
\description{
Value in an \code{\link{AnvilGraph}}. This is a mutable class.
}
