% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\alias{repositories}
\alias{repository}
\alias{BINARY_BASE_URL}
\alias{add_libpaths}
\title{Discover and install binary packages for fast installation}
\usage{
install(
  pkgs,
  lib = .libPaths()[1],
  ...,
  version = BiocManager::version(),
  binary_base_url = BINARY_BASE_URL,
  verbose = getOption("verbose")
)

repositories(
  version = BiocManager::version(),
  binary_base_url = BINARY_BASE_URL
)

repository(version = BiocManager::version(), binary_base_url = BINARY_BASE_URL)

add_libpaths(paths)
}
\arguments{
\item{pkgs}{`character()` packages to install from binary repository.}

\item{lib}{`character(1)` library path (directory) in which to
install `pkgs`; defaults to `.libPaths()[1]`.}

\item{...}{additional arguments, passed to `install.packages()`.}

\item{version}{`character(1)` or `package_version` Bioconductor
version, e.g., "3.12".}

\item{binary_base_url}{`character(1)` host and base path for binary
package 'CRAN-style' repository; not usually required by the
end-user.}

\item{verbose}{`logical(1)` report on package installation
progress?}

\item{paths}{`character()`: vector of directories to add to
`.libPaths()`. Paths that do not exist will be created.}
}
\value{
`install()`: return value of `install.packages()`.

`repositories()`: character() of binary (if available),
    Bioconductor, and CRAN repositories.

`repository()`: character(1) location of binary repository,
    if available, or character(0) if not.

`add_libpaths()`: updated .libPaths(), invisibly.
}
\description{
`install()`: install R / Bioconductor packages, using
    fast pre-built 'binary' libraries if available.

`repositories()`: repositories to search for binary
    (if available), Bioconductor, and CRAN packages.

`repository()`: the location of the repository of
    binary packages for fast installation, if available.

`add_libpaths()`: Add local library paths to
    `.libPaths()`.
}
\details{
`repositories()` prepends an additional repository URI to
    `BiocManager::repositories()`. The URI is formed by
    concatenating `binary_base_url`, the environment variables
    `TERRA_R_PLATFORM` and the 'major' and 'minor' components of
    `TERRA_R_PLATFORM_BINARY_VERSION` and
    `BiocManager::version()`. The URI is only prepended if a
    CRAN-style repostiory exists at that location, with binary
    package tar.gz content described by `src/contrib/PACKAGES.gz`.

The unexported URL to the base repository is available
    with `AnVIL:::BINARY_BASE_URL`.
}
\examples{
\dontrun{install(c('BiocParallel', 'BiocGenerics'))}

repositories()

repository()

\dontrun{add_libpaths("/tmp/host-site-library")}

}
