% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{IotaTensor}
\alias{IotaTensor}
\title{Iota Tensor Class}
\usage{
IotaTensor(shape, dtype, dimension, start = 1L, ambiguous = FALSE)
}
\arguments{
\item{shape}{(\code{\link[stablehlo:Shape]{stablehlo::Shape}} | \code{integer()})\cr
The shape of the tensor.}

\item{dtype}{(\code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
The data type.}

\item{dimension}{(\code{integer(1)})\cr
The dimension along which values increase.}

\item{start}{(\code{integer(1)})\cr
The starting value.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the type is ambiguous.
Ambiguous types usually arise from R literals (e.g., \code{1L}, \code{1.0}) and follow special promotion rules.
See the vignette "Type Promotion" for more details.}
}
\description{
An \code{\link{AbstractTensor}} representing a tensor where the data is a sequence of integers.
}
