% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{IotaTensor}
\alias{IotaTensor}
\title{Iota Tensor Class}
\usage{
IotaTensor(shape, dimension, dtype, ambiguous = FALSE)
}
\arguments{
\item{shape}{(\code{\link[stablehlo:Shape]{stablehlo::Shape}} | \code{integer()})\cr
The shape of the tensor.}

\item{dimension}{(\code{integer(1)})\cr}
}
\description{
A \code{\link{AbstractTensor}} representing a tensor where the data is a R scalar literal (e.g., \code{1L}, \code{2.5})
or an \code{\link{AnvilTensor}}.
Usually, their type is ambiguous, unless created via \code{\link{nv_fill}}.
}
