% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpreter-jit.R
\name{jit}
\alias{jit}
\title{JIT compile a function}
\usage{
jit(f, static = character())
}
\arguments{
\item{f}{(\code{function})\cr
Function to compile.}

\item{static}{(\code{character()})\cr
Which parameters of \code{f} are static.}
}
\value{
(\code{function})
}
\description{
Convert a function to a JIT compiled function.
}
