% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpreter-jit.R
\name{jit}
\alias{jit}
\title{JIT compile a function}
\usage{
jit(f)
}
\arguments{
\item{f}{(\code{function})\cr
Function to compile.}
}
\value{
(\code{function})
}
\description{
Convert a function to a JIT compiled function.
}
