% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jit.R
\name{jit}
\alias{jit}
\title{JIT compile a function}
\usage{
jit(f, static = character(), device = NULL, cache_size = 100L)
}
\arguments{
\item{f}{(\code{function})\cr
Function to compile.}

\item{static}{(\code{character()})\cr
Which parameters of \code{f} are static.}

\item{device}{(\code{NULL} | \code{character(1)} | \code{\link[pjrt:pjrt_device]{PJRTDevice}})\cr
The device to use for the compiled function.
The default (\code{NULL}) uses the \code{PJRT_PLATFORM} environment variable or defaults to "cpu".}

\item{cache_size}{(\code{integer(1)})\cr
The size of the cache for the jit-compiled functions.}
}
\value{
(\code{function})
}
\description{
Convert a function to a JIT compiled function.
}
