% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jit.R
\name{jit}
\alias{jit}
\title{JIT compile a function}
\usage{
jit(
  f,
  static = character(),
  cache_size = 100L,
  donate = character(),
  device = NULL
)
}
\arguments{
\item{f}{(\code{function})\cr
Function to compile.}

\item{static}{(\code{character()})\cr
Which parameters of \code{f} are static.}

\item{cache_size}{(\code{integer(1)})\cr
The size of the cache for the jit-compiled functions.}

\item{donate}{(\code{character()})\cr
Names of the arguments whose buffers should be donated.
Donated buffers can be aliased with outputs of the same type,
allowing in-place operations and reducing memory usage.}

\item{device}{(\code{NULL} | \code{character(1)} | \code{\link[pjrt:pjrt_device]{PJRTDevice}})\cr
The device to use if no input tensors are provided.
The default is to infer it from the arguments.}
}
\value{
(\code{function})
}
\description{
Convert a function to a JIT compiled function.
}
