% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jit.R
\name{jit_eval}
\alias{jit_eval}
\title{Jit an Evaluate an Expression}
\usage{
jit_eval(expr)
}
\arguments{
\item{expr}{(\code{expression})\cr
Expression to run.}
}
\value{
(\code{any})\cr
Result of the expression.
}
\description{
Compiles and evaluates an expression.
Evaluates in a copy of the calling environment to avoid side effects.
}
