% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{LiteralTensor}
\alias{LiteralTensor}
\title{Literal Tensor Class}
\usage{
LiteralTensor(data, shape, dtype = default_dtype(data))
}
\arguments{
\item{data}{(\code{numeric(1)} | \code{integer(1)} | \code{logical(1)})\cr
The scalar value.}

\item{shape}{(\code{\link[stablehlo:Shape]{stablehlo::Shape}} | \code{integer()})\cr
The shape of the tensor.}

\item{dtype}{(\code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
The data type. Defaults to \code{f32} for numeric, \code{i32} for integer, \code{i1} for logical.}
}
\description{
A \code{\link{ShapedTensor}} representing a tensor where the data is a R scalar literal (e.g., \code{1L}, \code{2.5}, \code{TRUE}).
Their type is ambiguous, and they adapt (if possible) to the types of
non-literal tensors they interact with.
}
\seealso{
\link{ShapedTensor}, \link{ConcreteTensor}
}
