% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{local_descriptor}
\alias{local_descriptor}
\title{Create a graph}
\usage{
local_descriptor(..., envir = parent.frame())
}
\arguments{
\item{envir}{(\code{environment})\cr
Environment where exit handler will be registered for cleaning up the
\code{\link{Graph}} if it was not returned yet.}
}
\value{
A \code{\link{Graph}} object.
}
\description{
Creates a new \code{\link{Graph}} which is afterwards accessible via \code{\link[=.current_descriptor]{.current_descriptor()}}.
The graph is automatically removed when exiting the current scope.
After the graph is either cleaned up automatically (by exiting the scope)
or finalized, the previously built graph is restored,
i.e., accessible via \code{\link[=.current_descriptor]{.current_descriptor()}}.
}
