% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize.R
\name{localize}
\alias{localize}
\alias{delocalize}
\title{Copy packages, folders, or files to or from google buckets.}
\usage{
localize(source, destination, dry = TRUE)

delocalize(source, destination, unlink = FALSE, dry = TRUE)
}
\arguments{
\item{source}{`character(1)`, a google storage bucket or local file
system directory location.}

\item{destination}{`character(1)`, a google storage bucket or local
file system directory location.}

\item{dry}{`logical(1)`, when `TRUE` (default), return the
consequences of the operation without actually performing the
operation.}

\item{unlink}{`logical(1)` remove (unlink) the file or directory
in `source`. Default: `FALSE`.}
}
\value{
`localize()`: exit status of function `gsutil_rsync()`.

`delocalize()`: exit status of function `gsutil_rsync()`
}
\description{
`localize()`: recursively synchronizes files from a
    Google storage bucket (`source`) to the local file system
    (`destination`). This command acts recursively on the `source`
    directory, and does not delete files in `destination` that are
    not in `source.

`delocalize()`: synchronize files from a local file
    system (`source`) to a Google storage bucket
    (`destination`). This command acts recursively on the `source`
    directory, and does not delete files in `destination` that are
    not in `source`.
}
