% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize.R
\name{localize}
\alias{localize}
\alias{delocalize}
\alias{add_libpaths}
\alias{install}
\title{Copy packages, folders, or files to or from google buckets.}
\usage{
localize(source, destination, dry = TRUE)

delocalize(source, destination, unlink = FALSE, dry = TRUE)

add_libpaths(paths)

install(
  pkgs,
  lib = .libPaths()[1],
  ...,
  binary_base_url = "https://storage.googleapis.com",
  verbose = getOption("verbose")
)
}
\arguments{
\item{source}{`character(1)`, a google storage bucket or local file
system directory location.}

\item{destination}{`character(1)`, a google storage bucket or local
file system directory location.}

\item{dry}{`logical(1)`, when `TRUE` (default), return the
consequences of the operation without actually performing the
operation.}

\item{unlink}{`logical(1)` remove (unlink) the file or directory
in `source`. Default: `FALSE`.}

\item{paths}{`character()`: vector of directories to add to
`.libPaths()`. Paths that do not exist will be created.}

\item{pkgs}{`character()` packages to install from binary repository.}

\item{lib}{`character(1)` library path (directory) in which to
install `pkgs`; defaults to `.libPaths()[1]`.}

\item{...}{additional arguments, passed to `install.packages()`.}

\item{binary_base_url}{`character(1)` host and base path for binary
package 'CRAN-style' repository; not usually required by the
end-user.}

\item{verbose}{`logical(1)` report on package installation
progress?}
}
\value{
`localize()`: exit status of function `gsutil_rsync()`.

`delocalize()`: exit status of function `gsutil_rsync()`

`add_libpaths()`: updated .libPaths(), invisibly.

`install()`: return value of `install.packages()`.
}
\description{
`localize()`: recursively synchronizes files from a
    Google storage bucket (`source`) to the local file system
    (`destination`). This command acts recursively on the `source`
    directory, and does not delete files in `destination` that are
    not in `source.

`delocalize()`: synchronize files from a local file
    system (`source`) to a Google storage bucket
    (`destination`). This command acts recursively on the `source`
    directory, and does not delete files in `destination` that are
    not in `source`.

`add_libpaths()`: Add local library paths to
    `.libPaths()`.

`install()`: install R / Bioconductor packages, using
    fast pre-built 'binary' libraries if available.
}
\details{
`install()` prepends an additional repository URI to
    `BiocManager::repositories()`. The URI is formed by
    concatenating `binary_base_url`, the environment variables
    `R_PLATFORM` and the 'major' and 'minor' components of
    `R_PLATFORM_BINARY_VERSION` and `BiocManager::version()`. The
    URI is only prepended if a CRAN-style repostiory exists at that
    location, with binary package tar.gz content described by
    `src/contrib/PACKAGES`.
}
\examples{
\dontrun{
add_libpaths("/tmp/host-site-library")
install(packages = c('BiocParallel', 'BiocGenerics'))
}

}
