% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize.R
\name{localize}
\alias{localize}
\alias{delocalize}
\alias{add_libpaths}
\alias{install}
\alias{install_precompiled}
\title{Copy packages, folders, or files to or from google buckets.}
\usage{
localize(source, destination, dry = TRUE)

delocalize(source, destination, unlink = FALSE, dry = TRUE)

add_libpaths(paths)

install(pkgs, lib = .libPaths()[1], lib.loc = NULL)

install_precompiled(
  pkgs,
  lib = .libPaths()[1],
  lib.loc = NULL,
  repos = "gs://biocbbs_2020a/zpacks",
  verbose = getOption("verbose")
)
}
\arguments{
\item{source}{`character(1)`, a google storage bucket or local file
system directory location.}

\item{destination}{`character(1)`, a google storage bucket or local
file system directory location.}

\item{dry}{`logical(1)`, when `TRUE` (default), return the
consequences of the operation without actually performing the
operation.}

\item{unlink}{`logical(1)` remove (unlink) the file or directory
in `source`. Default: `FALSE`.}

\item{paths}{`character()`: vector of directories to add to
`.libPaths()`. Paths that do not exist will be created.}

\item{pkgs}{`character()` packages to install from binary repository.}

\item{lib}{`character(1)` library path (directory) in which to
install `pkgs`; defaults to `.libPaths()[1]`.}

\item{lib.loc}{`character()` library locations to search for
currently installed packages. Default is all paths in
`.libPaths()`.}
}
\value{
`localize()`: exit status of function `gsutil_rsync()`.

`delocalize()`: exit status of function `gsutil_rsync()`

`add_libpaths()`: updated .libPaths(), invisibly.

`install()`: location of installed packages, invisibly.

`install_precompiled()`: NULL, invisibly.
}
\description{
`localize()`: recursively synchronizes files from a
    Google storage bucket (`source`) to the local file system
    (`destination`). This command acts recursively on the `source`
    directory, and does not delete files in `destination` that are
    not in `source.

`delocalize()`: synchronize files from a local file
    system (`source`) to a Google storage bucket
    (`destination`). This command acts recursively on the `source`
    directory, and does not delete files in `destination` that are
    not in `source`.

`add_libpaths()`: Add local library paths to
    `.libPaths()`.

`install()`: DEPRECATED; use `install_precompiled()`

`install_precompiled()`: install packages and their
    dependencies from pre-compiled versions of the packages,
    archived in a google cloud bucket.
}
\examples{

\dontrun{
   localize("gs://anvil-bioc", "/tmp/test-localize")
}


\dontrun{
    delocalize("/tmp/test-localize/", "gs://anvil-bioc")
}


\dontrun{
   add_libpaths('/tmp/my_library')
   localize('gs://bioconductor-full-devel', '/tmp/my_library')
}

\dontrun{
add_libpaths("/tmp/host-site-library")
install(packages = c('BiocParallel', 'BiocGenerics'))
}

\dontrun{
add_libpaths("/tmp/host-site-library")
install(packages = c('BiocParallel', 'BiocGenerics'))
}

}
