% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize.R
\name{localize}
\alias{localize}
\title{Localize a Google storage bucket to the container.}
\usage{
localize(source, destination, ...)
}
\arguments{
\item{source}{character, a valid path to a google storage
bucket.}

\item{destination}{character vector, representing a local path.}

\item{...}{additional arguments to `gsutil_rsync()`}
}
\value{
Exit status of function, `gsutil_rsync()`
}
\description{
Localize a Google storage bucket to the container.
}
\details{
localize a bucket from the google cloud storage to a local
    path on the machine.
}
\examples{

\dontrun{
   localize("gs://anvil-bioc", "/tmp/test-localize")
}

}
