% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localize.R
\name{localize}
\alias{localize}
\alias{delocalize}
\alias{install}
\alias{repositories}
\alias{add_libpaths}
\title{Copy packages, folders, or files to or from google buckets.}
\usage{
localize(source, destination, dry = TRUE)

delocalize(source, destination, unlink = FALSE, dry = TRUE)

install(
  pkgs,
  lib = .libPaths()[1],
  ...,
  version = BiocManager::version(),
  binary_base_url = "https://storage.googleapis.com",
  verbose = getOption("verbose")
)

repositories(
  version = BiocManager::version(),
  binary_base_url = "https://storage.googleapis.com"
)

add_libpaths(paths)
}
\arguments{
\item{source}{`character(1)`, a google storage bucket or local file
system directory location.}

\item{destination}{`character(1)`, a google storage bucket or local
file system directory location.}

\item{dry}{`logical(1)`, when `TRUE` (default), return the
consequences of the operation without actually performing the
operation.}

\item{unlink}{`logical(1)` remove (unlink) the file or directory
in `source`. Default: `FALSE`.}

\item{pkgs}{`character()` packages to install from binary repository.}

\item{lib}{`character(1)` library path (directory) in which to
install `pkgs`; defaults to `.libPaths()[1]`.}

\item{...}{additional arguments, passed to `install.packages()`.}

\item{version}{`character(1)` or `package_version` Bioconductor
version, e.g., "3.12".}

\item{binary_base_url}{`character(1)` host and base path for binary
package 'CRAN-style' repository; not usually required by the
end-user.}

\item{verbose}{`logical(1)` report on package installation
progress?}

\item{paths}{`character()`: vector of directories to add to
`.libPaths()`. Paths that do not exist will be created.}
}
\value{
`localize()`: exit status of function `gsutil_rsync()`.

`delocalize()`: exit status of function `gsutil_rsync()`

`install()`: return value of `install.packages()`.

`repositories()`: character() of binary (if available),
    Bioconductor, and CRAN repositories.

`add_libpaths()`: updated .libPaths(), invisibly.
}
\description{
`localize()`: recursively synchronizes files from a
    Google storage bucket (`source`) to the local file system
    (`destination`). This command acts recursively on the `source`
    directory, and does not delete files in `destination` that are
    not in `source.

`delocalize()`: synchronize files from a local file
    system (`source`) to a Google storage bucket
    (`destination`). This command acts recursively on the `source`
    directory, and does not delete files in `destination` that are
    not in `source`.

`install()`: install R / Bioconductor packages, using
    fast pre-built 'binary' libraries if available.

`repositories()`: repositories to search for binary
    (if available), Bioconductor, and CRAN packages.

`add_libpaths()`: Add local library paths to
    `.libPaths()`.
}
\details{
`repositories()` prepends an additional repository URI to
    `BiocManager::repositories()`. The URI is formed by
    concatenating `binary_base_url`, the environment variables
    `TERRA_R_PLATFORM` and the 'major' and 'minor' components of
    `TERRA_R_PLATFORM_BINARY_VERSION` and
    `BiocManager::version()`. The URI is only prepended if a
    CRAN-style repostiory exists at that location, with binary
    package tar.gz content described by `src/contrib/PACKAGES.gz`.
}
\examples{
\dontrun{install(c('BiocParallel', 'BiocGenerics'))}

repositories()

\dontrun{add_libpaths("/tmp/host-site-library")}

}
