% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{ndims}
\alias{ndims}
\title{Get the number of dimensions of a tensor}
\usage{
ndims(x)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
A tensor-like object.}
}
\value{
\code{integer(1)}
}
\description{
Returns the number of dimensions (sometimes also refered to as rank) of a tensor.
Equivalent to \code{length(shape(x))}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_tensor(1:4, dtype = "f32")
ndims(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[tengen:ndims]{tengen::ndims()}}
}
