% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_abs}
\alias{nv_abs}
\title{Absolute Value}
\usage{
nv_abs(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise absolute value. You can also use \code{abs()}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(-1, 2, -3))
  abs(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_abs]{nvl_abs()}} for the underlying primitive.
}
