% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_abs}
\alias{nv_abs}
\title{Absolute Value}
\usage{
nv_abs(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise absolute value.
}
