% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_binary_ops}
\alias{nv_binary_ops}
\alias{nv_add}
\alias{nv_mul}
\alias{nv_sub}
\alias{nv_div}
\title{Binary Operations}
\usage{
nv_add(lhs, rhs)

nv_mul(lhs, rhs)

nv_sub(lhs, rhs)

nv_div(lhs, rhs)
}
\arguments{
\item{lhs}{(\code{\link{nv_tensor}})}

\item{rhs}{(\code{\link{nv_tensor}})}
}
\value{
\code{\link{nv_tensor}}
}
\description{
Binary operations on tensors.
}
