% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_binary_ops}
\alias{nv_binary_ops}
\alias{nv_add}
\alias{nv_mul}
\alias{nv_sub}
\alias{nv_div}
\alias{nv_pow}
\alias{nv_eq}
\alias{nv_ne}
\alias{nv_gt}
\alias{nv_ge}
\alias{nv_lt}
\alias{nv_le}
\alias{nv_max}
\alias{nv_min}
\alias{nv_remainder}
\alias{nv_and}
\alias{nv_or}
\alias{nv_xor}
\alias{nv_shift_left}
\alias{nv_shift_right_logical}
\alias{nv_shift_right_arithmetic}
\alias{nv_atan2}
\title{Binary Operations}
\usage{
nv_add(lhs, rhs)

nv_mul(lhs, rhs)

nv_sub(lhs, rhs)

nv_div(lhs, rhs)

nv_pow(lhs, rhs)

nv_eq(lhs, rhs)

nv_ne(lhs, rhs)

nv_gt(lhs, rhs)

nv_ge(lhs, rhs)

nv_lt(lhs, rhs)

nv_le(lhs, rhs)

nv_max(lhs, rhs)

nv_min(lhs, rhs)

nv_remainder(lhs, rhs)

nv_and(lhs, rhs)

nv_or(lhs, rhs)

nv_xor(lhs, rhs)

nv_shift_left(lhs, rhs)

nv_shift_right_logical(lhs, rhs)

nv_shift_right_arithmetic(lhs, rhs)

nv_atan2(lhs, rhs)
}
\arguments{
\item{lhs}{(\code{\link{nv_tensor}})}

\item{rhs}{(\code{\link{nv_tensor}})}
}
\value{
\code{\link{nv_tensor}}
}
\description{
Binary operations on tensors.
}
\examples{
# Comparison operators such `nv_eq`, `nv_le`, `nv_gt`, etc
# are nondifferentiable and contribute zero to gradients.
relu <- function(x) {
  nv_convert(x > nv_scalar(0), "f32")*x
}
# df/dx = 1 if x > 0 else 0
g_relu <- jit(gradient(relu, "x"))

g_relu(nv_scalar(1, dtype = "f32"))
g_relu(nv_scalar(-1, dtype = "f32"))
}
