% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_bitcast_convert}
\alias{nv_bitcast_convert}
\title{Bitcast Conversion}
\usage{
nv_bitcast_convert(operand, dtype)
}
\arguments{
\item{operand}{tensor}

\item{dtype}{requested dtype}
}
\description{
Reinterpret Bits
}
