% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_broadcast_scalars}
\alias{nv_broadcast_scalars}
\title{Broadcast Scalars to Common Shape}
\usage{
nv_broadcast_scalars(...)
}
\arguments{
\item{...}{(\code{\link{nv_tensor}})\cr
Tensors to broadcast. Scalars will be broadcast to the common non-scalar shape.}
}
\value{
(\code{list()} of \code{\link{nv_tensor}})
}
\description{
Broadcast scalar tensors to match the shape of non-scalar tensors.
All non-scalar tensors must have the same shape.
}
