% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_broadcast_tensors}
\alias{nv_broadcast_tensors}
\title{Broadcast Tensors to a Common Shape}
\usage{
nv_broadcast_tensors(...)
}
\arguments{
\item{...}{(\code{list()} of \code{\link{nv_tensor}})}
}
\value{
(\code{list()} of \code{\link{nv_tensor}})
}
\description{
Broadcast tensors to a common shape.
}
\section{Broadcasting Rules}{

We follow the standard NumPy broadcasting rules:
\enumerate{
\item If the tensors have different numbers of dimensions, prepend 1s to the shape of the smaller tensor.
\item For each dimensions, if:
\itemize{
\item If the sizes are the same, do nothing.
\item One of the tensors has size 1, expand it to the size of the other tensor.
\item If the sizes are different and neither is 1, raise an error.
}
}
}

