% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_broadcast_tensors}
\alias{nv_broadcast_tensors}
\title{Broadcast Tensors to a Common Shape}
\usage{
nv_broadcast_tensors(...)
}
\arguments{
\item{...}{(\code{list()} of \code{\link{nv_tensor}})}
}
\value{
(\code{list()} of \code{\link{nv_tensor}})
}
\description{
Broadcast tensors to a common shape.
}
