% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_broadcast_to}
\alias{nv_broadcast_to}
\title{Broadcast}
\usage{
nv_broadcast_to(operand, shape)
}
\arguments{
\item{operand}{(\code{\link{nv_tensor}})\cr
Operand.}

\item{shape}{(\code{integer()})\cr
Output shape.}
}
\value{
(\code{\link{nv_tensor}})
}
\description{
Broadcast
}
