% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_cbrt}
\alias{nv_cbrt}
\title{Cube Root}
\usage{
nv_cbrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise cube root.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 8, 27))
  nv_cbrt(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_cbrt]{nvl_cbrt()}} for the underlying primitive.
}
