% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_cbrt}
\alias{nv_cbrt}
\title{Cube Root}
\usage{
nv_cbrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise cube root.
}
