% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_ceil}
\alias{nv_ceil}
\title{Ceiling}
\usage{
nv_ceil(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise ceiling (round toward positive infinity).
}
