% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_concatenate}
\alias{nv_concatenate}
\title{Concatenate}
\usage{
nv_concatenate(..., dimension = NULL)
}
\arguments{
\item{...}{tensors}

\item{dimension}{(\code{integer()})\cr
The dimension to concatenate along to. Other dimensions must be the same.
If this is \code{NULL} (default), it assumes all ranks are at most 1 and the concatenation dimension is 1.}
}
\value{
\code{\link{tensorish}}
}
\description{
Concatenate a variadic number of tensors.
}
