% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_concatenate}
\alias{nv_concatenate}
\title{Concatenate}
\usage{
nv_concatenate(..., dimension)
}
\arguments{
\item{...}{tensors}

\item{dimension}{(\code{integer()})\cr
The dimension to concatenate along to. Other dimensions must be the same.}

\item{operand}{(\code{\link{nv_tensor}})\cr
Operand.}
}
\value{
\code{\link{nv_tensor}}
}
\description{
Concatenate a variadic amaount of tensors.
}
