% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_convert}
\alias{nv_convert}
\title{Convert Tensor to Different Data Type}
\usage{
nv_convert(operand, dtype)
}
\arguments{
\item{operand}{(\code{\link{nv_tensor}})\cr
Operand.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}
}
\value{
\code{\link{nv_tensor}}
}
\description{
Convert a tensor to a different data type.
}
