% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_cosine}
\alias{nv_cosine}
\title{Cosine}
\usage{
nv_cosine(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise cosine. You can also use \code{cos()}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(0, pi / 2, pi))
  cos(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_cosine]{nvl_cosine()}} for the underlying primitive.
}
