% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/api-generics.R
\name{nv_crossprod}
\alias{nv_crossprod}
\alias{crossprod.AnvlBox}
\title{Cross Product (Matrix)}
\usage{
nv_crossprod(x, y = NULL)

\method{crossprod}{AnvlBox}(x, y = NULL, ...)
}
\arguments{
\item{x}{(\code{\link{arrayish}})\cr
An array with at least 2 dimensions.}

\item{y}{(\code{\link{arrayish}} | \code{NULL})\cr
Optional second array. If \code{NULL}, uses \code{x}.}

\item{...}{Unused.}
}
\value{
\code{\link{arrayish}}
}
\description{
Computes \code{t(x) \%*\% y}. If \code{y} is missing, computes \code{t(x) \%*\% x}.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(matrix(1:6, nrow = 3), dtype = "f32")
nv_crossprod(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_tcrossprod]{nv_tcrossprod()}}, \code{\link[=nv_matmul]{nv_matmul()}}
}
