% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/api-generics.R
\name{nv_crossprod}
\alias{nv_crossprod}
\alias{crossprod.AnvilBox}
\title{Cross Product (Matrix)}
\usage{
nv_crossprod(x, y = NULL)

\method{crossprod}{AnvilBox}(x, y = NULL, ...)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
A tensor with at least 2 dimensions.}

\item{y}{(\code{\link{tensorish}} | \code{NULL})\cr
Optional second tensor. If \code{NULL}, uses \code{x}.}
}
\value{
\code{\link{tensorish}}
}
\description{
Computes \code{t(x) \%*\% y}. If \code{y} is missing, computes \code{t(x) \%*\% x}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(matrix(1:6, nrow = 3), dtype = "f32")
  nv_crossprod(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_tcrossprod]{nv_tcrossprod()}}, \code{\link[=nv_matmul]{nv_matmul()}}
}
