% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{nv_device}
\alias{nv_device}
\title{Create a Device}
\usage{
nv_device(x, backend = default_backend())
}
\arguments{
\item{x}{(\code{character(1)})\cr
Identifier for the device.
E.g. \code{"cpu"}, \code{"cuda"}, or \code{"cuda:<n>"} (for the n-th GPU).}

\item{backend}{(\code{character(1)})\cr
The backend for which to create the device.}
}
\value{
A backend-specific device object (e.g. \code{PJRTDevice} for \code{"xla"},
\code{\link{quickr_device}} for \code{"quickr"}).
}
\description{
Constructs a backend-specific device object.

A device identifies a compute resources, such as CPU, or a specific GPU.
It is relevant for data allocation (e.g. via \code{\link[=nv_array]{nv_array()}}) but also compilation (\link{jit}).
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
# Create CPU device for xla backend:
nv_device("cpu", "xla")
# Create CPU device for quickr backend:
nv_device("cpu", "quickr")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=backend]{backend()}}, \code{\link[=AnvilBackend]{AnvilBackend()}}.
}
