% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_div}
\alias{nv_div}
\title{Division}
\usage{
nv_div(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise division of two tensors.
}
