% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_eq}
\alias{nv_eq}
\title{Equal}
\usage{
nv_eq(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise equality comparison.
}
