% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_exp}
\alias{nv_exp}
\title{Exponential}
\usage{
nv_exp(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise exponential function.
}
