% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_expm1}
\alias{nv_expm1}
\title{Exponential Minus One}
\usage{
nv_expm1(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise exp(x) - 1, more accurate for small x.
}
