% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_extract_diag}
\alias{nv_extract_diag}
\title{Extract Diagonal}
\usage{
nv_extract_diag(x)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
A 2-D tensor (matrix).}
}
\value{
\code{\link{tensorish}}\cr
A 1-D tensor of length \code{min(nrow, ncol)} containing the diagonal elements.
}
\description{
Extracts the diagonal elements from a 2-D tensor.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(1:9, shape = c(3, 3))
  nv_extract_diag(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_diag]{nv_diag()}} for creating a diagonal matrix, \code{\link[=nv_trace]{nv_trace()}}
}
