% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_eye}
\alias{nv_eye}
\title{Identity Matrix}
\usage{
nv_eye(n, dtype = "f32")
}
\arguments{
\item{n}{(\code{integer(1)})\cr
Size of the identity matrix.}

\item{dtype}{(\code{character(1)} | \code{\link[tengen:TensorDataType]{tengen::TensorDataType}})\cr
Data type.}
}
\value{
\code{\link{tensorish}}\cr
An \verb{n x n} identity matrix.
}
\description{
Creates an \verb{n x n} identity matrix.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval(nv_eye(3L))
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_diag]{nv_diag()}} for general diagonal matrices.
}
