% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_fill}
\alias{nv_fill}
\title{Constant}
\usage{
nv_fill(value, shape, dtype = NULL, ambiguous = FALSE)
}
\arguments{
\item{value}{(any)\cr
Value.}

\item{shape}{(integer())\cr
Shape.}

\item{dtype}{(character(1))\cr
Data type.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the type is ambiguous.
Ambiguous types usually arise from R literals (e.g., \code{1L}, \code{1.0}) and follow special promotion rules.
See the \code{vignette("type-promotion")} for more details.}
}
\value{
\code{\link{tensorish}}
}
\description{
Create a constant.
}
