% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_floor}
\alias{nv_floor}
\title{Floor}
\usage{
nv_floor(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise floor (round toward negative infinity). You can also use \code{floor()}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1.2, 2.7, -1.5))
  floor(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_floor]{nvl_floor()}} for the underlying primitive.
}
