% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_ge}
\alias{nv_ge}
\title{Greater Than or Equal}
\usage{
nv_ge(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise greater than or equal comparison.
}
