% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_if}
\alias{nv_if}
\title{If}
\usage{
nv_if(pred, true, false)
}
\arguments{
\item{pred}{(\link{nv_tensor})\cr
Flag.}

\item{true}{(NSE)\cr
Expression to evaluate if the condition is true.}

\item{false}{(NSE)\cr
Expression to evaluate if the condition is false.}
}
\value{
\code{\link{nv_tensor}}
}
\description{
Functional if statement.
}
