% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_ifelse}
\alias{nv_ifelse}
\title{Conditional Element Selection}
\usage{
nv_ifelse(pred, true_value, false_value)
}
\arguments{
\item{pred}{(\code{\link{arrayish}} of boolean type)\cr
Predicate array. Must be scalar or the same shape as \code{true_value}.}

\item{true_value}{(\code{\link{arrayish}})\cr
Values to return where \code{pred} is \code{TRUE}.}

\item{false_value}{(\code{\link{arrayish}})\cr
Values to return where \code{pred} is \code{FALSE}.
Must have the same shape and data type as \code{true_value}.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as \code{true_value}.
}
\description{
Selects elements from \code{true_value} or \code{false_value} based on \code{pred},
analogous to R's \code{\link[=ifelse]{ifelse()}}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  pred <- nv_array(c(TRUE, FALSE, TRUE))
  nv_ifelse(pred, nv_array(c(1, 2, 3)), nv_array(c(4, 5, 6)))
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_ifelse]{nvl_ifelse()}} for the underlying primitive.
}
