% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_iota}
\alias{nv_iota}
\title{Iota}
\usage{
nv_iota(dim, shape, dtype = "i32", start = 1)
}
\arguments{
\item{dim}{(\code{integer(1)})\cr
The dimension along which to generate increasing values.}

\item{shape}{(\code{integer()})\cr
Shape.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{start}{(\code{numeric(1)})\cr
The value to start the sequence at (default 1).}
}
\description{
Create a tensor with increasing values along a dimension.
}
