% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_iota}
\alias{nv_iota}
\title{Iota}
\usage{
nv_iota(dim, dtype, shape)
}
\arguments{
\item{dim}{(\code{integer(1)})\cr
Dimension along which values increase.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{shape}{(\code{integer()})\cr
Shape.}
}
\value{
\code{\link{tensorish}}
}
\description{
Creates a tensor with values increasing along the specified dimension.
}
