% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/api-generics.R
\name{nv_is_finite}
\alias{nv_is_finite}
\alias{is.finite.AnvilBox}
\title{Is Finite}
\usage{
nv_is_finite(operand)

\method{is.finite}{AnvilBox}(x)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Operand.}

\item{x}{(\code{\link{arrayish}})\cr Operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape as the input and boolean data type.
}
\description{
Element-wise check if values are finite (not \code{Inf}, \code{-Inf}, or \code{NaN}).
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(1, Inf, NaN, -Inf, 0))
nv_is_finite(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=prim_is_finite]{prim_is_finite()}} for the underlying primitive.
}
