% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_is_finite}
\alias{nv_is_finite}
\title{Is Finite}
\usage{
nv_is_finite(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}} of boolean type
}
\description{
Element-wise check if values are finite (not Inf, -Inf, or NaN).
}
