% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_is_finite}
\alias{nv_is_finite}
\title{Is Finite}
\usage{
nv_is_finite(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape as the input and boolean data type.
}
\description{
Element-wise check if values are finite (not \code{Inf}, \code{-Inf}, or \code{NaN}).
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, Inf, NaN, -Inf, 0))
  nv_is_finite(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_is_finite]{nvl_is_finite()}} for the underlying primitive.
}
