% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/api-generics.R
\name{nv_is_infinite}
\alias{nv_is_infinite}
\alias{is.infinite.AnvilBox}
\title{Is Infinite}
\usage{
nv_is_infinite(operand)

\method{is.infinite}{AnvilBox}(x)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape as the input and boolean data type.
}
\description{
Element-wise check if values are infinite (\code{Inf} or \code{-Inf}).
You can also use \code{is.infinite()}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, NaN, Inf, -Inf, 0))
  nv_is_infinite(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_is_finite]{nv_is_finite()}}, \code{\link[=nv_is_nan]{nv_is_nan()}}
}
