% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_log}
\alias{nv_log}
\title{Natural Logarithm}
\usage{
nv_log(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise natural logarithm. You can also use \code{log()}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2.718, 7.389))
  log(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_log]{nvl_log()}} for the underlying primitive.
}
