% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_log}
\alias{nv_log}
\title{Natural Logarithm}
\usage{
nv_log(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise natural logarithm.
}
