% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_log1p}
\alias{nv_log1p}
\title{Log Plus One}
\usage{
nv_log1p(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise log(1 + x), more accurate for small x.
}
