% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_log1p}
\alias{nv_log1p}
\title{Log Plus One}
\usage{
nv_log1p(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise \code{log(1 + x)}, more accurate for small \code{x}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(0, 0.001, 1))
  nv_log1p(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_log1p]{nvl_log1p()}} for the underlying primitive.
}
