% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_log2}
\alias{nv_log2}
\title{Base-2 Logarithm}
\usage{
nv_log2(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise base-2 logarithm. You can also use \code{log2()}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2, 4, 8))
  nv_log2(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_log]{nv_log()}}, \code{\link[=nv_log10]{nv_log10()}}
}
