% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_logistic}
\alias{nv_logistic}
\title{Logistic (Sigmoid)}
\usage{
nv_logistic(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logistic sigmoid: 1 / (1 + exp(-x)).
}
