% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_lt}
\alias{nv_lt}
\title{Less Than}
\usage{
nv_lt(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise less than comparison.
}
