% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_matmul}
\alias{nv_matmul}
\title{Matrix Multiplication}
\usage{
nv_matmul(lhs, rhs)
}
\arguments{
\item{lhs}{(\code{\link{tensorish}})}

\item{rhs}{(\code{\link{tensorish}})}
}
\value{
\code{\link{tensorish}}
}
\description{
Matrix multiplication of two tensors.
}
\section{Shapes}{

\itemize{
\item \code{lhs}: \verb{(b1, ..., bk, m, n)}
\item \code{rhs}: \verb{(b1, ..., bk, n, p)}
\item output: \verb{(b1, ..., bk, m, p)}
}
}

