% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_negate}
\alias{nv_negate}
\title{Negation}
\usage{
nv_negate(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Negates a tensor element-wise. You can also use the unary \code{-} operator.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, -2, 3))
  -x
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_negate]{nvl_negate()}} for the underlying primitive.
}
