% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_negate}
\alias{nv_negate}
\title{Negation}
\usage{
nv_negate(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise negation.
}
