% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_not}
\alias{nv_not}
\title{Logical Not}
\usage{
nv_not(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise logical NOT. You can also use the \code{!} operator.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(TRUE, FALSE, TRUE))
  !x
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_not]{nvl_not()}} for the underlying primitive.
}
