% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_popcnt}
\alias{nv_popcnt}
\title{Population Count}
\usage{
nv_popcnt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise population count (number of set bits in integer).
}
