% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_promote_to_common}
\alias{nv_promote_to_common}
\title{Promote Tensors to a Common Dtype}
\usage{
nv_promote_to_common(...)
}
\arguments{
\item{...}{(\code{\link{nv_tensor}})\cr
Tensors to promote.}
}
\value{
(\code{list()} of \code{\link{nv_tensor}})
}
\description{
Promote tensors to a common type.
}
