% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-rng.R
\name{nv_rbinom}
\alias{nv_rbinom}
\title{Sample from a Binomial Distribution}
\usage{
nv_rbinom(shape, initial_state, n = 1L, prob = 0.5, dtype = "i32")
}
\arguments{
\item{shape}{(\code{integer()})\cr
Shape.}

\item{initial_state}{([\code{ui64[2]}][tensorish])\cr
RNG state.}

\item{n}{(\code{integer(1)})\cr
Number of trials. Default is 1 (Bernoulli).}

\item{prob}{(\code{numeric(1)})\cr
Probability of success on each trial. Default is 0.5.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}
}
\value{
(\code{list()} of \code{\link{tensorish}})\cr
List of two tensors: the new RNG state and the generated random samples.
}
\description{
Sample from a binomial distribution with $n$ trials and success probability $p$.
}
