% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-rng.R
\name{nv_rbinom}
\alias{nv_rbinom}
\title{Random Binomial Samples}
\usage{
nv_rbinom(initial_state, dtype = "i32", shape)
}
\arguments{
\item{initial_state}{(\code{\link{tensorish}})\cr
Tensor of type \code{ui64[2]}.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{shape}{(\code{integer()})\cr
Shape.}
}
\value{
(\code{list()} of \code{\link{tensorish}})\cr
List of two tensors: the new RNG state and the generated random samples (0 or 1).
}
\description{
Generate random Binomial(1, 0.5) samples (0 or 1) by extracting individual bits
from the random number generator. This is equivalent to Bernoulli(0.5) samples.
}
