% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-rng.R
\name{nv_rdunif}
\alias{nv_rdunif}
\title{Sample from a Discrete Uniform Distribution}
\usage{
nv_rdunif(shape, initial_state, n, dtype = "i32")
}
\arguments{
\item{shape}{(\code{integer()})\cr
Shape.}

\item{initial_state}{([\code{ui64[2]}][tensorish])\cr
RNG state.}

\item{n}{(\code{integer(1)})\cr
Number of categories to sample from (samples integers 1 to n).}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}
}
\value{
(\code{list()} of \code{\link{tensorish}})\cr
List of two tensors: the new RNG state and the sampled integers.
}
\description{
Sample from a discrete distribution, analogous to R's \code{sample()} function.
Samples integers from 1 to n with uniform probability and with replacement.
}
