% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{nv_read}
\alias{nv_read}
\title{Read tensors from a file}
\usage{
nv_read(path, device = NULL)
}
\arguments{
\item{path}{(\code{character(1)})\cr
Path to the safetensors file.}

\item{device}{(\code{NULL} | \code{character(1)} | \code{\link[pjrt:pjrt_device]{PJRTDevice}})\cr
The device on which to place the loaded tensors (\code{"cpu"}, \code{"cuda"}, ...).
Default is to use the CPU.}
}
\value{
Named \code{list} of \code{\link{AnvilTensor}} objects.
}
\description{
Loads tensors from a file in the
\href{https://huggingface.co/docs/safetensors/index}{safetensors} format.
}
\details{
This is a convenience wrapper around \code{\link[=nv_unserialize]{nv_unserialize()}} that opens and
closes a file connection.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded("cpu")) withAutoprint(\{ # examplesIf}
x <- nv_tensor(array(1:6, dim = c(2, 3)))
x
path <- tempfile(fileext = ".safetensors")
nv_save(list(x = x), path)
nv_read(path)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_save]{nv_save()}}, \code{\link[=nv_serialize]{nv_serialize()}}, \code{\link[=nv_unserialize]{nv_unserialize()}}
}
