% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_reduce_prod}
\alias{nv_reduce_prod}
\title{Product Reduction}
\usage{
nv_reduce_prod(operand, dims, drop = TRUE)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dims}{(\code{integer()})\cr
Dimensions to reduce.}

\item{drop}{(\code{logical(1)})\cr
Whether to drop reduced dimensions.}
}
\value{
\code{\link{tensorish}}\cr
Has the same data type as the input.
When \code{drop = TRUE}, the reduced dimensions are removed.
When \code{drop = FALSE}, the reduced dimensions are set to 1.
}
\description{
Multiplies tensor elements along the specified dimensions.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(matrix(1:6, nrow = 2))
  nv_reduce_prod(x, dims = 1L)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_reduce_prod]{nvl_reduce_prod()}} for the underlying primitive.
}
