% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_remainder}
\alias{nv_remainder}
\title{Remainder (Truncating)}
\usage{
nv_remainder(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{arrayish}})\cr
Left and right operand.
Operands are \link[=nv_promote_to_common]{promoted to a common data type}.
Scalars are \link[=nv_broadcast_scalars]{broadcast} to the shape of the other operand.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and the promoted common data type of the inputs.
}
\description{
Element-wise remainder.
This differs from base R's \code{\%\%}, use \code{\link[=nv_mod]{nv_mod()}}/\code{\%\%} instead.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(7, 8, 9))
y <- nv_array(c(3, 3, 4))
nv_remainder(x, y)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_mod]{nv_mod()}} for the flooring remainder, \code{\link[=prim_remainder]{prim_remainder()}} for the
underlying primitive.
}
