% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_remainder}
\alias{nv_remainder}
\title{Remainder}
\usage{
nv_remainder(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise remainder of division.
}
