% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_reshape}
\alias{nv_reshape}
\title{Reshape}
\usage{
nv_reshape(operand, shape)
}
\arguments{
\item{operand}{(\code{\link{nv_tensor}})\cr
The tensor.}

\item{shape}{(\code{integer()})\cr
The new shape.}
}
\value{
\code{\link{nv_tensor}}
}
\description{
Reshape a tensor.
}
