% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_reshape}
\alias{nv_reshape}
\title{Reshape}
\usage{
nv_reshape(operand, shape)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{shape}{(\code{integer()})\cr
The new shape.}
}
\value{
\code{\link{tensorish}}
}
\description{
Reshape a tensor.
Note that row-major order is used, which differs from R's column-major order.
}
