% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_reverse}
\alias{nv_reverse}
\title{Reverse}
\usage{
nv_reverse(operand, dims)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dimensions}{(\code{integer()})\cr
Dimensions to reverse (1-indexed).}
}
\value{
\code{\link{tensorish}}
}
\description{
Reverses the order of elements along specified dimensions.
}
