% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nv_rng_state}
\alias{nv_rng_state}
\title{Generate random state}
\usage{
nv_rng_state(seed)
}
\arguments{
\item{seed}{(\code{integer(1)})\cr
Seed value}
}
\value{
\code{\link{nv_tensor}} of dtype \code{ui64} and shape (2)
}
\description{
lightweight function to generate an initial state
}
