% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_rnorm}
\alias{nv_rnorm}
\title{Random Normal Numbers}
\usage{
nv_rnorm(initial_state, shape_out, mu = 0, sigma = 1)
}
\arguments{
\item{initial_state}{state seed}

\item{shape_out}{output shape}

\item{mu}{scalar: expected value}

\item{sigma}{scalar: standard deviation
#' @section Covariance:
To implement a covariance structure use cholesky decomposition}
}
\description{
generate random normal numbers
}
