% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_rnorm}
\alias{nv_rnorm}
\title{Random Normal Numbers}
\usage{
nv_rnorm(initial_state, dtype, shape, mu = 0, sigma = 1)
}
\arguments{
\item{initial_state}{(\code{\link{tensorish}})\cr
Tensor of type \code{ui64[2]}.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{shape}{(\code{integer()})\cr
Shape.}

\item{mu}{(\code{numeric(1)})\cr
Expected value.}

\item{sigma}{(\code{numeric(1)})\cr
Standard deviation.
#' @section Covariance:
To implement a covariance structure use cholesky decomposition.}
}
\value{
(\code{list()} of \code{\link{tensorish}})\cr
List of two tensors: the new RNG state and the generated random numbers.
}
\description{
generate random normal numbers
}
