% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-rng.R
\name{nv_rnorm}
\alias{nv_rnorm}
\title{Sample from a Normal Distribution}
\usage{
nv_rnorm(shape, initial_state, dtype = "f32", mu = 0, sigma = 1)
}
\arguments{
\item{shape}{(\code{integer()})\cr
Shape.}

\item{initial_state}{([\code{ui64[2]}][tensorish])\cr
RNG state.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{mu}{(\code{numeric(1)})\cr
Expected value.}

\item{sigma}{(\code{numeric(1)})\cr
Standard deviation.}
}
\value{
(\code{list()} of \code{\link{tensorish}})\cr
List of two tensors: the new RNG state and the generated random numbers.
}
\description{
Sample from a normal distribution with mean \eqn{\mu} and standard deviation \eqn{\sigma}.
}
\section{Covariance}{

To implement a covariance structure use cholesky decomposition.
}

