% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_round}
\alias{nv_round}
\title{Round}
\usage{
nv_round(operand, method = "nearest_even")
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{method}{(\code{character(1)})\cr
Method to use for rounding.
Either \code{"nearest_even"} (default) or \code{"afz"} (away from zero).}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise rounding.
}
