% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_rsqrt}
\alias{nv_rsqrt}
\title{Reciprocal Square Root}
\usage{
nv_rsqrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise reciprocal square root.
}
