% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-rng.R
\name{nv_unif_rand}
\alias{nv_unif_rand}
\alias{nv_runif}
\title{Sample from a Uniform Distribution}
\usage{
nv_unif_rand(shape, initial_state, dtype = "f64")

nv_runif(shape, initial_state, dtype = "f32", lower = 0, upper = 1)
}
\arguments{
\item{shape}{(\code{integer()})\cr
Shape.}

\item{initial_state}{(\code{\link{tensorish}})\cr
RNG state (\code{ui64[2]}).}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{lower, upper}{(\code{numeric(1)})\cr
Lower and upper bound.}
}
\value{
(\code{list()} of \code{\link{tensorish}})\cr
List of two tensors: the new RNG state and the generated random numbers.
}
\description{
Sample from a uniform distribution in the open interval (lower, upper).
Note that \code{nv_rand_unif()} generates values in [0, 1), while \code{nv_runif()} generates values in (lower, upper).
}
