% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_runif}
\alias{nv_runif}
\title{Random Uniform Numbers}
\usage{
nv_runif(initial_state, shape_out, lower = 0, upper = 1)
}
\arguments{
\item{initial_state}{state seed}

\item{shape_out}{output shape}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\description{
generate random uniform numbers
}
