% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-rng.R
\name{nv_runif}
\alias{nv_runif}
\alias{nv_unif_rand}
\title{Internal: Random Unit Uniform Numbers}
\usage{
nv_unif_rand(shape, initial_state, dtype = "f64")

nv_runif(shape, initial_state, dtype = "f32", lower = 0, upper = 1)
}
\arguments{
\item{shape}{(\code{integer()})\cr
Shape.}

\item{initial_state}{([\code{ui64[2]}][tensorish])\cr
RNG state.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{lower, upper}{(\code{numeric(1)})\cr
Lower and upper bound.}
}
\value{
(\code{list()} of \code{\link{tensorish}})\cr
List of two tensors: the new RNG state and the generated random numbers.
}
\description{
generate random uniform numbers in [0, 1)

Sample from a uniform distribution in the open interval (lower, upper).
}
