% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/api-internal.R
\name{nv_runif}
\alias{nv_runif}
\alias{nv_unif_rand}
\title{Random Uniform Numbers}
\usage{
nv_runif(initial_state, dtype = "f64", shape_out, lower = 0, upper = 1)

nv_unif_rand(initial_state, dtype = "f64", shape_out)
}
\arguments{
\item{initial_state}{state seed}

\item{dtype}{output dtype either "f32" or "f64"}

\item{shape_out}{output shape}

\item{lower}{lower bound}

\item{upper}{upper bound}
}
\description{
generate random uniform numbers in ]lower, upper[

generate random uniform numbers in [0, 1)
}
