% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_runif}
\alias{nv_runif}
\title{Random Uniform Numbers}
\usage{
nv_runif(
  initial_state,
  dtype = "f64",
  shape_out,
  lower = 0,
  upper = 1,
  include_lower = FALSE
)
}
\arguments{
\item{initial_state}{state seed}

\item{dtype}{output dtype either "f32" or "f64"}

\item{shape_out}{output shape}

\item{lower}{lower bound}

\item{upper}{upper bound}

\item{include_upper}{Flag, whether to sample from upper bound}
}
\description{
generate random uniform numbers
}
