% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_select}
\alias{nv_select}
\title{Select}
\usage{
nv_select(pred, true_value, false_value)
}
\arguments{
\item{pred}{condition}

\item{true_value}{on true}

\item{false_value}{on false}
}
\value{
\code{\link{tensorish}}
}
\description{
return values from true_value and false_value conditioned on pred
}
