% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_seq}
\alias{nv_seq}
\title{Sequence}
\usage{
nv_seq(start, end, dtype = "i32", ambiguous = FALSE)
}
\arguments{
\item{start, end}{(\code{integer(1)})\cr
Start and end values.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the type is ambiguous.
Ambiguous types usually arise from R literals (e.g., \code{1L}, \code{1.0}) and follow special promotion rules.
See the vignette "Type Promotion" for more details.}
}
\value{
\code{\link{tensorish}}
}
\description{
Creates a tensor with values increasing from start to end.
}
