% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_seq}
\alias{nv_seq}
\title{Sequence}
\usage{
nv_seq(start, end, dtype = "i32", ambiguous = FALSE)
}
\arguments{
\item{start, end}{(\code{integer(1)})\cr
Start and end values. Must satisfy \code{start <= end}.}

\item{dtype}{(\code{character(1)} | \code{\link[tengen:TensorDataType]{tengen::TensorDataType}})\cr
Data type.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the type is ambiguous.
Ambiguous types usually arise from R literals (e.g., \code{1L}, \code{1.0}) and follow special promotion rules.
See the \code{vignette("type-promotion")} for more details.}
}
\value{
\code{\link{tensorish}}\cr
1-D tensor of length \code{end - start + 1}.
}
\description{
Creates a 1-D tensor with integer values from \code{start} to \code{end} (inclusive),
analogous to R's \code{seq(start, end)}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval(nv_seq(3, 7))
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_iota]{nv_iota()}} for multi-dimensional sequences.
}
