% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_shift_right_arithmetic}
\alias{nv_shift_right_arithmetic}
\title{Arithmetic Shift Right}
\usage{
nv_shift_right_arithmetic(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise bitwise arithmetic right shift.
}
