% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_sign}
\alias{nv_sign}
\title{Sign}
\usage{
nv_sign(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise sign function. You can also use \code{sign()}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(-3, 0, 5))
  sign(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_sign]{nvl_sign()}} for the underlying primitive.
}
