% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_sign}
\alias{nv_sign}
\title{Sign}
\usage{
nv_sign(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise sign function.
}
