% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_sqrt}
\alias{nv_sqrt}
\title{Square Root}
\usage{
nv_sqrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise square root.
}
