% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_sqrt}
\alias{nv_sqrt}
\title{Square Root}
\usage{
nv_sqrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise square root. You can also use \code{sqrt()}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 4, 9))
  sqrt(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_sqrt]{nvl_sqrt()}} for the underlying primitive.
}
