% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_static_slice}
\alias{nv_static_slice}
\title{Slice}
\usage{
nv_static_slice(operand, start_indices, limit_indices, strides)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{start_indices}{start of slice}

\item{limit_indices}{end of slice}

\item{strides}{stride size}
}
\value{
\code{\link{tensorish}}
}
\description{
return slice of operand.
}
