% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_sub}
\alias{nv_sub}
\title{Subtraction}
\usage{
nv_sub(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise subtraction of two tensors.
}
