% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_sub}
\alias{nv_sub}
\title{Subtraction}
\usage{
nv_sub(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.
Operands are \link[=nv_promote_to_common]{promoted to a common data type}.
Scalars are \link[=nv_broadcast_scalars]{broadcast} to the shape of the other operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and the promoted common data type of the inputs.
}
\description{
Subtracts two tensors element-wise. You can also use the \code{-} operator.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(4, 5, 6))
  y <- nv_tensor(c(1, 2, 3))
  x - y
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_sub]{nvl_sub()}} for the underlying primitive.
}
