% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-generics.R, R/api-subset.R
\name{[.AnvilBox}
\alias{[.AnvilBox}
\alias{[.AnvilTensor}
\alias{nv_subset}
\title{Subset a Tensor}
\usage{
\method{[}{AnvilBox}(x, ...)

\method{[}{AnvilTensor}(x, ...)

nv_subset(x, ...)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
Input tensor to subset.}

\item{...}{Subset specifications.}
}
\value{
\code{\link{tensorish}}
}
\description{
Extracts a subset from a tensor.
See vignette("subsetting") for more details.
}
