% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-generics.R, R/api-subset.R
\name{[<-.AnvilBox}
\alias{[<-.AnvilBox}
\alias{[<-.AnvilTensor}
\alias{nv_subset_assign}
\title{Update Subset}
\usage{
\method{[}{AnvilBox}(x, ...) <- value

\method{[}{AnvilTensor}(x, ...) <- value

nv_subset_assign(x, ..., value)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
Input tensor to update.}

\item{...}{Subset specifications.}

\item{value}{(\code{\link{tensorish}})\cr
Values to write. Scalars are broadcast to the subset shape.
Non-scalar values must have a shape matching the subset shape.}
}
\value{
\code{\link{tensorish}} A new tensor with the subset updated.
}
\description{
Updates elements of a tensor at specified positions.
See vignette("subsetting") for more details.
}
\seealso{
\code{\link[=nv_subset]{nv_subset()}}
}
