% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_tan}
\alias{nv_tan}
\title{Tangent}
\usage{
nv_tan(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
}
\description{
Element-wise tangent. You can also use \code{tan()}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(0, 0.5, 1))
  tan(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_tan]{nvl_tan()}} for the underlying primitive.
}
