% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_tanh}
\alias{nv_tanh}
\title{Hyperbolic Tangent}
\usage{
nv_tanh(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise hyperbolic tangent.
}
