% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/api-generics.R
\name{nv_tcrossprod}
\alias{nv_tcrossprod}
\alias{tcrossprod.AnvlBox}
\title{Transpose Cross Product (Matrix)}
\usage{
nv_tcrossprod(x, y = NULL)

\method{tcrossprod}{AnvlBox}(x, y = NULL, ...)
}
\arguments{
\item{x}{(\code{\link{arrayish}})\cr
An array with at least 2 dimensions.}

\item{y}{(\code{\link{arrayish}} | \code{NULL})\cr
Optional second array. If \code{NULL}, uses \code{x}.}

\item{...}{Unused.}
}
\value{
\code{\link{arrayish}}
}
\description{
Computes \code{x \%*\% t(y)}. If \code{y} is missing, computes \code{x \%*\% t(x)}.
}
\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(matrix(1:6, nrow = 2), dtype = "f32")
nv_tcrossprod(x)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_crossprod]{nv_crossprod()}}, \code{\link[=nv_matmul]{nv_matmul()}}
}
