% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/api-generics.R
\name{nv_tcrossprod}
\alias{nv_tcrossprod}
\alias{tcrossprod.AnvilBox}
\title{Transpose Cross Product (Matrix)}
\usage{
nv_tcrossprod(x, y = NULL)

\method{tcrossprod}{AnvilBox}(x, y = NULL, ...)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
A tensor with at least 2 dimensions.}

\item{y}{(\code{\link{tensorish}} | \code{NULL})\cr
Optional second tensor. If \code{NULL}, uses \code{x}.}
}
\value{
\code{\link{tensorish}}
}
\description{
Computes \code{x \%*\% t(y)}. If \code{y} is missing, computes \code{x \%*\% t(x)}.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(matrix(1:6, nrow = 2), dtype = "f32")
  nv_tcrossprod(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_crossprod]{nv_crossprod()}}, \code{\link[=nv_matmul]{nv_matmul()}}
}
