% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{nv_tensor}
\alias{nv_tensor}
\alias{nv_scalar}
\alias{nv_empty}
\title{Tensor}
\usage{
nv_tensor(data, dtype = NULL, platform = NULL, shape = NULL)

nv_scalar(data, dtype = NULL, platform = NULL)

nv_empty(dtype, shape, platform = NULL)
}
\arguments{
\item{data}{(any)\cr
Object convertible to a \code{\link[pjrt:pjrt_buffer]{PJRTBuffer}}.}

\item{dtype}{(\code{NULL} | \code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
One of pred, i8, i16, i32, i64, ui8, ui16, ui32, ui64, f32, f64 or a \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}}.
The default (\code{NULL}) uses \code{f32} for numeric data, \code{i32} for integer data, and \code{pred} for logical data.}

\item{platform}{(\code{character(1)})\cr
The platform name for the tensor (\code{"cpu"}, \code{"cuda"}, \code{"metal"}).
Default is to use the CPU, unless the data is already a \code{\link[pjrt:pjrt_buffer]{PJRTBuffer}}.
You can change the default by setting the \code{PJRT_PLATFORM} environment variable.}

\item{shape}{(\code{NULL} | \code{integer()})\cr
Shape.
The default (\code{NULL}) is to infer it from the data if possible.
Note that \code{\link{nv_tensor}} interpretes length 1 vectors as having shape \code{(1)}.
To create a "scalar" with dimension \verb{()}, use \code{\link{nv_scalar}}.}
}
\value{
(\code{AnvilTensor})
}
\description{
Create a tensor.
}
\details{
Internally calls \code{\link[pjrt:pjrt_buffer]{pjrt_buffer}}.
}
