% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_trace}
\alias{nv_trace}
\title{Matrix Trace}
\usage{
nv_trace(x)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
A 2-D tensor (matrix).}
}
\value{
\code{\link{tensorish}}\cr
A scalar tensor with the same data type as \code{x}.
}
\description{
Computes the trace (sum of diagonal elements) of a 2-D tensor.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 0, 0, 0, 2, 0, 0, 0, 3), shape = c(3, 3))
  nv_trace(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_extract_diag]{nv_extract_diag()}}, \code{\link[=nv_diag]{nv_diag()}}
}
