% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/api-generics.R
\name{nv_transpose}
\alias{nv_transpose}
\alias{t.AnvilBox}
\title{Transpose}
\usage{
nv_transpose(x, permutation = NULL)

\method{t}{AnvilBox}(x)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
Tensor to transpose.}

\item{permutation}{(\code{integer()} | \code{NULL})\cr
New ordering of dimensions. If \code{NULL} (default), reverses the dimensions.}
}
\value{
\code{\link{tensorish}}\cr
Has the same data type as \code{x} and shape \code{nv_shape(x)[permutation]}.
}
\description{
Permutes the dimensions of a tensor. You can also use \code{t()} for matrices.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(matrix(1:6, nrow = 2))
  t(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_transpose]{nvl_transpose()}} for the underlying primitive.
}
