% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/api-group-generics.R
\name{nv_transpose}
\alias{nv_transpose}
\alias{t.anvil::Box}
\title{Transpose}
\usage{
nv_transpose(x, permutation = NULL)

\method{t}{`anvil::Box`}(x)
}
\arguments{
\item{x}{(\code{\link{nv_tensor}})}

\item{permutation}{(\code{integer()} | \code{NULl})\cr
Permutation of dimensions. If \code{NULL} (default), reverses the dimensions.}
}
\value{
\code{\link{nv_tensor}}
}
\description{
Transpose a tensor.
}
