% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_tril}
\alias{nv_tril}
\title{Lower Triangular Matrix}
\usage{
nv_tril(x, diagonal = 0L)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
A 2-D tensor (matrix).}

\item{diagonal}{(\code{integer(1)})\cr
Diagonal offset. \code{0} (default) is the main diagonal, positive values
include diagonals above, negative values exclude diagonals below.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as \code{x}.
}
\description{
Returns the lower triangular part of a 2-D tensor, setting elements above
the specified diagonal to zero.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_fill(1, c(3, 3))
  nv_tril(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_triu]{nv_triu()}}
}
