% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_unary_ops}
\alias{nv_unary_ops}
\alias{nv_neg}
\alias{nv_abs}
\alias{nv_sqrt}
\alias{nv_rsqrt}
\alias{nv_log}
\alias{nv_tanh}
\alias{nv_tan}
\alias{nv_floor}
\alias{nv_ceil}
\alias{nv_sign}
\alias{nv_exp}
\alias{nv_round}
\title{Unary Operations}
\usage{
nv_neg(operand)

nv_abs(operand)

nv_sqrt(operand)

nv_rsqrt(operand)

nv_log(operand)

nv_tanh(operand)

nv_tan(operand)

nv_floor(operand)

nv_ceil(operand)

nv_sign(operand)

nv_exp(operand)

nv_round(operand, method = "nearest_even")
}
\arguments{
\item{operand}{(\code{\link{nv_tensor}})}
}
\value{
\code{\link{nv_tensor}}
}
\description{
Unary operations on tensors.
}
