% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_while}
\alias{nv_while}
\title{While}
\usage{
nv_while(init, cond, body)
}
\arguments{
\item{init}{(\code{list()})\cr
Initial state.}

\item{cond}{(\code{function})\cr
Condition function: \code{f: state -> bool}.}

\item{body}{(\code{function})\cr
Body function. \code{f: state -> state}.}
}
\value{
\code{\link{nv_tensor}}
}
\description{
Functional while loop.
}
