% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_xor}
\alias{nv_xor}
\title{Logical Xor}
\usage{
nv_xor(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.
Operands are \link[=nv_promote_to_common]{promoted to a common data type}.
Scalars are \link[=nv_broadcast_scalars]{broadcast} to the shape of the other operand.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and the promoted common data type of the inputs.
}
\description{
Element-wise logical XOR.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(TRUE, FALSE, TRUE))
  y <- nv_tensor(c(TRUE, TRUE, FALSE))
  nv_xor(x, y)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nvl_xor]{nvl_xor()}} for the underlying primitive.
}
