% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{nv_xor}
\alias{nv_xor}
\title{Logical Xor}
\usage{
nv_xor(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logical XOR operation.
}
