% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_abs}
\alias{nvl_abs}
\title{Primitive Absolute Value}
\usage{
nvl_abs(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise absolute value.
}
