% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_abs}
\alias{nvl_abs}
\title{Primitive Absolute Value}
\usage{
nvl_abs(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of signed integer or floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise absolute value.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_abs]{stablehlo::hlo_abs()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(-1, 2, -3))
  nvl_abs(x)
})
\dontshow{\}) # examplesIf}
}
