% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_add}
\alias{nvl_add}
\title{Primitive Addition}
\usage{
nvl_add(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Adds two tensors element-wise.
}
