% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_and}
\alias{nvl_and}
\title{Primitive And}
\usage{
nvl_and(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logical AND.
}
