% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_atan2}
\alias{nvl_atan2}
\title{Primitive Atan2}
\usage{
nvl_atan2(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}} of floating-point type)\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise atan2 operation.
}
\section{Shapes}{

\code{lhs} and \code{rhs} must have the same shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_atan2]{stablehlo::hlo_atan2()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  y <- nv_tensor(c(1, 0, -1))
  x <- nv_tensor(c(0, 1, 0))
  nvl_atan2(y, x)
})
\dontshow{\}) # examplesIf}
}
