% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_atan2}
\alias{nvl_atan2}
\title{Primitive Atan2}
\usage{
nvl_atan2(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise arc tangent of y/x.
}
