% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_bitcast_convert}
\alias{nvl_bitcast_convert}
\title{Primitive Bitcast Convert}
\usage{
nvl_bitcast_convert(operand, dtype)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}
}
\value{
\code{\link{tensorish}}
}
\description{
Reinterprets tensor bits as a different dtype.
}
\section{Shapes}{

If the source and target types have the same bit width, the output
has the same shape as \code{operand}. Otherwise the last dimension is
adjusted based on the bit-width ratio.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_bitcast_convert]{stablehlo::hlo_bitcast_convert()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(1L)
  nvl_bitcast_convert(x, dtype = "f32")
})
\dontshow{\}) # examplesIf}
}
