% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_bitcast_convert}
\alias{nvl_bitcast_convert}
\title{Primitive Bitcast Convert}
\usage{
nvl_bitcast_convert(operand, dtype)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}
}
\value{
\code{\link{tensorish}}
}
\description{
Reinterprets tensor bits as a different dtype.
}
