% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_broadcast_in_dim}
\alias{nvl_broadcast_in_dim}
\title{Primitive Broadcast}
\usage{
nvl_broadcast_in_dim(operand, shape_out, broadcast_dimensions)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{shape_out}{(\code{integer()})\cr
Target shape.}

\item{broadcast_dimensions}{(\code{integer()})\cr
Dimension mapping.}
}
\value{
\code{\link{tensorish}}
}
\description{
Broadcasts a tensor to a new shape.
}
