% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_cbrt}
\alias{nvl_cbrt}
\title{Primitive Cube Root}
\usage{
nvl_cbrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise cube root.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_cbrt]{stablehlo::hlo_cbrt()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 8, 27))
  nvl_cbrt(x)
})
\dontshow{\}) # examplesIf}
}
