% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_cbrt}
\alias{nvl_cbrt}
\title{Primitive Cube Root}
\usage{
nvl_cbrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of data type floating-point.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise cube root.
Is the same as \code{\link[=nv_cbrt]{nv_cbrt()}}.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_cbrt]{stablehlo::hlo_cbrt()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 8, 27))
  nvl_cbrt(x)
})
\dontshow{\}) # examplesIf}
}
