% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_ceil}
\alias{nvl_ceil}
\title{Primitive Ceiling}
\usage{
nvl_ceil(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise ceiling.
}
