% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_ceil}
\alias{nvl_ceil}
\title{Primitive Ceiling}
\usage{
nvl_ceil(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise ceiling.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_ceil]{stablehlo::hlo_ceil()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1.2, 2.7, -1.5))
  nvl_ceil(x)
})
\dontshow{\}) # examplesIf}
}
