% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_clamp}
\alias{nvl_clamp}
\title{Primitive Clamp}
\usage{
nvl_clamp(min_val, operand, max_val)
}
\arguments{
\item{min_val}{(\code{\link{tensorish}})\cr
Minimum value. Must be scalar or the same shape as \code{operand}.}

\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of any data type.}

\item{max_val}{(\code{\link{tensorish}})\cr
Maximum value. Must be scalar or the same shape as \code{operand}.}
}
\value{
\code{\link{tensorish}}\cr
Has the same data type and shape as \code{operand}.
It is ambiguous if the input is ambiguous.
}
\description{
Clamps every element of \code{operand} to the range \verb{[min_val, max_val]},
i.e. \code{max(min_val, min(operand, max_val))}.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_clamp]{stablehlo::hlo_clamp()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(-1, 0.5, 2))
  nvl_clamp(nv_scalar(0), x, nv_scalar(1))
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_clamp]{nv_clamp()}}
}
