% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_concatenate}
\alias{nvl_concatenate}
\title{Primitive Concatenate}
\usage{
nvl_concatenate(..., dimension)
}
\arguments{
\item{...}{(\code{\link{tensorish}})\cr
Tensors to concatenate.}

\item{dimension}{(\code{integer(1)})\cr
Dimension to concatenate along.}
}
\value{
\code{\link{tensorish}}
}
\description{
Concatenates tensors along a dimension.
}
