% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_concatenate}
\alias{nvl_concatenate}
\title{Primitive Concatenate}
\usage{
nvl_concatenate(..., dimension)
}
\arguments{
\item{...}{(\code{\link{tensorish}})\cr
Tensors to concatenate.}

\item{dimension}{(\code{integer(1)})\cr
Dimension to concatenate along.}
}
\value{
\code{\link{tensorish}}
}
\description{
Concatenates tensors along a dimension.
}
\section{Shapes}{

All inputs must have the same shape except along \code{dimension}. The output dimension size is the sum of the input dimension sizes along \code{dimension}.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_concatenate]{stablehlo::hlo_concatenate()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  y <- nv_tensor(c(4, 5, 6))
  nvl_concatenate(x, y, dimension = 1L)
})
}
}
