% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_convert}
\alias{nvl_convert}
\title{Primitive Convert}
\usage{
nvl_convert(operand, dtype, ambiguous = FALSE)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the result type is ambiguous.}
}
\value{
\code{\link{tensorish}}
}
\description{
Converts tensor to a different dtype.
}
