% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_convert}
\alias{nvl_convert}
\title{Primitive Convert}
\usage{
nvl_convert(operand, dtype, ambiguous = FALSE)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the type is ambiguous.
Ambiguous types usually arise from R literals (e.g., \code{1L}, \code{1.0}) and follow special promotion rules.
See the vignette "Type Promotion" for more details.}
}
\value{
\code{\link{tensorish}}
}
\description{
Converts tensor to a different dtype.
}
\section{Shapes}{

Output has the same shape as \code{operand}.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_convert]{stablehlo::hlo_convert()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(1L, 2L, 3L))
  nvl_convert(x, dtype = "f32")
})
}
}
