% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_cosine}
\alias{nvl_cosine}
\title{Primitive Cosine}
\usage{
nvl_cosine(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of data type floating-point.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise cosine.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_cosine]{stablehlo::hlo_cosine()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(0, pi / 2, pi))
  nvl_cosine(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_cosine]{nv_cosine()}}, \code{\link[=cos]{cos()}}
}
