% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_div}
\alias{nvl_div}
\title{Primitive Division}
\usage{
nvl_div(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Divides two tensors element-wise.
}
