% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_div}
\alias{nvl_div}
\title{Primitive Division}
\usage{
nvl_div(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}} of integer or floating-point type)\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Divides two tensors element-wise.
}
\section{Shapes}{

\code{lhs} and \code{rhs} must have the same shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_divide]{stablehlo::hlo_divide()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(10, 20, 30))
  y <- nv_tensor(c(2, 5, 10))
  nvl_div(x, y)
})
}
}
