% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_dot_general}
\alias{nvl_dot_general}
\title{Primitive Dot General}
\usage{
nvl_dot_general(lhs, rhs, contracting_dims, batching_dims)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}

\item{contracting_dims}{(\code{list()})\cr
Dimensions to contract.}

\item{batching_dims}{(\code{list()})\cr
Batch dimensions.}
}
\value{
\code{\link{tensorish}}
}
\description{
General dot product of two tensors.
}
