% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_dot_general}
\alias{nvl_dot_general}
\title{Primitive Dot General}
\usage{
nvl_dot_general(lhs, rhs, contracting_dims, batching_dims)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}

\item{contracting_dims}{(\code{list()})\cr
Dimensions to contract.}

\item{batching_dims}{(\code{list()})\cr
Batch dimensions.}
}
\value{
\code{\link{tensorish}}
}
\description{
General dot product of two tensors.
}
\section{Shapes}{

Contracting dimensions in \code{lhs} and \code{rhs} must have matching sizes. Batching dimensions must also have matching sizes. The output shape is the batching dimensions followed by the remaining (non-contracted, non-batched) dimensions of \code{lhs}, then \code{rhs}.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_dot_general]{stablehlo::hlo_dot_general()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(matrix(1:6, nrow = 2))
  y <- nv_tensor(matrix(1:6, nrow = 3))
  nvl_dot_general(x, y,
    contracting_dims = list(2L, 1L),
    batching_dims = list(integer(0), integer(0))
  )
})
\dontshow{\}) # examplesIf}
}
