% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_dynamic_slice}
\alias{nvl_dynamic_slice}
\title{Primitive Dynamic Slice}
\usage{
nvl_dynamic_slice(operand, ..., slice_sizes)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{...}{(\code{\link{tensorish}})\cr
Start indices as scalar tensors (1-based).}

\item{slice_sizes}{(\code{integer()})\cr
Size of the slice in each dimension.}
}
\value{
\code{\link{tensorish}}
}
\description{
Extracts a dynamically positioned slice from a tensor.
}
