% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_dynamic_update_slice}
\alias{nvl_dynamic_update_slice}
\title{Primitive Dynamic Update Slice}
\usage{
nvl_dynamic_update_slice(operand, update, ...)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{update}{(\code{\link{tensorish}})\cr
Update tensor with the new values.}

\item{...}{(\code{\link{tensorish}})\cr
Start indices as scalar tensors (1-based).}
}
\value{
\code{\link{tensorish}}
}
\description{
Updates a dynamically positioned slice in a tensor.
}
