% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_exp}
\alias{nvl_exp}
\title{Primitive Exponential}
\usage{
nvl_exp(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise exponential.
}
