% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_exp}
\alias{nvl_exp}
\title{Primitive Exponential}
\usage{
nvl_exp(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise exponential.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_exponential]{stablehlo::hlo_exponential()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(0, 1, 2))
  nvl_exp(x)
})
}
}
