% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_exp}
\alias{nvl_exp}
\title{Primitive Exponential}
\usage{
nvl_exp(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of data type floating-point.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise exponential.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_exponential]{stablehlo::hlo_exponential()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(0, 1, 2))
  nvl_exp(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_exp]{nv_exp()}}, \code{\link[=exp]{exp()}}
}
