% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_expm1}
\alias{nvl_expm1}
\title{Primitive Exponential Minus One}
\usage{
nvl_expm1(operand)
}
\arguments{
\item{operand}{(\code{\link{arrayish}})\cr
Arrayish value of data type floating-point.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise exp(x) - 1, more accurate for small x.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{reverse}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_exponential_minus_one]{stablehlo::hlo_exponential_minus_one()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_array(c(0, 0.001, 1))
  nvl_expm1(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_expm1]{nv_expm1()}}
}
