% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_expm1}
\alias{nvl_expm1}
\title{Primitive Exponential Minus One}
\usage{
nvl_expm1(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise exp(x) - 1, more accurate for small x.
}
