% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_expm1}
\alias{nvl_expm1}
\title{Primitive Exponential Minus One}
\usage{
nvl_expm1(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise exp(x) - 1, more accurate for small x.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_exponential_minus_one]{stablehlo::hlo_exponential_minus_one()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(0, 0.001, 1))
  nvl_expm1(x)
})
}
}
