% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_fill}
\alias{nvl_fill}
\title{Primitive Fill}
\usage{
nvl_fill(value, shape, dtype)
}
\arguments{
\item{value}{(\code{numeric(1)})\cr
Scalar value.}

\item{shape}{(\code{integer()})\cr
Shape.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}
}
\value{
\code{\link{tensorish}}
}
\description{
Creates a tensor filled with a scalar value.
}
