% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_fill}
\alias{nvl_fill}
\title{Primitive Fill}
\usage{
nvl_fill(value, shape, dtype, ambiguous = FALSE)
}
\arguments{
\item{value}{(\code{numeric(1)})\cr
Scalar value.}

\item{shape}{(\code{integer()})\cr
Shape.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the type is ambiguous.
Ambiguous types usually arise from R literals (e.g., \code{1L}, \code{1.0}) and follow special promotion rules.
See the vignette "Type Promotion" for more details.}
}
\value{
\code{\link{tensorish}}
}
\description{
Creates a tensor filled with a scalar value.
}
\section{Shapes}{

Output shape is \code{shape}.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_constant]{stablehlo::hlo_constant()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval(nvl_fill(3.14, shape = c(2, 3), dtype = "f32"))
\dontshow{\}) # examplesIf}
}
