% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_floor}
\alias{nvl_floor}
\title{Primitive Floor}
\usage{
nvl_floor(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise floor.
}
