% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_floor}
\alias{nvl_floor}
\title{Primitive Floor}
\usage{
nvl_floor(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise floor.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_floor]{stablehlo::hlo_floor()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(1.2, 2.7, -1.5))
  nvl_floor(x)
})
}
}
