% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_gather}
\alias{nvl_gather}
\title{Primitive Gather}
\usage{
nvl_gather(
  operand,
  start_indices,
  slice_sizes,
  offset_dims,
  collapsed_slice_dims,
  operand_batching_dims,
  start_indices_batching_dims,
  start_index_map,
  index_vector_dim,
  indices_are_sorted = FALSE,
  unique_indices = FALSE
)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{start_indices}{(\code{\link{tensorish}})\cr
Tensor of integer type containing the starting indices for the gather operation.}

\item{slice_sizes}{(\code{integer()})\cr
The sizes of the slices to gather in each dimension.}

\item{offset_dims}{(\code{integer()})\cr
Dimensions of the operand to gather from.}

\item{collapsed_slice_dims}{(\code{integer()})\cr
Dimensions of the slice to gather.}

\item{operand_batching_dims}{(\code{integer()})\cr
Dimensions of the operand to gather from.}

\item{start_indices_batching_dims}{(\code{integer()})\cr
Dimensions of the start_indices to gather from.}

\item{start_index_map}{(\code{integer()})\cr
Mapping from the start_indices to the operand dimensions.}

\item{index_vector_dim}{(\code{integer(1)})\cr
Dimension of the index vector.}

\item{indices_are_sorted}{(\code{logical(1)})\cr
Whether indices are guaranteed to be sorted.}

\item{unique_indices}{(\code{logical(1)})\cr
Whether indices are guaranteed to be unique (no duplicates).}
}
\value{
\code{\link{tensorish}}
}
\description{
Gathers slices from the operand at positions specified by start_indices.
}
