% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_gather}
\alias{nvl_gather}
\title{Primitive Gather}
\usage{
nvl_gather(
  operand,
  start_indices,
  gather_dimension_numbers,
  slice_sizes,
  indices_are_sorted = FALSE
)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{start_indices}{(\code{\link{tensorish}})\cr
Tensor of integer type containing the starting indices for the gather operation.}

\item{gather_dimension_numbers}{(\code{\link{GatherDimensionNumbers}})\cr
Dimension configuration for the gather operation.}

\item{slice_sizes}{(\code{integer()})\cr
The sizes of the slices to gather in each dimension.}

\item{indices_are_sorted}{(\code{logical(1)})\cr
Whether indices are guaranteed to be sorted.}
}
\value{
\code{\link{tensorish}}
}
\description{
Gathers slices from the operand at positions specified by start_indices.
}
