% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_ge}
\alias{nvl_ge}
\title{Primitive Greater Equal}
\usage{
nvl_ge(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}} (boolean)
}
\description{
Element-wise greater than or equal comparison.
}
\section{Shapes}{

\code{lhs} and \code{rhs} must have the same shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_compare]{stablehlo::hlo_compare()}} with \code{comparison_direction = "GE"}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  y <- nv_tensor(c(3, 2, 1))
  nvl_ge(x, y)
})
}
}
