% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_gt}
\alias{nvl_gt}
\title{Primitive Greater Than}
\usage{
nvl_gt(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}} (boolean)
}
\description{
Element-wise greater than comparison.
}
