% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_if}
\alias{nvl_if}
\title{Primitive If}
\usage{
nvl_if(pred, true, false)
}
\arguments{
\item{pred}{(\code{\link{tensorish}})\cr
Scalar boolean predicate.}

\item{true}{(\code{expression})\cr
Expression for true branch.}

\item{false}{(\code{expression})\cr
Expression for false branch.}
}
\value{
Result of the executed branch.
}
\description{
Conditional execution of branches.
}
