% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_if}
\alias{nvl_if}
\title{Primitive If}
\usage{
nvl_if(pred, true, false)
}
\arguments{
\item{pred}{(\code{\link{tensorish}} of boolean type, scalar)\cr
Predicate.}

\item{true}{(\code{expression})\cr
Expression for true branch.}

\item{false}{(\code{expression})\cr
Expression for false branch.}
}
\value{
Result of the executed branch.
}
\description{
Conditional execution of branches.
}
\section{Shapes}{

\code{pred} must be scalar. Both branches must return outputs with the same shapes.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_if]{stablehlo::hlo_if()}}.
}

\examples{
\dontrun{
jit_eval(nvl_if(nv_scalar(TRUE), nv_scalar(1), nv_scalar(2)))
}
}
