% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_ifelse}
\alias{nvl_ifelse}
\title{Primitive Ifelse}
\usage{
nvl_ifelse(pred, true_value, false_value)
}
\arguments{
\item{pred}{(\code{\link{tensorish}} of boolean type)\cr
Predicate tensor.}

\item{true_value}{(\code{\link{tensorish}})\cr
Value when pred is true.}

\item{false_value}{(\code{\link{tensorish}})\cr
Value when pred is false.}
}
\value{
\code{\link{tensorish}}
}
\description{
Selects elements based on a predicate.
}
\section{Shapes}{

\code{pred} must be either scalar or the same shape as \code{true_value}. \code{true_value} and \code{false_value} must have the same shape. Output has the shape of \code{true_value}.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_select]{stablehlo::hlo_select()}}.
}

\examples{
\dontrun{
jit_eval({
  pred <- nv_tensor(c(TRUE, FALSE, TRUE))
  nvl_ifelse(pred, nv_tensor(c(1, 2, 3)), nv_tensor(c(4, 5, 6)))
})
}
}
