% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_ifelse}
\alias{nvl_ifelse}
\title{Primitive Ifelse}
\usage{
nvl_ifelse(pred, true_value, false_value)
}
\arguments{
\item{pred}{(\code{\link{tensorish}})\cr
Boolean predicate tensor.}

\item{true_value}{(\code{\link{tensorish}})\cr
Value when pred is true.}

\item{false_value}{(\code{\link{tensorish}})\cr
Value when pred is false.}
}
\value{
\code{\link{tensorish}}
}
\description{
Selects elements based on a predicate.
}
