% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_iota}
\alias{nvl_iota}
\title{Primitive Iota}
\usage{
nvl_iota(dim, dtype, shape)

nvl_iota(dim, dtype, shape)
}
\arguments{
\item{dim}{(\code{integer(1)})\cr
Dimension along which values increase (1-indexed).}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{shape}{(\code{integer()})\cr
Shape of the output tensor.}

\item{start}{(\code{numeric(1)})\cr
The value to start the sequence at.}
}
\value{
\code{\link{tensorish}}

\code{\link{tensorish}}
}
\description{
Creates a tensor with increasing values along a dimension.

Creates a tensor with values increasing along the specified dimension.
}
