% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_iota}
\alias{nvl_iota}
\title{Primitive Iota}
\usage{
nvl_iota(dim, shape, dtype, start)
}
\arguments{
\item{dim}{(\code{integer(1)})\cr
The dimension along which to generate increasing values.}

\item{shape}{(\code{integer()})\cr
Shape.}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{start}{(\code{numeric(1)})\cr
The value to start the sequence at.}
}
\value{
\code{\link{tensorish}}
}
\description{
Creates a tensor with increasing values along a dimension.
}
