% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_iota}
\alias{nvl_iota}
\title{Primitive Iota}
\usage{
nvl_iota(dim, dtype, shape, start = 1L, ambiguous = FALSE)
}
\arguments{
\item{dim}{(\code{integer(1)})\cr
Dimension along which values increase (1-indexed).}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{shape}{(\code{integer()})\cr
Shape of the output tensor.}

\item{start}{(\code{integer(1)})\cr
Starting value.}

\item{ambiguous}{(\code{logical(1)})\cr
Whether the type is ambiguous.
Ambiguous types usually arise from R literals (e.g., \code{1L}, \code{1.0}) and follow special promotion rules.
See the \code{vignette("type-promotion")} for more details.}
}
\value{
\code{\link{tensorish}}
}
\description{
Creates a tensor with values increasing along the specified dimension.
}
