% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_is_finite}
\alias{nvl_is_finite}
\title{Primitive Is Finite}
\usage{
nvl_is_finite(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}} of boolean type
}
\description{
Element-wise check if values are finite (not Inf, -Inf, or NaN).
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_is_finite]{stablehlo::hlo_is_finite()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(1, Inf, NaN, -Inf, 0))
  nvl_is_finite(x)
})
}
}
