% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_log}
\alias{nvl_log}
\title{Primitive Logarithm}
\usage{
nvl_log(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise natural logarithm.
}
