% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_log}
\alias{nvl_log}
\title{Primitive Logarithm}
\usage{
nvl_log(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise natural logarithm.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_log]{stablehlo::hlo_log()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2.718, 7.389))
  nvl_log(x)
})
\dontshow{\}) # examplesIf}
}
