% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_log1p}
\alias{nvl_log1p}
\title{Primitive Log Plus One}
\usage{
nvl_log1p(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise log(1 + x), more accurate for small x.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_log_plus_one]{stablehlo::hlo_log_plus_one()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(0, 0.001, 1))
  nvl_log1p(x)
})
}
}
