% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_log1p}
\alias{nvl_log1p}
\title{Primitive Log Plus One}
\usage{
nvl_log1p(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise log(1 + x), more accurate for small x.
}
