% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_logistic}
\alias{nvl_logistic}
\title{Primitive Logistic (Sigmoid)}
\usage{
nvl_logistic(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logistic sigmoid: 1 / (1 + exp(-x)).
}
