% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_logistic}
\alias{nvl_logistic}
\title{Primitive Logistic (Sigmoid)}
\usage{
nvl_logistic(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logistic sigmoid: 1 / (1 + exp(-x)).
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_logistic]{stablehlo::hlo_logistic()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(-2, 0, 2))
  nvl_logistic(x)
})
}
}
