% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_max}
\alias{nvl_max}
\title{Primitive Maximum}
\usage{
nvl_max(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise maximum of two tensors.
}
