% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_min}
\alias{nvl_min}
\title{Primitive Minimum}
\usage{
nvl_min(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise minimum of two tensors.
}
\section{Shapes}{

\code{lhs} and \code{rhs} must have the same shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_minimum]{stablehlo::hlo_minimum()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(1, 5, 3))
  y <- nv_tensor(c(4, 2, 6))
  nvl_min(x, y)
})
}
}
