% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_mul}
\alias{nvl_mul}
\title{Primitive Multiplication}
\usage{
nvl_mul(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Multiplies two tensors element-wise.
}
