% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_mul}
\alias{nvl_mul}
\title{Primitive Multiplication}
\usage{
nvl_mul(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Multiplies two tensors element-wise.
}
\section{Shapes}{

\code{lhs} and \code{rhs} must have the same shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_multiply]{stablehlo::hlo_multiply()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  y <- nv_tensor(c(4, 5, 6))
  nvl_mul(x, y)
})
\dontshow{\}) # examplesIf}
}
