% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_negate}
\alias{nvl_negate}
\title{Primitive Negation}
\usage{
nvl_negate(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of integer or floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Negates a tensor element-wise.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_negate]{stablehlo::hlo_negate()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, -2, 3))
  nvl_negate(x)
})
\dontshow{\}) # examplesIf}
}
