% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_not}
\alias{nvl_not}
\title{Primitive Not}
\usage{
nvl_not(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of boolean or integer type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logical NOT.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_not]{stablehlo::hlo_not()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(TRUE, FALSE, TRUE))
  nvl_not(x)
})
\dontshow{\}) # examplesIf}
}
