% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_not}
\alias{nvl_not}
\title{Primitive Not}
\usage{
nvl_not(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logical NOT.
}
