% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_or}
\alias{nvl_or}
\title{Primitive Or}
\usage{
nvl_or(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logical OR.
}
