% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_or}
\alias{nvl_or}
\title{Primitive Or}
\usage{
nvl_or(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}} of boolean or integer type)\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logical OR.
}
\section{Shapes}{

\code{lhs} and \code{rhs} must have the same shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_or]{stablehlo::hlo_or()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(TRUE, FALSE, TRUE))
  y <- nv_tensor(c(TRUE, TRUE, FALSE))
  nvl_or(x, y)
})
\dontshow{\}) # examplesIf}
}
