% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_pad}
\alias{nvl_pad}
\title{Primitive Pad}
\usage{
nvl_pad(
  operand,
  padding_value,
  edge_padding_low,
  edge_padding_high,
  interior_padding
)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{padding_value}{(\code{\link{tensorish}})\cr
Scalar value to use for padding.}

\item{edge_padding_low}{(\code{integer()})\cr
Amount of padding to add at the start of each dimension.}

\item{edge_padding_high}{(\code{integer()})\cr
Amount of padding to add at the end of each dimension.}

\item{interior_padding}{(\code{integer()})\cr
Amount of padding to add between elements in each dimension.}
}
\value{
\code{\link{tensorish}}
}
\description{
Pads a tensor with a given padding value.
}
