% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_pad}
\alias{nvl_pad}
\title{Primitive Pad}
\usage{
nvl_pad(
  operand,
  padding_value,
  edge_padding_low,
  edge_padding_high,
  interior_padding
)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{padding_value}{(\code{\link{tensorish}})\cr
Scalar value to use for padding.}

\item{edge_padding_low}{(\code{integer()})\cr
Amount of padding to add at the start of each dimension.}

\item{edge_padding_high}{(\code{integer()})\cr
Amount of padding to add at the end of each dimension.}

\item{interior_padding}{(\code{integer()})\cr
Amount of padding to add between elements in each dimension.}
}
\value{
\code{\link{tensorish}}
}
\description{
Pads a tensor with a given padding value.
}
\section{Shapes}{

\code{padding_value} must be scalar. \code{edge_padding_low}, \code{edge_padding_high}, and \code{interior_padding} must each have length equal to \code{rank(operand)}.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_pad]{stablehlo::hlo_pad()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  nvl_pad(x, nv_scalar(0),
    edge_padding_low = 2L, edge_padding_high = 1L, interior_padding = 0L
  )
})
}
}
