% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_popcnt}
\alias{nvl_popcnt}
\title{Primitive Population Count}
\usage{
nvl_popcnt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of data type integer or unsigned integer.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise population count (number of set bits).
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_popcnt]{stablehlo::hlo_popcnt()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(7L, 3L, 15L))
  nvl_popcnt(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_popcnt]{nv_popcnt()}}
}
