% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_popcnt}
\alias{nvl_popcnt}
\title{Primitive Population Count}
\usage{
nvl_popcnt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of integer type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise population count (number of set bits).
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_popcnt]{stablehlo::hlo_popcnt()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(7L, 3L, 15L))
  nvl_popcnt(x)
})
}
}
