% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_pow}
\alias{nvl_pow}
\title{Primitive Power}
\usage{
nvl_pow(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Raises lhs to the power of rhs element-wise.
}
