% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_pow}
\alias{nvl_pow}
\title{Primitive Power}
\usage{
nvl_pow(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}} of integer or floating-point type)\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Raises lhs to the power of rhs element-wise.
}
\section{Shapes}{

\code{lhs} and \code{rhs} must have the same shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_power]{stablehlo::hlo_power()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(2, 3, 4))
  y <- nv_tensor(c(3, 2, 1))
  nvl_pow(x, y)
})
\dontshow{\}) # examplesIf}
}
