% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_print}
\alias{nvl_print}
\title{Primitive Print}
\usage{
nvl_print(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Prints a tensor during execution.
Returns the input unchanged.
Note: Currently only works on CPU backend.
}
