% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_reduce_all}
\alias{nvl_reduce_all}
\title{Primitive All Reduction}
\usage{
nvl_reduce_all(operand, dims, drop = TRUE)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of boolean type)\cr
Operand.}

\item{dims}{(\code{integer()})\cr
Dimensions to reduce.}

\item{drop}{(\code{logical(1)})\cr
Whether to drop reduced dimensions.}
}
\value{
\code{\link{tensorish}} (boolean)
}
\description{
Logical AND along dimensions.
}
\section{Shapes}{

\itemize{
\item \code{operand}: any shape
\item output (\code{drop = TRUE}): shape of \code{operand} with \code{dims} removed
\item output (\code{drop = FALSE}): shape of \code{operand} with \code{dims} set to 1
}
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_reduce]{stablehlo::hlo_reduce()}} with \code{\link[stablehlo:hlo_and]{stablehlo::hlo_and()}} as the reducer.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(matrix(c(TRUE, FALSE, TRUE, TRUE), nrow = 2))
  nvl_reduce_all(x, dims = 1L)
})
}
}
