% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_reduce_any}
\alias{nvl_reduce_any}
\title{Primitive Any Reduction}
\usage{
nvl_reduce_any(operand, dims, drop = TRUE)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dims}{(\code{integer()})\cr
Dimensions to reduce.}

\item{drop}{(\code{logical(1)})\cr
Whether to drop reduced dimensions.}
}
\value{
\code{\link{tensorish}}
}
\description{
Logical OR along dimensions.
}
