% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_reduce_min}
\alias{nvl_reduce_min}
\title{Primitive Min Reduction}
\usage{
nvl_reduce_min(operand, dims, drop = TRUE)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dims}{(\code{integer()})\cr
Dimensions to reduce.}

\item{drop}{(\code{logical(1)})\cr
Whether to drop reduced dimensions.}
}
\value{
\code{\link{tensorish}}
}
\description{
Finds minimum along dimensions.
}
