% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_reduce_prod}
\alias{nvl_reduce_prod}
\title{Primitive Product Reduction}
\usage{
nvl_reduce_prod(operand, dims, drop = TRUE)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dims}{(\code{integer()})\cr
Dimensions to reduce.}

\item{drop}{(\code{logical(1)})\cr
Whether to drop reduced dimensions.}
}
\value{
\code{\link{tensorish}}
}
\description{
Multiplies tensor elements along dimensions.
}
\section{Shapes}{

\itemize{
\item \code{operand}: any shape
\item output (\code{drop = TRUE}): shape of \code{operand} with \code{dims} removed
\item output (\code{drop = FALSE}): shape of \code{operand} with \code{dims} set to 1
}
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_reduce]{stablehlo::hlo_reduce()}} with \code{\link[stablehlo:hlo_multiply]{stablehlo::hlo_multiply()}} as the reducer.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(matrix(1:6, nrow = 2))
  nvl_reduce_prod(x, dims = 1L)
})
\dontshow{\}) # examplesIf}
}
