% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_remainder}
\alias{nvl_remainder}
\title{Primitive Remainder}
\usage{
nvl_remainder(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}} of integer or floating-point type)\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise remainder of division.
}
\section{Shapes}{

\code{lhs} and \code{rhs} must have the same shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_remainder]{stablehlo::hlo_remainder()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(7, 10, 15))
  y <- nv_tensor(c(3, 4, 6))
  nvl_remainder(x, y)
})
\dontshow{\}) # examplesIf}
}
