% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_remainder}
\alias{nvl_remainder}
\title{Primitive Remainder}
\usage{
nvl_remainder(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise remainder of division.
}
