% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_reshape}
\alias{nvl_reshape}
\title{Primitive Reshape}
\usage{
nvl_reshape(operand, shape)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{shape}{(\code{integer()})\cr
Shape.}
}
\value{
\code{\link{tensorish}}
}
\description{
Reshapes a tensor to a new shape.
}
