% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_reshape}
\alias{nvl_reshape}
\title{Primitive Reshape}
\usage{
nvl_reshape(operand, shape)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{shape}{(\code{integer()})\cr
Shape.}
}
\value{
\code{\link{tensorish}}
}
\description{
Reshapes a tensor to a new shape.
}
\section{Shapes}{

\code{shape} must have the same number of elements as \code{operand}.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_reshape]{stablehlo::hlo_reshape()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(1:6)
  nvl_reshape(x, shape = c(2, 3))
})
\dontshow{\}) # examplesIf}
}
