% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_reverse}
\alias{nvl_reverse}
\title{Primitive Reverse}
\usage{
nvl_reverse(operand, dims)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dims}{(\code{integer()})\cr
Dimensions to reverse (1-indexed).}
}
\value{
\code{\link{tensorish}}
}
\description{
Reverses the order of elements along specified dimensions.
}
\section{Shapes}{

Output has the same shape as \code{operand}.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_reverse]{stablehlo::hlo_reverse()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(1, 2, 3, 4, 5))
  nvl_reverse(x, dims = 1L)
})
}
}
