% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_reverse}
\alias{nvl_reverse}
\title{Primitive Reverse}
\usage{
nvl_reverse(operand, dims)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{dimensions}{(\code{integer()})\cr
Dimensions to reverse (1-indexed).}
}
\value{
\code{\link{tensorish}}
}
\description{
Reverses the order of elements along specified dimensions.
}
