% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_rng_bit_generator}
\alias{nvl_rng_bit_generator}
\title{Primitive RNG Bit Generator}
\usage{
nvl_rng_bit_generator(
  initial_state,
  rng_algorithm = "THREE_FRY",
  dtype,
  shape_out
)
}
\arguments{
\item{initial_state}{(\code{\link{tensorish}})\cr
RNG state tensor.}

\item{rng_algorithm}{(\code{character(1)})\cr
Algorithm name (default "THREE_FRY").}

\item{dtype}{(\code{character(1)} | \code{\link[stablehlo:TensorDataType]{stablehlo::TensorDataType}})\cr
Data type.}

\item{shape_out}{(\code{integer()})\cr
Output shape.}
}
\value{
List of new state and random tensor.
}
\description{
Generates random bits using the specified algorithm.
}
