% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_round}
\alias{nvl_round}
\title{Primitive Round}
\usage{
nvl_round(operand, method = "nearest_even")
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{method}{(\code{character(1)})\cr
Rounding method ("nearest_even" or "afz").}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise rounding.
}
