% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_round}
\alias{nvl_round}
\title{Primitive Round}
\usage{
nvl_round(operand, method = "nearest_even")
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of floating-point type)\cr
Operand.}

\item{method}{(\code{character(1)})\cr
Rounding method ("nearest_even" or "afz").}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise rounding.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_round_nearest_even]{stablehlo::hlo_round_nearest_even()}} or
\code{\link[stablehlo:hlo_round_nearest_afz]{stablehlo::hlo_round_nearest_afz()}} depending on the \code{method} parameter.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(1.4, 2.5, 3.6))
  nvl_round(x)
})
}
}
