% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_rsqrt}
\alias{nvl_rsqrt}
\title{Primitive Reciprocal Square Root}
\usage{
nvl_rsqrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise reciprocal square root.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_rsqrt]{stablehlo::hlo_rsqrt()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(1, 4, 9))
  nvl_rsqrt(x)
})
}
}
