% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_rsqrt}
\alias{nvl_rsqrt}
\title{Primitive Reciprocal Square Root}
\usage{
nvl_rsqrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of data type floating-point.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise reciprocal square root.
For a more user-friendly interface, see \code{\link[=nv_rsqrt]{nv_rsqrt()}}.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_rsqrt]{stablehlo::hlo_rsqrt()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 4, 9))
  nvl_rsqrt(x)
})
\dontshow{\}) # examplesIf}
}
