% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_rsqrt}
\alias{nvl_rsqrt}
\title{Primitive Reciprocal Square Root}
\usage{
nvl_rsqrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise reciprocal square root.
}
