% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_scatter}
\alias{nvl_scatter}
\title{Primitive Scatter}
\usage{
nvl_scatter(
  input,
  scatter_indices,
  update,
  scatter_dimension_numbers,
  indices_are_sorted = FALSE,
  unique_indices = FALSE,
  update_computation
)
}
\arguments{
\item{input}{(\code{\link{tensorish}})\cr
Input tensor to scatter into.}

\item{scatter_indices}{(\code{\link{tensorish}})\cr
Tensor of integer type containing indices.}

\item{update}{(\code{\link{tensorish}})\cr
Update values tensor.}

\item{scatter_dimension_numbers}{(\code{\link{ScatterDimensionNumbers}})\cr
Dimension configuration for scatter operation.}

\item{indices_are_sorted}{(\code{logical(1)})\cr
Whether indices are sorted.}

\item{unique_indices}{(\code{logical(1)})\cr
Whether indices are unique.}

\item{update_computation}{(\code{function})\cr
Binary function to combine existing and update values.}
}
\value{
\code{\link{tensorish}}
}
\description{
Produces a result tensor equal to the input tensor except that
slices specified by scatter_indices are updated with values from the update tensor.
}
