% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_scatter}
\alias{nvl_scatter}
\title{Primitive Scatter}
\usage{
nvl_scatter(
  input,
  scatter_indices,
  update,
  update_window_dims = integer(),
  inserted_window_dims = integer(),
  input_batching_dims = integer(),
  scatter_indices_batching_dims = integer(),
  scatter_dims_to_operand_dims,
  index_vector_dim,
  indices_are_sorted = FALSE,
  unique_indices = FALSE,
  update_computation = NULL
)
}
\arguments{
\item{input}{(\code{\link{tensorish}})\cr
Input tensor to scatter into.}

\item{scatter_indices}{(\code{\link{tensorish}})\cr
Tensor of integer type containing indices.}

\item{update}{(\code{\link{tensorish}})\cr
Update values tensor.}

\item{update_window_dims}{(\code{integer()})\cr
1-based dimensions of the update tensor that correspond to window dimensions.}

\item{inserted_window_dims}{(\code{integer()})\cr
1-based dimensions inserted into the update tensor.}

\item{input_batching_dims}{(\code{integer()})\cr
1-based batch dimensions in the input tensor.}

\item{scatter_indices_batching_dims}{(\code{integer()})\cr
1-based batch dimensions in the scatter indices.}

\item{scatter_dims_to_operand_dims}{(\code{integer()})\cr
1-based mapping from scatter indices to operand dimensions.}

\item{index_vector_dim}{(\code{integer(1)})\cr
1-based dimension in scatter_indices containing the index vectors.}

\item{indices_are_sorted}{(\code{logical(1)})\cr
Whether indices are sorted.}

\item{unique_indices}{(\code{logical(1)})\cr
Whether indices are unique.}

\item{update_computation}{(\code{function})\cr
Binary function to combine existing and update values.}
}
\value{
\code{\link{tensorish}}
}
\description{
Produces a result tensor equal to the input tensor except that
slices specified by scatter_indices are updated with values from the update tensor.
}
