% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_shift_left}
\alias{nvl_shift_left}
\title{Primitive Shift Left}
\usage{
nvl_shift_left(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise left bit shift.
}
