% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_shift_right_arithmetic}
\alias{nvl_shift_right_arithmetic}
\title{Primitive Arithmetic Shift Right}
\usage{
nvl_shift_right_arithmetic(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise arithmetic right bit shift.
}
