% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_sign}
\alias{nvl_sign}
\title{Primitive Sign}
\usage{
nvl_sign(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of signed integer or floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise sign.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_sign]{stablehlo::hlo_sign()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(-3, 0, 5))
  nvl_sign(x)
})
\dontshow{\}) # examplesIf}
}
