% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_sqrt}
\alias{nvl_sqrt}
\title{Primitive Square Root}
\usage{
nvl_sqrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise square root.
}
