% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_sqrt}
\alias{nvl_sqrt}
\title{Primitive Square Root}
\usage{
nvl_sqrt(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Tensorish value of data type floating-point.}
}
\value{
\code{\link{tensorish}}\cr
Has the same shape and data type as the input.
It is ambiguous if the input is ambiguous.
}
\description{
Element-wise square root.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{backward}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_sqrt]{stablehlo::hlo_sqrt()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(1, 4, 9))
  nvl_sqrt(x)
})
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_sqrt]{nv_sqrt()}}, \code{\link[=sqrt]{sqrt()}}
}
