% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_static_slice}
\alias{nvl_static_slice}
\title{Primitive Slice}
\usage{
nvl_static_slice(operand, start_indices, limit_indices, strides)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{start_indices}{(\code{integer()})\cr
Start indices (1-based).}

\item{limit_indices}{(\code{integer()})\cr
End indices (exclusive).}

\item{strides}{(\code{integer()})\cr
Step sizes.}
}
\value{
\code{\link{tensorish}}
}
\description{
Extracts a slice from a tensor.
}
