% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_sub}
\alias{nvl_sub}
\title{Primitive Subtraction}
\usage{
nvl_sub(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Subtracts two tensors element-wise.
}
