% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_sub}
\alias{nvl_sub}
\title{Primitive Subtraction}
\usage{
nvl_sub(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}} of integer or floating-point type)\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Subtracts two tensors element-wise.
}
\section{Shapes}{

\code{lhs} and \code{rhs} must have the same shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_subtract]{stablehlo::hlo_subtract()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(1, 2, 3))
  y <- nv_tensor(c(4, 5, 6))
  nvl_sub(x, y)
})
}
}
