% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_tan}
\alias{nvl_tan}
\title{Primitive Tangent}
\usage{
nvl_tan(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise tangent.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_tan]{stablehlo::hlo_tan()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(0, 0.5, 1))
  nvl_tan(x)
})
}
}
