% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_tan}
\alias{nvl_tan}
\title{Primitive Tangent}
\usage{
nvl_tan(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise tangent.
}
