% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_tanh}
\alias{nvl_tanh}
\title{Primitive Hyperbolic Tangent}
\usage{
nvl_tanh(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise hyperbolic tangent.
}
