% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_tanh}
\alias{nvl_tanh}
\title{Primitive Hyperbolic Tangent}
\usage{
nvl_tanh(operand)
}
\arguments{
\item{operand}{(\code{\link{tensorish}} of floating-point type)\cr
Operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise hyperbolic tangent.
}
\section{Shapes}{

Operand can have any shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_tanh]{stablehlo::hlo_tanh()}}.
}

\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
jit_eval({
  x <- nv_tensor(c(-1, 0, 1))
  nvl_tanh(x)
})
\dontshow{\}) # examplesIf}
}
