% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_transpose}
\alias{nvl_transpose}
\title{Primitive Transpose}
\usage{
nvl_transpose(operand, permutation)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{permutation}{(\code{integer()})\cr
Dimension permutation.}
}
\value{
\code{\link{tensorish}}
}
\description{
Transposes a tensor according to a permutation.
}
