% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_transpose}
\alias{nvl_transpose}
\title{Primitive Transpose}
\usage{
nvl_transpose(operand, permutation)
}
\arguments{
\item{operand}{(\code{\link{tensorish}})\cr
Operand.}

\item{permutation}{(\code{integer()})\cr
Dimension permutation.}
}
\value{
\code{\link{tensorish}}
}
\description{
Transposes a tensor according to a permutation.
}
\section{Shapes}{

Output shape is \code{shape(operand)[permutation]}.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_transpose]{stablehlo::hlo_transpose()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(matrix(1:6, nrow = 2))
  nvl_transpose(x, permutation = c(2L, 1L))
})
}
}
