% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_while}
\alias{nvl_while}
\title{Primitive While Loop}
\usage{
nvl_while(init, cond, body)
}
\arguments{
\item{init}{(\code{list()})\cr
Named list of initial state values.}

\item{cond}{(\code{function})\cr
Condition function returning boolean.}

\item{body}{(\code{function})\cr
Body function returning updated state.}
}
\value{
Final state after loop terminates.
}
\description{
Executes a while loop.
}
