% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_while}
\alias{nvl_while}
\title{Primitive While Loop}
\usage{
nvl_while(init, cond, body)
}
\arguments{
\item{init}{(\code{list()})\cr
Named list of initial state values.}

\item{cond}{(\code{function})\cr
Condition function returning boolean.}

\item{body}{(\code{function})\cr
Body function returning updated state.}
}
\value{
Final state after loop terminates.
}
\description{
Executes a while loop.
}
\section{Shapes}{

\code{cond} must return a scalar boolean. \code{body} must return the same shapes as \code{init}.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_while]{stablehlo::hlo_while()}}.
}

\examples{
\dontrun{
jit_eval({
  nvl_while(
    init = list(i = nv_scalar(0L), total = nv_scalar(0L)),
    cond = function(i, total) nvl_lt(i, nv_scalar(5L)),
    body = function(i, total) list(
      i = nvl_add(i, nv_scalar(1L)),
      total = nvl_add(total, i)
    )
  )
})
}
}
