% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_xor}
\alias{nvl_xor}
\title{Primitive Xor}
\usage{
nvl_xor(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}})\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logical XOR.
}
