% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{nvl_xor}
\alias{nvl_xor}
\title{Primitive Xor}
\usage{
nvl_xor(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{tensorish}} of boolean or integer type)\cr
Left and right operand.}
}
\value{
\code{\link{tensorish}}
}
\description{
Element-wise logical XOR.
}
\section{Shapes}{

\code{lhs} and \code{rhs} must have the same shape. The output has the same shape.
}

\section{StableHLO}{

Calls \code{\link[stablehlo:hlo_xor]{stablehlo::hlo_xor()}}.
}

\examples{
\dontrun{
jit_eval({
  x <- nv_tensor(c(TRUE, FALSE, TRUE))
  y <- nv_tensor(c(TRUE, TRUE, FALSE))
  nvl_xor(x, y)
})
}
}
