% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{platform}
\alias{platform}
\title{Platform}
\usage{
platform(x, ...)
}
\arguments{
\item{x}{A tensor or PJRT object.}

\item{...}{Additional arguments (unused).}
}
\value{
The platform string.
}
\description{
Get the platform of a tensor or PJRT object.
}
