% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{platform}
\alias{platform}
\title{Platform}
\usage{
platform(x, ...)
}
\arguments{
\item{x}{(\code{anvil::Box})\cr
The tensor.}

\item{...}{(\code{any})\cr
Additional argument (unused).}
}
\value{
(\code{character(1)})
}
\description{
Get the platform of a tensor.
}
