% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{prim}
\alias{prim}
\title{Get a Primitive}
\usage{
prim(name = NULL)
}
\arguments{
\item{name}{(\code{character()} | \code{NULL})\cr
The name of the primitive.
If \code{NULL}, returns a list of all primitives.}
}
\value{
(\code{AnvilPrimitive})
}
\description{
Get a primitive by name.
}
