% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.R
\name{prim}
\alias{prim}
\title{Get a Primitive}
\usage{
prim(name)
}
\arguments{
\item{name}{(\code{character()})\cr
The name of the primitive.}
}
\value{
(\code{Primitive})
}
\description{
Get a primitive by name.
}
