% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{prim_or}
\alias{prim_or}
\title{Primitive Or}
\usage{
prim_or(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{arrayish}})\cr
Arrayish values of data type boolean, integer, or unsigned integer.
Must have the same shape.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the inputs.
It is ambiguous if both inputs are ambiguous.
}
\description{
Element-wise logical OR.
}
\section{Access}{

Access this primitive via \code{prim$or}.
}

\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{quickr}
\item \code{reverse}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_or]{stablehlo::hlo_or()}}.
}

\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_array(c(TRUE, FALSE, TRUE))
y <- nv_array(c(TRUE, TRUE, FALSE))
prim$or(x, y)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_or]{nv_or()}}, \code{|}
}
