% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{prim_remainder}
\alias{prim_remainder}
\title{Primitive Remainder}
\usage{
prim_remainder(lhs, rhs)
}
\arguments{
\item{lhs, rhs}{(\code{\link{arrayish}})\cr
Arrayish values of data type integer, unsigned integer, or floating-point.
Must have the same shape.}
}
\value{
\code{\link{arrayish}}\cr
Has the same shape and data type as the inputs.
It is ambiguous if both inputs are ambiguous.
}
\description{
Element-wise remainder.
Result has sign of the divident, which differs from base R's \code{\%\%}, which is available
via \code{\link[=nv_mod]{nv_mod()}} and has sign of divisor.
}
\section{Implemented Rules}{

\itemize{
\item \code{stablehlo}
\item \code{reverse}
}
}

\section{StableHLO}{

Lowers to \code{\link[stablehlo:hlo_remainder]{stablehlo::hlo_remainder()}}.
}

\examples{
\dontshow{if (pjrt::plugins_downloaded()) withAutoprint(\{ # examplesIf}
prim_remainder(1, -3)
1 \%\% -3
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=nv_remainder]{nv_remainder()}}
}
