% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{PrimitiveCall}
\alias{PrimitiveCall}
\title{Primitive Call}
\usage{
PrimitiveCall(
  primitive = Primitive(),
  inputs = list(),
  params = list(),
  outputs = list()
)
}
\arguments{
\item{primitive}{(\code{Primitive})\cr
The function.}

\item{inputs}{(\code{list(GraphValue)})\cr
The (tensor) inputs to the primitive.}

\item{params}{(\verb{list(<any>)})\cr
The (static) parameters of the function call.}

\item{outputs}{(\code{list(GraphValue)})\cr
The (tensor) outputs of the primitive.}
}
\description{
Call of a primitive in a \code{\link{Graph}}
Note that a primitive call also be a call into another graph (\code{p_graph}).
}
