% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpreter-pullback.R
\name{pullback}
\alias{pullback}
\title{Pullback of a function}
\usage{
pullback(f, ...)
}
\arguments{
\item{f}{(\code{function})\cr
Function to compute the pullback of.}

\item{...}{(\code{any})\cr
Example arguments to pass to the function.}
}
\value{
(\code{function})
}
\description{
Compute the pullback (transposed derivative) of a function.
}
