% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/Response.R
\docType{methods}
\name{Response}
\alias{Response}
\alias{flatten}
\alias{flatten,response-method}
\alias{str,response-method}
\alias{as.list.response}
\title{Process service responses to tibble and other data structures.}
\usage{
flatten(x)

\S4method{str}{response}(object)

\method{as.list}{response}(x, ..., as = "text")
}
\arguments{
\item{x}{A `response` object returned by the service.}

\item{object}{A `response` object returned by the service.}

\item{\dots}{not currently used}

\item{as}{a character(1); one of 'raw', 'text', 'parsed'}
}
\value{
`flatten()` returns a `tibble` where each row correseponds
    to a top-level list element of the return value, and columns
    are the unlisted second and more-nested elements.

`str()` displays a compact representation of the list-like
    JSON response; it returns `NULL`.

a list with the content of the response
}
\description{
Process service responses to tibble and other data structures.
}
\examples{
\donttest{leonardo$listClusters() \%>\% flatten()}

\donttest{leonardo$getSystemStatus() \%>\% str()}

}
