% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response.R
\name{Response}
\alias{Response}
\alias{flatten,response-method}
\alias{flatten}
\alias{str,response-method}
\alias{as.list.response}
\title{Process service responses to tibble and other data structures.}
\usage{
flatten(x)

\S4method{str}{response}(object)

\method{as.list}{response}(x, ..., as = c("text", "raw", "parsed"))
}
\arguments{
\item{x}{A `response` object returned by the service.}

\item{object}{A `response` object returned by the service.}

\item{\dots}{not currently used}

\item{as}{character(1); one of 'raw', 'text', 'parsed'}
}
\value{
`flatten()` returns a `tibble` where each row correseponds
    to a top-level list element of the return value, and columns
    are the unlisted second and more-nested elements.

`str()` displays a compact representation of the list-like
    JSON response; it returns `NULL`.

`as.list()` retruns the content of the web service request
    as a list.
}
\description{
Process service responses to tibble and other data structures.
}
\examples{
if (gcloud_exists()) {
    leonardo <- Leonardo()
    leonardo$listClusters() \%>\% flatten()
}

if (gcloud_exists())
   leonardo$getSystemStatus() \%>\% str()

if (gcloud_exists())
    leonardo$getSystemStatus() \%>\% as.list()

}
