% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R, R/Service.R, R/Leonardo.R,
%   R/Terra.R, R/Dockstore.R, R/Gen3.R
\docType{methods}
\name{operations}
\alias{operations}
\alias{schemas}
\alias{Service}
\alias{.DollarNames.Service}
\alias{operations,Service-method}
\alias{schemas,Service-method}
\alias{show,Service-method}
\alias{Service-class}
\alias{tags}
\alias{$,Service-method}
\alias{leonardo}
\alias{terra}
\alias{Terra-class}
\alias{operations,Terra-method}
\alias{schemas,Terra-method}
\alias{dockstore}
\alias{Dockstore-class}
\alias{operations,Dockstore-method}
\alias{gen3_fence}
\alias{gen3_indexd}
\alias{gen3_sheepdog}
\alias{gen3_peregrine}
\title{RESTful services useful for AnVIL developers}
\format{An object of class \code{Service} of length 1.}
\usage{
operations(x)

schemas(x)

tags(x)

\S4method{$}{Service}(x, name)

leonardo

terra

dockstore

gen3_fence

gen3_indexd

gen3_sheepdog

gen3_peregrine
}
\arguments{
\item{x}{A `Service` instance, usually a singleton provided by the
package and documented on this page, e.g., `leonardo` or
`terra`.}

\item{name}{A symbol representing a defined operation, e.g.,
`leonardo$listClusters()`.}
}
\value{
`leonardo` represents the API of the Leonard container
    deployment service at https://leonardo.dev.anvilproject.org/.

`terra` represents the API of the Terra cloud computational
    environemnt at https://api.firecloud.org/.

`dockstore` represents the API of the Dockstore platform to
    share Docker-based tools in CWL or WDL or Nextflow at
    https://dockstore.org

`gen3_*` APIs are not fully implemented, because a service
    endpoint has not been identifiied.

    `gen3_fence` provides authentication at
    https://raw.githubusercontent.com/uc-cdis/fence/master/openapis/swagger.yaml

`gen3_indexd` provides an indexing service documented at
    https://raw.githubusercontent.com/uc-cdis/indexd/master/openapis/swagger.yaml

`gen3_sheepdog` provides submission services at
    https://raw.githubusercontent.com/uc-cdis/sheepdog/master/openapi/swagger.yml

`gen3_peregrine` provides graphQL query services at
    https://raw.githubusercontent.com/uc-cdis/peregrine/master/openapis/swagger.yaml
}
\description{
RESTful services useful for AnVIL developers
}
\keyword{datasets}
