% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Service.R
\name{Service}
\alias{Service}
\title{RESTful service constructor}
\usage{
Service(
  service,
  host,
  config = httr::config(),
  authenticate = TRUE,
  api_url = character(),
  package = "AnVIL",
  schemes = "https",
  api_reference_url = api_url,
  api_reference_md5sum = character()
)
}
\arguments{
\item{service}{character(1) The `Service` class name, e.g., `"terra"`.}

\item{host}{character(1) host name that provides the API resource,
e.g., `"api.firecloud.org"`.}

\item{config}{httr::config() curl options}

\item{authenticate}{logical(1) use credentials from authentication
service file 'auth.json' in the specified package?}

\item{api_url}{optional character(1) url location of OpenAPI
`.json` or `.yaml` service definition.}

\item{package}{character(1) (default `AnVIL`) The package where
'api.json' yaml and (optionally) 'auth.json' files are located.}

\item{schemes}{character(1) (default 'https') Specifies the
transfer protocol supported by the API service.}

\item{api_reference_url}{character(1) path to reference API. See
Details.}

\item{api_reference_md5sum}{character(1) the result of
`tools::md5sum()` applied to the reference API.}
}
\value{
An object of class \code{Service}.
}
\description{
RESTful service constructor
}
\details{
This function creates a RESTful interface to a service
    provided by a host, e.g., "api.firecloud.org". The function
    requires an OpenAPI `.json` or `.yaml` specifcation as well as
    an (optional) `.json` authentication token. These files are
    located in the source directory of a pacakge, at
    `<package>/inst/service/<service>/api.json` and
    `<package>/inst/service/<service>/auth.json`, or at `api_url`.

    When provided, the `api_reference_md5sum` is used to check that
    the file described at `api_reference_url` has the same checksum
    as an author-validated version.

    The service is usually a singleton, created at the package
    level during `.onLoad()`.
}
\examples{
.MyService <- setClass("MyService", contains = "Service")

MyService <- function() {
    .MyService(Service("my_service", host="my.api.org"))
}

}
