% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-generics.R
\name{[.AnvilBox}
\alias{[.AnvilBox}
\alias{[.AnvilTensor}
\title{Subset Tensor}
\usage{
\method{[}{AnvilBox}(x, ...)

\method{[}{AnvilTensor}(x, ...)
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
Tensor to subset.}

\item{...}{Slice specifications.}

\item{drop}{(\code{logical(1)})\cr
Whether to drop dimensions of size 1.}
}
\value{
\code{\link{tensorish}}
}
\description{
Extract elements from a tensor using \code{[} indexing.
}
