% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-generics.R
\name{[<-.AnvilBox}
\alias{[<-.AnvilBox}
\title{Update Tensor Slice}
\usage{
\method{[}{AnvilBox}(x, ...) <- value
}
\arguments{
\item{x}{(\code{\link{tensorish}})\cr
Tensor to update.}

\item{...}{Slice specifications.}

\item{value}{(\code{\link{tensorish}})\cr
Values to assign.}
}
\value{
\code{\link{tensorish}}
}
\description{
Update elements of a tensor using \verb{[<-} indexing.
}
