% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tensorish}
\alias{tensorish}
\title{Tensor-like Objects}
\description{
A value that is either an \code{\link[=nv_tensor]{AnvilTensor}}, can be converted to it, or
represents an abstract version of it.
This also includes atomic R vectors.
}
\examples{
\dontshow{if (pjrt::plugin_is_downloaded()) withAutoprint(\{ # examplesIf}
x <- nv_tensor(1:4, dtype = "f32")
x
\dontshow{\}) # examplesIf}
}
\seealso{
\link{nv_tensor}, \link{ConcreteTensor}, \link{AbstractTensor}, \link{LiteralTensor}, \link{GraphBox}
}
