% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{to_abstract}
\alias{to_abstract}
\title{Convert to Abstract Tensor}
\usage{
to_abstract(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to convert.}
}
\value{
\code{\link{AbstractTensor}}
}
\description{
Convert an object to its abstract tensor representation (\code{\link{AbstractTensor}}).
}
