% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{trace_fn}
\alias{trace_fn}
\title{Trace an R function into a Graph}
\usage{
trace_fn(
  f,
  args = NULL,
  desc = NULL,
  toplevel = FALSE,
  flat_inputs = NULL,
  in_tree = NULL
)
}
\arguments{
\item{f}{(\code{function})\cr
The function to trace_fn.}

\item{args}{(\code{list} of (\code{\link{AnvilTensor}} | \code{\link{AbstractTensor}}))\cr
The arguments to the function. Can be \code{NULL} if \code{flat_inputs} is provided.}

\item{desc}{(\code{NULL} | \code{GraphDescriptor})\cr
The descriptor to use for the graph.}

\item{toplevel}{(\code{logical(1)})\cr
Whether the function is being traced at the top level.
If this is \code{TRUE}, inputs that are \code{AnvilTensor}s are treated as unknown.
If this is \code{FALSE} (default), \code{AnvilTensor}s are treated as constants.}

\item{flat_inputs}{(\code{NULL} | \code{list})\cr
Pre-flattened inputs. If provided, \code{args} is ignored and \code{in_tree} must be provided.}

\item{in_tree}{(\code{NULL} | \code{Node})\cr
The input tree structure. Required if \code{flat_inputs} is provided.}
}
\value{
(\code{\link{AnvilGraph}})
}
\description{
Create a graph representation of an R function by tracing.
}
