% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{trace_fn}
\alias{trace_fn}
\title{Trace an R function into a Graph}
\usage{
trace_fn(f, args, desc = NULL)
}
\arguments{
\item{f}{(\code{function})\cr
The function to trace_fn.}

\item{args}{(\code{list} of (\code{\link{AnvilTensor}} | \code{\link{AbstractTensor}}))\cr
The arguments to the function.}

\item{desc}{(\code{NULL} | \code{GraphDescriptor})\cr
The descriptor to use for the graph.}
}
\value{
(\code{\link{Graph}})
}
\description{
Create a graph representation of an R function by tracing.
}
