% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{trace_fn}
\alias{trace_fn}
\title{Trace an R function into a Graph}
\usage{
trace_fn(
  f,
  args = NULL,
  desc = NULL,
  toplevel = FALSE,
  lit_to_tensor = FALSE,
  args_flat = NULL,
  in_tree = NULL
)
}
\arguments{
\item{f}{(\code{function})\cr
The function to trace_fn.}

\item{args}{(\code{list} of (\code{\link{AnvilTensor}} | \code{\link{AbstractTensor}}))\cr
The (unflattened) arguments to the function.}

\item{desc}{(\code{NULL} | \code{GraphDescriptor})\cr
The descriptor to use for the graph.}

\item{toplevel}{(\code{logical(1)})\cr
Whether the function is being traced at the top level.
If this is \code{TRUE}, inputs that are \code{AnvilTensor}s are treated as unknown.
If this is \code{FALSE} (default), \code{AnvilTensor}s are treated as constants.}

\item{lit_to_tensor}{(\code{logical(1)})\cr
Whether to convert literals to \code{AnvilTensor}s.
Should only be used for higher-order primitives like if and while, where no static inputs are possible.}

\item{args_flat}{(\code{list})\cr
The flattened arguments. Also requires passing \code{in_tree}.}

\item{in_tree}{(\code{Node})\cr
The tree structure of the arguments.}
}
\value{
(\code{\link{AnvilGraph}})
}
\description{
Create a graph representation of an R function by tracing.
}
