% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{trace_fn}
\alias{trace_fn}
\title{Trace an R function into a Graph}
\usage{
trace_fn(f, args, desc = NULL, toplevel = FALSE, tensorish_args = FALSE)
}
\arguments{
\item{f}{(\code{function})\cr
The function to trace_fn.}

\item{args}{(\code{list} of (\code{\link{AnvilTensor}} | \code{\link{AbstractTensor}}))\cr
The arguments to the function.}

\item{desc}{(\code{NULL} | \code{GraphDescriptor})\cr
The descriptor to use for the graph.}

\item{toplevel}{(\code{logical(1)})\cr
Whether the function is being traced at the top level.
If this is \code{TRUE}, inputs that are \code{AnvilTensor}s are treated as unknown.
If this is \code{FALSE} (default), \code{AnvilTensor}s are treated as constants.}

\item{tensorish_args}{(\code{logical(1)})\cr
Whether the arguments are all tensorish.
If this is \code{TRUE}, we convert R literals to scalar tensors.}
}
\value{
(\code{\link{AnvilGraph}})
}
\description{
Create a graph representation of an R function by tracing.
}
