% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward.R
\name{transform_gradient}
\alias{transform_gradient}
\title{Transform a graph to its gradient}
\usage{
transform_gradient(graph, wrt)
}
\arguments{
\item{graph}{(\code{\link{AnvilGraph}})\cr
The graph to transform.}

\item{wrt}{(\code{character})\cr
The names of the variables to compute the gradient with respect to.}
}
\value{
An \code{\link{AnvilGraph}} object.
}
\description{
Transform a graph to its gradient.
This is a low-level function that should usually not be used directly.
Use \code{\link[=gradient]{gradient()}} instead.
}
