% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.R
\name{unflatten}
\alias{unflatten}
\title{Unflatten}
\usage{
unflatten(node, x)
}
\arguments{
\item{node}{Tree node describing the structure.}

\item{x}{Flat list to unflatten.}
}
\value{
Reconstructed nested structure.
}
\description{
Reconstruct a nested structure from a flat list using a tree structure.
}
