% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{utilities}
\alias{utilities}
\alias{add_libpaths}
\title{Utilities for managing library paths}
\usage{
add_libpaths(paths)
}
\arguments{
\item{paths}{\code{character()}: vector of directories to add to
\code{.libPaths()}. Paths that do not exist will be created.}
}
\value{
\code{add_libpaths()}: updated \code{.libPaths()}, invisibly.
}
\description{
\code{add_libpaths()}: Add local library paths to
\code{.libPaths()}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
add_libpaths("/tmp/host-site-library")
\dontshow{\}) # examplesIf}
}
