% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{with_backend}
\alias{with_backend}
\title{Temporarily override the active backend}
\usage{
with_backend(backend, code)
}
\arguments{
\item{backend}{(\code{character(1)})\cr
Backend to use (\code{"xla"} or \code{"quickr"}).}

\item{code}{Expression to evaluate with the overridden backend.}
}
\value{
The result of evaluating \code{code}.
}
\description{
Temporarily override the active backend
}
\keyword{internal}
